% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{abundPlot}
\alias{abundPlot}
\title{Rank-specific abundance plot}
\usage{
abundPlot(rank, physeqFullObject, xAesLogic, numTopRanks, group)
}
\arguments{
\item{physeqFullObject}{(phyloseq object),x (rank)}
}
\value{
Returns a ggplot
}
\description{
Abundance distribution for a specific rank
}
