% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{add_centered_title}
\alias{add_centered_title}
\title{Formats ggplot title}
\usage{
add_centered_title(p, text)
}
\arguments{
\item{p}{(ggplot), text (title)}
}
\value{
A plot.grid
}
\description{
Centered and right size
}
