% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addCountResultSummary}
\alias{addCountResultSummary}
\title{Adds a summary of the count result}
\usage{
addCountResultSummary(doc, param, result)
}
\arguments{
\item{doc}{an object of the class bsdoc to add the table.}

\item{param}{a list of parameters to influence the output:
\itemize{
 \item{batch}{ a logical indicating whether the second factor was used.}
 \item{comparison}{ which comparison was used.}
 \item{normMethod}{ the normalization method.}
 \item{sigThresh}{ the threshold...}
 \item{useSigThresh}{ ...and whether it should be used.}
}}

\item{result}{a list of results.}
}
\description{
Adds a summary of the count result to a bsdoc object.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[ReporteRs]{addFlexTable}}
}

