% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\name{addDEXSeqCondition}
\alias{addDEXSeqCondition}
\title{Addition experimental conditions to input files}
\usage{
addDEXSeqCondition(psInput, pvCondition)
}
\description{
Addition experimental conditions to input files
}

