% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage-reports.R
\name{addGageTables}
\alias{addGageTables}
\title{Adds the gage tables}
\usage{
addGageTables(doc, param = NULL, gageResults = NULL)
}
\arguments{
\item{doc}{an object of the class bsdoc to add the tables.}

\item{param}{a list of parameters, possibly passed to other functions as well:
\itemize{
 \item{gageThreshold}{ the threshold for significant pathways.}
 \item{gageGeneThreshold}{ the threshold for genes selected within a pathway.}
 \item{pathview}{ a logical indicating whether to add links to the kegg pathview.}
}}

\item{gageResults}{the results of the gage analysis.}
}
\description{
Adds the gage tables to an html file.
}
\seealso{
\code{\link{gageAnalysis}}

\code{\link{twoGroupsGO}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
