% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addJavaScriptIgvStarter}
\alias{addJavaScriptIgvStarter}
\title{Adds a java function}
\usage{
addJavaScriptIgvStarter(htmlFile, projectId, doc)
}
\arguments{
\item{htmlFile}{a character, ending with .html, representing the file path of an html file.}

\item{projectId}{a character representing the project ID.}

\item{doc}{an object of the class bsdoc to add the JS starter.}
}
\description{
Adds a java function to start Igv from Jnlp.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

