% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addResultFile}
\alias{addResultFile}
\title{Adds a result file}
\usage{
addResultFile(
  doc,
  param,
  result,
  rawData,
  useInOutput = TRUE,
  file = paste0("result--", param$comparison, ".txt")
)
}
\arguments{
\item{doc}{an object of the class bsdoc to add the result.}

\item{param}{a list of parameters that pastes the \code{comparison} into the file name and does a zip file if \code{doZip} is true.}

\item{result}{a list of results.}

\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}

\item{useInOutput}{a logical specifying whether to use most of the result information.}

\item{file}{a character representing the name of the result file.}
}
\value{
Returns the name of the result file.
}
\description{
Adds a result file in text format or zipped.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
