% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addSignificantCounts}
\alias{addSignificantCounts}
\alias{getSignificantCountsTable}
\alias{getSignificantFoldChangeCountsTable}
\title{Adds tables of the significant counts}
\usage{
addSignificantCounts(doc, result, pThresh = c(0.1, 0.05, 1/10^(2:5)))

getSignificantCountsTable(result, pThresh = 1/10^(1:5), genes = NULL)

getSignificantFoldChangeCountsTable(result, pThresh = 1/10^(1:5),
  fcThresh = c(1, 1.5, 2, 3, 4, 8, 10), genes = NULL)
}
\arguments{
\item{doc}{an object of the class bsdoc to add the table.}

\item{result}{a list of results.}

\item{pThresh}{a numerical indicating the p-value threshold. genes}
}
\description{
Adds tables of the significant counts.
}
\section{Functions}{
\itemize{
\item \code{getSignificantCountsTable}: Gets the table containing the significant counts.

\item \code{getSignificantFoldChangeCountsTable}: Gets the table containing the significant fold change counts.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

