% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-reports.R
\name{addTestScatterPlots}
\alias{addTestScatterPlots}
\title{Adds test scatter plots}
\usage{
addTestScatterPlots(doc, param, x, result, seqAnno, resultFile, types = NULL)
}
\arguments{
\item{doc}{an object of the class bsdoc to add the plots.}

\item{param}{a list of parameters.}

\item{x}{a vector containing the signal.}

\item{result}{a list of results.}

\item{seqAnno}{the sequence annotation. This is used if types is NULL.}

\item{resultFile}{a character representing the link to the result file.}

\item{types}{a character vector containing the types.}
}
\description{
Adds test scatter plots to an html file.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

