% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addTitleWithAnchor}
\alias{addTitleWithAnchor}
\title{Adds a title with an anchor}
\usage{
addTitleWithAnchor(doc, title, level = 1)
}
\arguments{
\item{doc}{an object of the class bsdoc to add the anchored title to.}

\item{title}{a character specifying the title.}

\item{level}{an integer specifying the heading level.}
}
\description{
Adds a title with an anchor by using \code{addTitle()} and \code{addCodeBlock()} from the ReporteRs package.
}
\examples{
theDoc = openBsdocReport(title="My html report")
title = "My title"
ezAddBootstrapMenu(theDoc, paste0("#", title))
addTitleWithAnchor(theDoc, title)
closeBsdocReport(doc=theDoc, file="example.html")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[ReporteRs]{addTitle}}

\code{\link[ReporteRs]{addCodeBlock}}
}

