% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{addTranscriptsToGffExons}
\alias{addTranscriptsToGffExons}
\title{Adds transcripts from exons to the annotation}
\usage{
addTranscriptsToGffExons(gff)
}
\arguments{
\item{gff}{the annotation data.frame to get transcripts from.}
}
\value{
Returns an annotation data.frame with transcripts added to it.
}
\description{
Adds transcripts from exons to the annotation using a gtf or gff data.frame.
}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param,"./inst/extdata/genes.gtf")
addTranscriptsToGffExons(gtf)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

