% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addTxtLinksToReport}
\alias{addTxtLinksToReport}
\title{Adds text links}
\usage{
addTxtLinksToReport(doc, txtNames, mime = "text/plain")
}
\arguments{
\item{doc}{an object of the class bsdoc to add the text links.}

\item{txtNames}{a character representing the link names.}

\item{mime}{a character representing the type of the links.}

\item{con}{the connection to the file to write in.}
}
\description{
Adds texts link to an html file.
}
\examples{
param = ezParam()
htmlFile = "example.html"
doc =  openBsdocReport(title="My report")
addTxtLinksToReport(doc, "dataset.tsv")
closeBsdocReport(doc, htmlFile)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

