% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{aggregateCountsByGene}
\alias{aggregateCountsByGene}
\title{Aggregates counts by gene}
\usage{
aggregateCountsByGene(param, rawData)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{geneColumnSet}{ a character naming the columns.}
  \item{sigThresh}{ the threshold...}
  \item{useSigThresh}{ ...and whether it should be used.}
}}

\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}
}
\value{
Returns the aggregates raw data.
}
\description{
Aggregates counts by gene.
}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
param$ezRef@refFeatureFile = "./inst/extdata/genes.gtf"
param$ezRef@refAnnotationFile = "./inst/extdata/genes_annotation_example.txt"
param$ezRef@refFastaFile = "/srv/GT/reference/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/WholeGenomeFasta/genome.fa"
input = EzDataset$new(file=system.file("extdata/yeast_10k_STAR_featureCounts/dataset.tsv", package="ezRun", mustWork = TRUE))
rawData = loadCountDataset(input, param)
rawData$seqAnno = writeAnnotationFromGtf(param)
aggregateCountsByGene(param, rawData)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{getGeneMapping}}
}

