% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-macs2.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Annotates peaks}
\usage{
annotatePeaks(peakFile, peakSeqFile, param)
}
\arguments{
\item{param}{a list of parameters to extract the \code{ezRef@refFeatureFile} and the \code{ezRef@refAnnotationFile} from.}

\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}
}
\description{
Annotates peaks and writes them into a separate table.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
