% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{averageAcrossRows}
\alias{averageAcrossRows}
\title{Averages rows together}
\usage{
averageAcrossRows(data, by = labels, func = mean, ...)
}
\arguments{
\item{data}{the matrix whose rows should be averaged.}

\item{by}{an integer or vector by which to average and/or rearrange rows.}

\item{func}{the function to apply to the result. Default: mean.}

\item{...}{additional arguments to be passed to \code{aggregate()}.}
}
\value{
Returns a vector or matrix of averaged rows.
}
\description{
Rearranges and averages rows according to \code{by}.
}
\examples{
m1 = matrix(1:20,5)
averageRows(m1,c(1,1,2,2,3))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
