% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{averageColumns}
\alias{averageColumns}
\title{Averages columns together}
\usage{
averageColumns(x, by = NULL, func = function(x) {
    mean(x, na.rm = TRUE)
})
}
\arguments{
\item{x}{the matrix whose columns should be averaged.}

\item{by}{an integer or vector by which to average and/or rearrange columns.}

\item{func}{the function to apply to the result. Default: mean with removing NAs.}
}
\value{
Returns a vector or matrix of averaged columns.
}
\description{
Rearranges and averages columns according to \code{by}.
}
\examples{
m1 = matrix(1:20,5)
rownames(m1) = 1:5
m2 = averageColumns(m1,1)
m3 = averageColumns(m1,c(4,2,3,1))
m4 = averageColumns(m1,c(1,1,2,2))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
