% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamio.r
\name{buildScanBamParam}
\alias{buildScanBamParam}
\title{Builds the ScanBam parameters}
\usage{
buildScanBamParam(param, bamFile)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{readUnmapped}{ whether unmapped (TRUE), mapped (FALSE) or any read (NA) should be returned.}
  \item{paired}{ a logical indicating whether the samples are paired.}
  \item{subsetGenome}{ a logical. If true, only a subset (1/20) of the genome region will be used.}
  \item{seqNames}{ if provided, the function will only use these chromosomes.}
}}

\item{bamFile}{a character representing the bam file path.}
}
\value{
Returns an object of the class ScanBamParam.
}
\description{
Builds the ScanBam parameters from other parameters and a bam file.
}
\examples{
param = ezParam()
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
buildScanBamParam(param, bamFile)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[Rsamtools]{ScanBamParam}}
}

