% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurDataCleanAssociatedFunctions.R
\name{chimeraSummaryPlot}
\alias{chimeraSummaryPlot}
\title{Chimera identification}
\usage{
chimeraSummaryPlot(x)
}
\arguments{
\item{chimerFile}{ezTable from  mothur chimera file.}
}
\value{
Returns a pie chart plot.
}
\description{
Summarizes chimera rates from chimera file
}
