% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R
\name{chimeraSummaryTable}
\alias{chimeraSummaryTable}
\title{Chimera identification}
\usage{
chimeraSummaryTable(chimFile, groupFile)
}
\arguments{
\item{chimerFile}{mothur chimera file.}
}
\value{
Returns a data frame.
}
\description{
Summarizes chimera rates from chimera file for sample
}
