% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-edgerMultiGroups.R
\name{cleanupMultiGroupsInput}
\alias{cleanupMultiGroupsInput}
\title{Cleans up input from the edgeR multi groups app}
\usage{
cleanupMultiGroupsInput(input, param)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{param}{a list of parameters to use or clean up.}
}
\value{
Returns an object of the class EzDataset that is the modified \code{input}.
}
\description{
Cleans up input from  the edgeR multi groups app.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

