% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoGroups.R
\name{cleanupTwoGroupsInput}
\alias{cleanupTwoGroupsInput}
\title{Cleans up input from two group apps}
\usage{
cleanupTwoGroupsInput(input, param)
}
\arguments{
\item{input}{an object of the class EzDataset.}

\item{param}{a list of parameters to use or clean up.}
}
\value{
Returns an object of the class EzDataset that is the modified \code{input}.
}
\description{
Cleans up input from two group apps.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

