% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{closeHTML}
\alias{closeHTML}
\title{Closes and wraps up an html connection}
\usage{
closeHTML(html)
}
\arguments{
\item{html}{a connection to an html file.}
}
\description{
Closes and wraps up an html connection adding the closing tags and a "Finished..." line.
}
\examples{
param = ezParam()
htmlFile = "example_html"
html =  openHtmlReport(htmlFile,param)
closeHTML(html)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

