% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R
\name{clusterResults}
\alias{clusterResults}
\alias{clusterHeatmap}
\alias{goClusterResults}
\title{Plots the cluster heatmap}
\usage{
clusterResults(
  x,
  nClusters = 5,
  clusterColors = rainbow(nClusters),
  d = NULL,
  method = "ward.D2"
)

clusterHeatmap(
  x,
  param,
  result,
  file = "cluster-heatmap.png",
  method = "ward.D2",
  doClusterColumns = FALSE,
  columnDist = NULL,
  colColors = NULL,
  lim = c(-4, 4),
  lwid = c(1, 4),
  lhei = c(1, 5),
  cexRow = 1,
  cexCol = 1.5,
  labRow = rownames(x),
  margins = c(14, 9),
  colors = getBlueRedScale(),
  maxGenesWithLabel = 50,
  ...
)

goClusterResults(
  x,
  param,
  result,
  ontologies = c("BP", "MF", "CC"),
  seqAnno = NULL,
  universeGeneIds = NULL,
  universeProbeIds = NULL,
  keggOrganism = NA
)
}
\arguments{
\item{x}{the data to plot.}

\item{method}{a character representing the method to pass to \code{hclust()}.}

\item{param}{a list of parameters to extract the logical \code{showGeneClusterLabels} from.}

\item{result}{a list of cluster result properties obtained with \code{clusterResults()}.}

\item{file}{a character representing the name of the .png file to derive tables from.}

\item{doClusterColumns}{a logical indicating whether to do cluster columns.}

\item{columnDist}{the distance matrix to use for the cluster columns.}

\item{colColors}{used for the \code{ColSideColors} argument passed to \code{heatmap.2()}.}

\item{lim}{two integers used for \code{breaks} argument passed to \code{heatmap.2()}.}

\item{cexRow}{an integer passed to \code{heatmap.2()}.}

\item{cexCol}{an integer passed to \code{heatmap.2()}.}

\item{labRow}{a character vector, possibly modified, then passed to \code{heatmap.2()}.}

\item{margins}{an integer, possibly modified, then passed to \code{heatmap.2()}.}

\item{colors}{a character vector containing colors.}

\item{maxGenesWithLabel}{an integer specifying the maximum amount of genes with labels.}

\item{...}{additional arguments to be passed to \code{heatmap.2()}.}
}
\description{
Plots the cluster heatmap.
}
\section{Functions}{
\itemize{
\item \code{clusterResults}: Initializes and returns a list of results used by \code{clusterHeatmap()} and \code{goClusterResults()}.

\item \code{goClusterResults}: Applies a GO analysis to the cluster results if GO should be done.
}}

\seealso{
\code{\link[stats]{hclust}}

\code{\link[gplots]{heatmap.2}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
