% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{collapse}
\alias{collapse}
\title{Collapses a set of values}
\usage{
collapse(x, sep = "; ", na.rm = FALSE, empty.rm = FALSE,
  uniqueOnly = FALSE)
}
\arguments{
\item{x}{a vector, matrix or list.}

\item{sep}{the separator to use between values.}

\item{na.rm}{a logical specifying whether to remove \code{NA}'s.}

\item{empty.rm}{a logical specifying whether to remove empty values.}

\item{uniqueOnly}{a logical specifying whether to keep only unique values.}
}
\value{
Returns the values collapsed into one character.
}
\description{
Collapses a set of values into a character and pastes them. Useful for documentation purposes.
}
\examples{
l1 = list(a=c(1,"",6),c=c("rsrg","yjrt",NA,6))
collapse(l1,sep="_")
collapse(l1,na.rm=T,empty.rm=T,uniqueOnly=T)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

