% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{colorClusterLabels}
\alias{colorClusterLabels}
\alias{colorNode}
\title{Gets the color cluster labels}
\usage{
colorClusterLabels(hcd, colors)

colorNode(n, cols = NULL)
}
\arguments{
\item{hcd}{an object of the class dendrogram.}

\item{colors}{a character vector containing colors.}
}
\value{
Returns the modified dendrogram.
}
\description{
Gets the color cluster labels using \code{dendrapply()}.
}
\section{Functions}{
\itemize{
\item \code{colorNode()}: The functions used in \code{dendrapply()}.

}}
\examples{
hc = hclust(dist(USArrests), "ave")
dend = as.dendrogram(hc)
ccl = colorClusterLabels(dend, rainbow(6))
}
\seealso{
\code{\link[stats]{dendrapply}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
