% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{communityPercSummTable}
\alias{communityPercSummTable}
\title{Summary of community composition}
\usage{
communityPercSummTable(phyloseqOtuObj, rank)
}
\arguments{
\item{a}{phyloseq object}
}
\value{
A data.frame
}
\description{
It calculates the percentages of the organisms in the community
}
