% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DnaQC.R
\name{computeDnaBamStats}
\alias{computeDnaBamStats}
\title{Computes the DnaQC statistics}
\usage{
computeDnaBamStats(input, htmlFile, param, resultList = NULL)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{htmlFile}{a character, ending with .html, representing the file path of an html file.}

\item{resultList}{a list of results.}
}
\description{
Computes the DnaQC statistics.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
