% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangesTools.r
\name{computeRangeStats}
\alias{computeRangeStats}
\title{Computes stats of genomic ranges}
\usage{
computeRangeStats(cov, targetRanges, FUN = mean)
}
\arguments{
\item{cov}{an object of the class RleList.}

\item{targetRanges}{an object of the class GRanges.}

\item{FUN}{a function to apply to the views.}
}
\value{
Returns the computed values belonging to their genomic ranges.
}
\description{
Computes stats of genomic ranges.
}
\examples{
cov = RleList(chr1=Rle(5523:5539),chr2=Rle(6544:6557))
iRanges = IRanges(5000:5002,8000:7998)
targetRanges = GRanges(c("chr1", "chr1", "chr2"), iRanges, strand=c("+", "-", "+"))
computeRangeStats(cov, targetRanges)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[IRanges]{RleViewsList}}

\code{\link[IRanges]{viewApply}}
}

