% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-VeloCyto.R
\name{convertCramToBam}
\alias{convertCramToBam}
\title{Convert CRAM to BAM if needed}
\usage{
convertCramToBam(inputFile, outputBam, cores = 1)
}
\arguments{
\item{inputFile}{path to the input file (can be BAM or CRAM)}

\item{outputBam}{desired output BAM file path}

\item{cores}{number of CPU cores to use (must be a positive integer)}
}
\value{
Returns the path to the BAM file (either original or converted)
}
\description{
Checks if the input file is a CRAM file and converts it to BAM format using samtools.
Requires samtools to be available in the PATH.
}
\details{
This function will stop with an error if:
\itemize{
  \item The input file does not exist
  \item The cores parameter is not a valid positive integer
  \item The BAM conversion fails
  \item The BAM indexing fails
}
}
