% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatranscriptomicsAssociatedFunctions.R
\name{convertDiamondAnnotationToAbund}
\alias{convertDiamondAnnotationToAbund}
\title{Mothur Summary Table}
\usage{
convertDiamondAnnotationToAbund(annFile)
}
\arguments{
\item{summary, }{mothur summary files.}
}
\value{
Returns a list of data.frames.
}
\description{
Create summary table from mothur summary files.
}
