% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqCounting.R
\name{convertGtfToGff}
\alias{convertGtfToGff}
\title{Convert annotation file from GTF to GFF}
\usage{
convertGtfToGff(psGtfFile)
}
\arguments{
\item{psGtfFile}{name of the GTF annotation file}
}
\description{
Convert annotation file from GTF to GFF
}

