% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R,
%   R/mothurErrorEstimateAssociatedFunctions.R
\name{convStepTable}
\alias{convStepTable}
\title{Clustering steps}
\usage{
convStepTable(convStepFile)

convStepTable(convStepFile)
}
\arguments{
\item{convStepFile, }{mothur steps file.}
}
\value{
Returns a table

Returns a table
}
\description{
Converegence iteration

Converegence iteration
}
