% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R
\name{countAndAssignSeqsFromFasta}
\alias{countAndAssignSeqsFromFasta}
\title{Mothur fasta summary}
\usage{
countAndAssignSeqsFromFasta(fastaFile, filterStep, groupFile)
}
\arguments{
\item{sharedFile, }{mothur fasta  file}
}
\value{
Returns a data frame
}
\description{
Summarizes read count across samples for a fasta file
}
