% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsPlots.R
\name{countDensPlot}
\alias{countDensPlot}
\title{Plots the count densities}
\usage{
countDensPlot(cts, colors, main = "all transcripts", bw = 7)
}
\arguments{
\item{cts}{the signal to use for plotting.}

\item{colors}{a character vector containing colors.}

\item{main}{a character representing the plot title.}

\item{bw}{a numeric passed to \code{density()}.}
}
\description{
Plots the count densities.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
