% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngsio.r
\name{countReadsInFastq}
\alias{countReadsInFastq}
\title{Counts reads in FastQ files}
\usage{
countReadsInFastq(fastqFiles)
}
\arguments{
\item{fastqFiles}{a character vector representing file paths to FastQ files.}
}
\value{
Returns a named numeric vector
}
\description{
Counts reads in FastQ files and returns the amount for each file.
}
\examples{
input = EzDataset$new(file=system.file("extdata/yeast_10k/dataset.tsv", package="ezRun", mustWork = TRUE))
param = list()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
fqFiles = input$getFullPaths(param, "Read1")
countReadsInFastq(fqFiles)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

