% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotationQC_associatedFunctions.R
\name{createAbundTable}
\alias{createAbundTable}
\title{createAbundTable}
\usage{
createAbundTable(mergedSummaryBinDF)
}
\arguments{
\item{the}{data frame generated by mergeSummaryBinFiles}
}
\value{
Returns a data frame
}
\description{
It creates abundance table from the organims annotations to bins
}
