% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-macs2.R
\name{createBigWig}
\alias{createBigWig}
\title{Creates a bigwig file}
\usage{
createBigWig(input = NA, output = NA, param = NA)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to extract \code{paired} from.}
}
\description{
Creates and exports a bigwig file.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
