% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurDataCleanAssociatedFunctions.R
\name{createSaturationTableForKableExtra}
\alias{createSaturationTableForKableExtra}
\title{Creates saturation  convergence table for Kable_extra output}
\usage{
createSaturationTableForKableExtra(x)
}
\arguments{
\item{x}{list of mothur shared files for which to create the combined table}
}
\value{
Returns a ktable and the above header
}
\description{
Merges summary tables for Kable_extra output
}
