% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurDataCleanAssociatedFunctions.R
\name{createStepConvTableForKableExtra}
\alias{createStepConvTableForKableExtra}
\title{Creates step convergence table for Kable_extra output}
\usage{
createStepConvTableForKableExtra(x)
}
\arguments{
\item{x}{list of files for which to create the combined table}
}
\value{
Returns a ktable and the above header
}
\description{
Merges summary tables for Kable_extra output
}
