% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R,
%   R/mothurDataCleanAssociatedFunctions.R
\name{createSummaryTable}
\alias{createSummaryTable}
\title{Mothur Summary Table}
\usage{
createSummaryTable(summary)

createSummaryTable(summary)
}
\arguments{
\item{summary, }{ezTable from mothur summary files.}
}
\value{
Returns a data.frame.

Returns a data.frame.
}
\description{
Create summary table from mothur summary files.

Create summary table from mothur summary files.
}
