% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2associatedFunctions.R
\name{DADA2mergeSeqTabs}
\alias{DADA2mergeSeqTabs}
\title{DADA2 seqtab merging, removing chimeras and assigning taxonomy}
\usage{
DADA2mergeSeqTabs(listOfSeqTabsFiles, database)
}
\arguments{
\item{a}{list of  DADA2 seqTables.}
}
\value{
Returns a merged and annotated DADA2 seqtab object, without chimeras and chimera stats.
}
\description{
Merges   DADA2 seqTables.
}
