% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R
\name{datasetToMothur}
\alias{datasetToMothur}
\title{Creates Mothur input files from  Sushi dataset.}
\usage{
datasetToMothur(sushiInputDataset, param)
}
\arguments{
\item{taxaFileName, }{mothur taxonomy file.}
}
\value{
Returns the .groups and .fasta files
}
\description{
Converts Sushi dataset into Mothur input.
}
