% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt_report.R
\name{delete_e}
\alias{delete_e}
\title{delete "-e " strings}
\usage{
delete_e(linesRead)
}
\arguments{
\item{linesRead}{<string> vector output from readLines()}
}
\description{
delete "-e " strings in *.stats processed output document to create the required dataframe

for downstream processing
}
\author{
Miquel Anglada Girotto
}
