% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R
\name{doGo}
\alias{doGo}
\alias{hasGoAnnotation}
\title{Do GO?}
\usage{
doGo(param, seqAnno)

hasGoAnnotation(seqAnnoDF)
}
\arguments{
\item{param}{a list of parameters to extract the logical \code{runGO} and the character \code{featureLevel} from.}

\item{seqAnno}{the sequence annotation.}

\item{seqAnnoDF}{a data.frame containing the sequence annotation.}
}
\value{
Returns a logical.
}
\description{
Decides whether to do a gene ontologies analysis.
}
\section{Functions}{
\itemize{
\item \code{hasGoAnnotation}: Checks if the Annotation contains gene ontologies.
}}
\examples{
param = ezParam()
param$ezRef@refFeatureFile = system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE)
annoFile = system.file("extdata/genes_annotation_example.txt", package="ezRun", mustWork=TRUE)
param$ezRef@refAnnotationFile = annoFile
fp = "/srv/GT/reference/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/WholeGenomeFasta/genome.fa"
param$ezRef@refFastaFile = fp
seqAnno = writeAnnotationFromGtf(param)
doGo(param, seqAnno)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

