% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{enrichrAddList}
\alias{enrichrAddList}
\title{Add gene list}
\usage{
enrichrAddList(genes)
}
\arguments{
\item{genes}{A list or vector containing gene names}
}
\value{
Returns a list with userListId and shortId
}
\description{
Registers a gene list with the Enrichr server
}
\seealso{
\code{\link{runEnrichr}}
}
\author{
Rehrauer, Hubert

Schmid, Peter

Roman Briskine
}
