% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{enrichrEnrich}
\alias{enrichrEnrich}
\title{Enrichment analysis}
\usage{
enrichrEnrich(
  userListId,
  libNames = getEnrichrLibNames(),
  connectionN = 10,
  retryN = 3
)
}
\arguments{
\item{userListId}{userListId returned by the Enrichr server via \code{\link{enrichrAddList}}}

\item{libNames}{vector of library names that should be a subset of those specified at the
\href{http://amp.pharm.mssm.edu/Enrichr/#stats}{Enrichr} website. This function does not
 validate the list. If the value is NULL, the function will use the internal list.}

\item{connectionN}{maximum number of concurrent connections to make}

\item{retryN}{total number of times to retry a query if it fails}
}
\value{
\code{list} with two elements: success and failure. The first element is a \code{list}
of \code{data.frame} objects that contain enrichment results per library. The names of the
success list match the library names. The failure \code{list} contains error messages for the
libraries that failed enrichment tests. The name of this \code{list} may not match the library
names. You can determine which libraries failed by comparing the success list with the supplied
library names.
}
\description{
Runs enrichment analysis against the specified databases. By default, the function
will use the internal list (\code{\link{getEnrichrLibNames}}) of the library names.
}
\seealso{
\code{\link{runEnrichr}, \link{enrichrAddList}, \link{getEnrichrLibNames}}
}
\author{
Rehrauer, Hubert

Schmid, Peter

Roman Briskine
}
