% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R,
%   R/mothurErrorEstimateAssociatedFunctions.R
\name{errorRateSummaryPlot}
\alias{errorRateSummaryPlot}
\title{Error rate}
\usage{
errorRateSummaryPlot(errorCountFileName, avReadLen)

errorRateSummaryPlot(errorCountFileName, avReadLen)
}
\arguments{
\item{errorCountFile, }{mothur taxonomy file.}
}
\value{
Returns a number.

Returns a pie chart plot.
}
\description{
Summarizes error rate from error count file

Summarizes error rate from error count file
}
