% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rangesTools.r
\name{expandGRanges}
\alias{expandGRanges}
\title{1}
\usage{
expandGRanges(x, width = 2000)
}
\arguments{
\item{x}{}
}
\value{
Returns
}
\description{
1
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

