% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotationQC_associatedFunctions.R
\name{extractTopN}
\alias{extractTopN}
\title{Prepares all the interproscan-associated files; the sceond extracts topN from the list}
\usage{
extractTopN(DF, column, N)
}
\arguments{
\item{a}{interproscan.gff prediction file}
}
\value{
Returns the full DF
}
\description{
Prepares all the interproscan-associated plots
}
