% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ezAllPairScatter}
\alias{ezAllPairScatter}
\title{Does scatter plots of all pairs}
\usage{
ezAllPairScatter(
  x,
  main = "",
  shrink = FALSE,
  xylab = NULL,
  lim = range(x, na.rm = TRUE),
  isPresent = NULL,
  types = NULL,
  pch = 16,
  colors = rainbow(ncol(types)),
  legendPos = "bottomright",
  cex.main = 1,
  cex = 1,
  ...
)
}
\arguments{
\item{x}{a matrix containing the data to plot.}

\item{main}{a character specifying the main title of the plots.}

\item{shrink}{a logical specifying whether to shrink the values to range.}

\item{xylab}{a character vector containing the axis labels. If it is NULL, \code{colnames(x)} will be used.}

\item{lim}{integers specifying the x and y limits of the plot.}

\item{isPresent}{a logical specifying whether the signal is present.}

\item{types}{a character vector containing the types.}

\item{pch}{an integer specifying the look of plotted points.}

\item{colors}{a character vector containing colors.}

\item{legendPos}{a character vector or integer specifying the position of the legend.}

\item{cex.main}{a numeric specifying the size of main titles relative to \code{cex}.}

\item{cex}{a numeric specifying the size of text and symbols.}

\item{...}{additional arguments to be passed further.}
}
\description{
Does scatter plots of all pairs.
}
\examples{
ezAllPairScatter(x=matrix(1:10,5))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
