% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01classes.R
\docType{class}
\name{EzApp-class}
\alias{EzApp}
\alias{EzApp-class}
\title{The R5 class representing a runnable app}
\description{
This reference class is the basis of all other apps that inherit from it. It sets the framework to run different apps.
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the app.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}
\section{Methods}{

\describe{
\item{\code{appExitAction(param, output, appName = "unknown")}}{Executes actions on exit of an application. This includes links to the output and possibly sending an e-mail.}

\item{\code{exitMail(text, subject, param)}}{Sends a report e-mail to the specified e-mail address. If not valid, an e-mail will be sent to the administrator if there was an error.}

\item{\code{outputLinks(output, param)}}{Returns urls, that are tagged as Links, specified in the output list together with relevant metadata.}

\item{\code{run(input, output, param)}}{Runs the app with the provided \code{input}, \code{output} and \code{param}.}
}}
\examples{
ds = EzDataset$new(file=system.file("extdata/yeast_10k/dataset.tsv", package="ezRun", mustWork = TRUE))
NULLApp = EzApp$new(runMethod=function(input, output, param){},name="NULLApp")
NULLApp$run(input=ds, output=ds, param=list(process_mode="DATASET"))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzDataset}}

\code{\link{checkFreeDiskSpace}}
}

