% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-AtacSeqQC.R
\docType{class}
\name{EzAppATACSeqQC-class}
\alias{EzAppATACSeqQC-class}
\alias{EzAppATACSeqQC}
\title{The R5 class representing a runnable app using \code{ezMethodATACSeqQC(input=NA, output=NA, param=NA, htmlFile="00index.html")()}}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\description{
Use this reference class to run QC on bamFiles from ATAC-Seq
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}

\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\section{Functions}{

\itemize{
  \item{\code{ezMethodATACSeqQC(input=NA, output=NA, param=NA, htmlFile="00index.html") }}:
  {The function to run this application.}
}
}

\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}
}
\author{
Opitz, Lennart

Rehrauer, Hubert

Schmid, Peter
}
