% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-mapping.r
\docType{class}
\name{EzAppBWA-class}
\alias{EzAppBWA}
\alias{EzAppBWA-class}
\title{The R5 class representing a runnable app using \code{ezMethodBWA()}}
\description{
This reference class is based on \code{EzApp()} and uses the function \code{ezMethodBWA()}.
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}

\code{\link{ezMethodBWA}}
}

