% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\docType{class}
\name{EzAppDEXSeqAnalysis-class}
\alias{EzAppDEXSeqAnalysis}
\alias{EzAppDEXSeqAnalysis-class}
\title{The R5 class representing a runnable app using \code{ezMethodDEXSeqCounting()}}
\description{
Use this reference class to run analysis on differential exon usage
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

