% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqCounting.R
\docType{class}
\name{EzAppDEXSeqCounting-class}
\alias{EzAppDEXSeqCounting}
\alias{EzAppDEXSeqCounting-class}
\title{The R5 class representing a runnable app using \code{ezMethodDEXSeqCounting()}}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}

\item{htmlFile}{a character, ending with .html, representing the file path of an html file.}
}
\description{
Use this reference class to run
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}
\section{Functions}{

\itemize{
  \item{\code{ezMethodDEXSeqCounting(input=NA, output=NA, param=NA): }}
  {The function to run this application.}
}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}
}

