% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt.R
\docType{class}
\name{EzAppExceRpt-class}
\alias{EzAppExceRpt-class}
\alias{EzAppExceRpt}
\title{exceRpt_smallRNA app}
\description{
Use this reference class to run exceRpt_smallRNA parameters.
}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\author{
Miquel Anglada Girotto
}
