% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt_report.R
\docType{class}
\name{EzAppExceRptReport-class}
\alias{EzAppExceRptReport-class}
\alias{EzAppExceRptReport}
\title{exceRpt_smallRNA report app}
\description{
Use this reference class to process exceRpt_smallRNA's output after execution and perform a QC.
}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\author{
Miquel Anglada Girotto
}
