% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-trim.R
\docType{class}
\name{EzAppFastp-class}
\alias{EzAppFastp-class}
\alias{EzAppFastp}
\title{EzAppFastp app}
\description{
fast read pre-processing.
}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\author{
Miquel Anglada Girotto
}
