% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-fastQC_10x.R
\docType{class}
\name{EzAppFastqc_10x-class}
\alias{EzAppFastqc_10x-class}
\alias{EzAppFastqc_10x}
\title{The R5 class representing a runnable app using \code{ezMethodFastQC(input=NA, output=NA, param=NA, htmlFile="00index.html")()}}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\description{
Use this reference class to run
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}

\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\section{Functions}{

\itemize{
  \item{\code{ezMethodFastQC(input=NA, output=NA, param=NA, htmlFile="00index.html") }}:
  {The function to run this application.}
}


\itemize{
  \item{\code{plotReadCountToLibConc(dataset, colname): }}
  {Plots \code{colname} from \code{dataset} against read counts in millions.}
  \item{\code{getQualityMatrix(inputFile): }}
  {Gets a quality count matrix from a fastq or gziped fastq.gz file with dimensions read quality and read length.}
  \item{\code{plotQualityMatrixAsHeatmap(qualMatrixList, isR2=FALSE, xScale=1, yScale=1): }}
  {Returns a png table of quality matrices interpreted as heatmaps.}
  \item{\code{plotQualityHeatmap(result, name=NULL, colorRange=c(0,sqrt(40)), colors=gray((1:256)/256), main=NULL, pngFileName=NULL, xScale=1, yScale=1): }}
  {Creates and returns the images used by \code{plotQualityMatrixAsHeatmap()}.}
}
}

\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
