% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-fastqscreen.R
\docType{class}
\name{EzAppFastqScreen-class}
\alias{EzAppFastqScreen-class}
\alias{EzAppFastqScreen}
\title{The R5 class representing a runnable app using \code{ezMethodFastqScreen(input=NA, output=NA, param=NA, htmlFile="00index.html")()}}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\description{
Use this reference class to run
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}

\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\section{Functions}{

\itemize{
  \item{\code{ezMethodFastqScreen(input=NA, output=NA, param=NA, htmlFile="00index.html") }}:
  {The function to run this application.}
}


\itemize{
  \item{\code{executeFastqscreenCMD(param, files): }}
  {Executes the fastqscreen command with given parameters on the input files.}
  \item{\code{collectFastqscreenOutput(dataset, files, resultFiles): }}
  {Collects the fastqscreen output after the result files have been obtained by \code{executeFastqscreenCMD()}.}
  \item{\code{executeBowtie2CMD(param, input): }}
  {Executes the bowtie2 command with given parameters on the input files.}
  \item{\code{collectBowtie2Output(param, dataset, countFiles): }}
  {Collects the bowtie2 output after the count files have been obtained by \code{executeBowtie2CMD()}.}
  \item{\code{fastqscreenReport(dataset, param, htmlFile="00index.html", fastqData, speciesPercentageTop): }}
  {Generates a report with plots and other information about the outcome of the run.}
}
}

\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
