% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-JunctionSeq.R
\docType{class}
\name{EzAppJunctionSeq-class}
\alias{EzAppJunctionSeq-class}
\alias{EzAppJunctionSeq}
\title{The R5 class representing a runnable app using \code{ezMethodRunQoRTs(input=NA, output=NA, param=NA, htmlFile="00index.html")()}}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\description{
Use this reference class to run JunctionSeq on bamFiles for alternative Splicing analysis
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}

\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\section{Functions}{

\itemize{
  \item{\code{ezMethodRunQoRTs(input=NA, output=NA, param=NA, htmlFile="00index.html") }}:
  {The function to run this application.}
}
}

\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}
}
\author{
Opitz, Lennart

Rehrauer, Hubert

Schmid, Peter
}
