% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ncpro.R
\docType{class}
\name{EzAppNcpro-class}
\alias{EzAppNcpro}
\alias{EzAppNcpro-class}
\title{Reference Class Constructor for ncPro App}
\description{
Reference Class Constructor for ncPro App

This reference class is based on \code{EzApp} and uses the function \code{ezMethodNcpro()}.
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}

\code{\link{ezMethodNcpro}}
}

