% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-PeakCombiner.R
\docType{class}
\name{EzAppPeakCombiner-class}
\alias{EzAppPeakCombiner-class}
\alias{EzAppPeakCombiner}
\title{}
\description{

}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

