% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-rsem.R
\docType{class}
\name{EzAppRSEM-class}
\alias{EzAppRSEM}
\alias{EzAppRSEM-class}
\title{The R5 class representing a runnable app using \code{ezMethodRSEM()}}
\description{
This reference class is based on \code{EzApp} and uses the function \code{ezMethodRSEM()}.
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}

\code{\link{ezMethodRSEM}}
}

