% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SCEVANApp.R
\docType{class}
\name{EzAppSCEVANApp-class}
\alias{EzAppSCEVANApp-class}
\alias{EzAppSCEVANApp}
\title{The R5 class representing a runnable app using \code{ezMethodSCEVANApp(input=NA, output=NA, param=NA, htmlFile="00index.html")()}}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\description{
Use this reference class to run SCEVAN CNV analysis
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}

\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\section{Functions}{

\itemize{
  \item{\code{ezMethodSCEVANApp(input=NA, output=NA, param=NA, htmlFile="00index.html") }}:
  {The function to run this application.}
}


\describe{
  \item{initialize}{Initializes the application using its specific defaults.}
}
}

\section{Slots}{

\describe{
  \item{runMethod}{The method to run (ezMethodSCEVANApp).}
  \item{name}{The name of the application (EzAppSCEVANApp).}
  \item{appDefaults}{Default parameters for the application.}
}
}

\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
