% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{ezBuildAttributeField}
\alias{ezBuildAttributeField}
\title{Builds an annotation attributes field}
\usage{
ezBuildAttributeField(x, format = "gtf")
}
\arguments{
\item{x}{a subset of an annotation data.frame to combine the attributes from.}

\item{format}{the annotation format to build \code{x} with. Either \code{"gtf"} or \code{"gff"}.}
}
\value{
Returns a gtf or gff attribute field.
}
\description{
Builds a compatible attributes field for either the gtf or gff annotation format.
}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param, system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE))
attrNew = ezBuildAttributeField(gtf[ , c("gene_id", "gene_name", "transcript_id", "tss_id")],"gtf")
gtf$attributes2 = attrNew
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
