% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{ezChromSizesFromVcf}
\alias{ezChromSizesFromVcf}
\title{Gets chromosome sizes from a VCF file}
\usage{
ezChromSizesFromVcf(vcfFile)
}
\arguments{
\item{vcfFile}{the VCF file to get the chromosome sizes from.}
}
\value{
Returns a names vector of the chromosome sizes.
}
\description{
Gets chromosome sizes from a VCF file by accessing its \code{contig} column.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
