% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-functions.R
\name{ezComputeBias}
\alias{ezComputeBias}
\title{compute and plot the GC content and gene-length associated bias in RNA-seq data}
\usage{
ezComputeBias(
  dsFile,
  dsName = NULL,
  param = NULL,
  qcSummaryDir = "/srv/GT/analysis/p2220/RNA-seq-bias-results",
  refBuildMap = getRefBuildMap(),
  minReadsPerSample = 30000,
  maxReadsPerSample = 5e+06,
  minReadsPerGene = 3,
  minPresentFraction = 0.2
)
}
\arguments{
\item{dsFile}{}

\item{dsName}{}

\item{param}{}

\item{qcSummaryDir}{}

\item{refBuildMap}{}

\item{minReadsPerSample}{}

\item{maxReadsPerSample}{}

\item{minReadsPerGene}{}

\item{minPresentFraction}{}
}
\value{

}
\description{
compute and plot the GC content and gene-length associated bias in RNA-seq data
}
\examples{
## needs the following modules loaded module add Tools/samtools QC/Trimmomatic QC/Flexbar Aligner/kallisto
options(error=recover)
require(ezRun)
dsFile = "/srv/gstore/projects/p3082/ISeq_20190425_iSeq35_o5404_DataDelivery/dataset.tsv"
ezComputeBias(dsFile)
}
