% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ezCorrelationPlot}
\alias{ezCorrelationPlot}
\title{Does a correlation plot}
\usage{
ezCorrelationPlot(
  z,
  cond = NULL,
  condOrder = NULL,
  main = "Correlation",
  labels = NULL,
  condLabels = NULL,
  plotLabels = nrow(z) < 100,
  colors = NULL
)
}
\arguments{
\item{z}{the data to plot.}

\item{cond}{a character vector specifying the conditions.}

\item{condOrder}{a sorted character vector specifying the order of the conditions.}

\item{main}{a character specifying the main title of the plots.}

\item{labels}{a character vector specifying axis labels.}

\item{condLabels}{a character vector specifying condition labels.}

\item{plotLabels}{a logical specifying whether to do plot labels.}

\item{colors}{a character vector containing colors.}
}
\description{
Does a correlation plot.
}
\examples{
ezCorrelationPlot(z=matrix(1:100,10))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
