% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rangesTools.r
\name{ezCountOverlaps}
\alias{ezCountOverlaps}
\alias{ezHasOverlaps}
\title{1}
\usage{
ezCountOverlaps(query, subject, ignoreStrand = FALSE, ...)

ezHasOverlaps(query, subject, ignore.strand = FALSE, ...)
}
\arguments{
\item{query}{}

\item{subject}{}

\item{ignoreStrand}{}

\item{...}{additional arguments to be passed to \code{\code{countOverlaps()}}.}
}
\value{
Returns
}
\description{
1
}
\section{Functions}{
\itemize{
\item \code{ezHasOverlaps}: 
}}
\examples{
query = GRanges(c("chr1", "chr1", "chr2"), IRanges(5000:5002,8000:8002), strand=c("+", "-", "+"))
subject = GRanges(c("chr2", "chr1", "chr3"), IRanges(4900:4902,5100:5102), strand=c("+", "-", "+"))
ezCountOverlaps(query, subject)
ezHasOverlaps(query,subject)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[GenomicRanges]{countOverlaps}}
}

