% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezCut}
\alias{ezCut}
\title{Convert numeric to factor}
\usage{
ezCut(x, breaks, prefix = NULL, labels = NULL)
}
\arguments{
\item{x}{a numeric vector to be cut.}

\item{breaks}{The values by which the vector should be cut. 2 or more are needed.}

\item{prefix}{adds a prefix to the level names.}

\item{labels}{use labels for the factors instead of displaying the ranges provided by breaks.}
}
\value{
Returns the numeric input as factors.
}
\description{
Cuts a numeric vector into factors using defined breaks.
}
\examples{
x = ezCut(1:10,breaks=c(2,5,7),prefix=letters[1:4])
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
