% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01classes.R
\docType{class}
\name{EzDataset-class}
\alias{EzDataset-class}
\alias{EzDataset}
\title{The R5 class representing a dataset}
\description{
Use this to create an object of the class EzDataset that contains the necessary fields for the input and output datasets.
}
\section{Fields}{

\describe{
\item{\code{file}}{a character representing the file path of the dataset's contents.}

\item{\code{meta}}{a data.frame containing the information about the samples.}

\item{\code{colNames}}{a character vector of the column names in \code{meta}.}

\item{\code{tags}}{a list with possible tags for \code{colNames}.}

\item{\code{isModified}}{whether the dataset has been modified.}
}}

\section{Methods}{

\describe{
\item{\code{columnHasTag(tag)}}{Checks each column whether its \code{tags} matches \code{tag}.}

\item{\code{getColumn(names)}}{Gets the column(s) selected with \code{names}.}

\item{\code{getFullPaths(name, checkExists = TRUE)}}{Gets the files in the named column prepended with the \code{dataRoot}.}

\item{\code{getLength()}}{Gets the number of samples.}

\item{\code{getNames()}}{Gets the row names.}

\item{\code{setColumn(name, values)}}{Sets the column selected with \code{name} to \code{values}. If \code{values} is \code{NULL} the column gets removed}

\item{\code{subset(samples)}}{Subsets the meta field keeping \code{samples} and generates a new EzDataset}
}}

\section{Functions}{

\itemize{
  \item{\code{ezTagListFromNames(names): }}{Gets the tags from \code{names} that are in the format [tag].}
}
}

\examples{
file = system.file("extdata/yeast_10k/dataset.tsv", package="ezRun", mustWork = TRUE)
dataRoot = system.file(package="ezRun", mustWork = TRUE)
ds = EzDataset$new(file=file, dataRoot=dataRoot)
ds$file
ds$meta
ds$getColumn("Read1")
ds$getFullPaths("Read1")
ds2 = ds$copy()
ds2$setColumn("Read1","replacement")
ds$columnHasTag("File")
ds$getNames()
ds$meta$"Genotype [Factor]"[1] = "a"
ds2 = EzDataset$new(meta = ds$meta, dataRoot=dataRoot)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
