% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\name{ezDesignFromDataset}
\alias{addReplicate}
\alias{ezConditionsFromDataset}
\alias{ezConditionsFromDesign}
\alias{ezDesignFromDataset}
\title{Gets the design from the dataset}
\usage{
ezDesignFromDataset(dataset, param = NULL)

ezConditionsFromDesign(design, maxFactors = 2)

ezConditionsFromDataset(dataset, param = NULL, maxFactors = 2)

addReplicate(x, sep = "_", repLabels = 1:length(x))
}
\arguments{
\item{dataset}{a data.frame to get the design from, usually the field meta of an EzDataset object.}

\item{param}{a list of parameters to specify the factors directly using \code{param$factors}.}
}
\value{
Returns the factorial design of the dataset.
}
\description{
Gets the design from the dataset or parameters, if specified.
}
\section{Functions}{
\itemize{
\item \code{ezConditionsFromDesign}: Gets the conditions from the design. Use \code{maxFactors} to limit the amount of factors.

\item \code{ezConditionsFromDataset}: A wrapper to get the conditions directly from the dataset.

\item \code{addReplicate}: Adds the names into the input character vector.
}}
\examples{
ds = EzDataset$new(file=system.file("extdata/yeast_10k/dataset.tsv", package="ezRun", mustWork = TRUE))
design = ezDesignFromDataset(ds$meta)
cond = ezConditionsFromDesign(design)
addReplicate(apply(design, 1, paste, collapse="_"))
addReplicate(cond)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

