% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{ezFlexTable}
\alias{ezFlexTable}
\title{Wrapper for \code{FlexTable()}}
\usage{
ezFlexTable(x, header = FALSE, ...)
}
\arguments{
\item{x}{a matrix or data.frame to turn into an object of the class FlexTable.}

\item{header}{a logical indicating whether to use a header for the table.}

\item{...}{additional arguments to be passed to \code{FlexTable}.}
}
\value{
Returns an object of the class FlexTable.
}
\description{
Wraps \code{FlexTable()} with defaults to remove the cell header and cell borders.
}
\examples{
ezFlexTable(cbind(a=1:5,b=11:15))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[ReporteRs]{FlexTable}}
}

