% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00defaults.R
\name{ezFrame}
\alias{ezFrame}
\title{Modified default of data.frame}
\usage{
ezFrame(..., row.names = NULL, check.rows = TRUE, check.names = FALSE,
  stringsAsFactors = FALSE)
}
\arguments{
\item{...}{additional arguments to be passed to \code{data.frame()}.}
}
\value{
Returns a data.frame.
}
\description{
Modified version of \code{data.frame()} with a different default.
}
\examples{
ezFrame(a=1:10,b=5)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

