% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezGeomean}
\alias{ezGeomean}
\title{Geometric mean}
\usage{
ezGeomean(x, ...)
}
\arguments{
\item{x}{a value, vector or matrix for which to calculate the geometric mean.}

\item{...}{additional arguments to be passed to \code{mean()}.}
}
\value{
Returns the geometric mean of \code{x}.
}
\description{
Calculates the geometric mean of a numeric argument.
}
\examples{
ezGeomean(1:10)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
