% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezGrepl}
\alias{ezGrepl}
\title{Matches patterns and returns a logical vector}
\usage{
ezGrepl(patterns, x, combine = "or", ...)
}
\arguments{
\item{patterns}{A value, vector or list to match the provided values with.}

\item{x}{the values to match.}

\item{combine}{whether just one value or all from \code{patterns} need to match to return TRUE:
\itemize{
 \item{"or"}{ only one value from \code{patterns} needs to match.}
 \item{"and"}{ all values from \code{patterns} need to match.}
}}

\item{...}{additional arguments to be passed to \code{grepl()}.}
}
\value{
Returns a logical vector containing the results of the tested pattern.
}
\description{
Searches for entries in \code{x} that match \code{patterns}.
}
\examples{
ezGrepl(3,1:40)
ezGrepl(c(2,4),1:100)
ezGrepl(c(2,4),1:100,combine="and")
}
\seealso{
\code{\link[base]{grepl}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
