% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R
\name{ezGroupGO}
\alias{ezGetGoByLevels}
\alias{ezGroupGO}
\title{1}
\usage{
ezGroupGO(selectedGenes, go2GeneList, onto = "CC", levels = 2:4,
  goSlim = NULL)

ezGetGoByLevels(onto, levels, goSlim = NULL)
}
\arguments{
\item{selectedGenes}{a character vector containing the selected genes.}

\item{go2GeneList}{1}

\item{onto}{a character representing the ontology to use.}

\item{levels}{1}

\item{goSlim}{1}
}
\value{
Returns the results of the GO analysis.
}
\description{
1
}
\section{Functions}{
\itemize{
\item \code{ezGetGoByLevels}: Gets the gene ontology by level.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

