% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsio.R
\name{ezHead}
\alias{ezHead}
\title{Writes the head of a file}
\usage{
ezHead(target = paste0(x, "_head"), x, n = 1000)
}
\arguments{
\item{target}{a character specifying the path of the output.}

\item{x}{a character specifying the path to a file to read from.}

\item{n}{an integer specifying the amount of lines to return.}
}
\value{
Returns the name of the output file.
}
\description{
Writes the head of a file into a newly created target file.
}
\examples{
description = system.file("DESCRIPTION", package="ezRun", mustWork=TRUE)
ezHead(x=description)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
