% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ezHeatmap}
\alias{ezHeatmap}
\title{Plots a heatmap}
\usage{
ezHeatmap(
  x,
  lim = c(-4, 4),
  colors = getBlueRedScale(),
  dendrogram = "none",
  margins = c(8, 6),
  cexCol = 1.1,
  Rowv = TRUE,
  Colv = TRUE,
  labCol = ezSplitLongLabels(colnames(x)),
  labRow = rownames(x),
  key = TRUE,
  lwid = c(1, 4),
  lhei = c(1, 5),
  ...
)
}
\arguments{
\item{x}{the data to plot.}

\item{lim}{two integers used for \code{breaks} argument passed to \code{heatmap.2()}.}

\item{colors}{a character vector containing colors.}

\item{dendrogram}{a character passed to \code{heatmap.2()}.}

\item{margins}{an integer, possibly modified, then passed to \code{heatmap.2()}.}

\item{cexCol}{an integer passed to \code{heatmap.2()}.}

\item{Rowv}{a logical passed to \code{heatmap.2()}.}

\item{Colv}{a logical passed to \code{heatmap.2()}.}

\item{labCol}{a character vector passed to \code{heatmap.2()}.}

\item{labRow}{a character vector, possibly modified, then passed to \code{heatmap.2()}.}

\item{key}{a logical passed to \code{heatmap.2()}. Will be set to FALSE if there is only one unique numeric in \code{x}.}

\item{...}{additional arguments to be passed to \code{heatmap.2()}.}
}
\description{
Plots a heatmap and adds a color key.
}
\examples{
x=matrix(1:100,10)
colnames(x) = letters[1:10]
ezHeatmap(x)
}
\seealso{
\code{\link[gplots]{heatmap.2}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
