% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reports.R
\name{ezImageFileLink}
\alias{ezImageFileLink}
\title{Gets an image link as html}
\usage{
ezImageFileLink(ezPlotter, file = NULL,
  mouseOverText = ezPlotter$mouseOverText, addPdfLink = TRUE, width = 480,
  height = 480, ...)
}
\arguments{
\item{ezPlotter}{an object of the class EzPlotter or inheriting from it.}

\item{file}{a character specifying the name of the image with a .png suffix.}

\item{mouseOverText}{a character specifying the text being displayed when mousing over the image.}

\item{addPdfLink}{a logical indicating whether to add a link on the image to a pdf version of itself.}

\item{width}{an integer specifying the width of each plot to create an image from.}

\item{height}{an integer specifying the height of each plot to create an image from.}

\item{...}{additional arguments to be passed to \code{ezPlotter}.}
}
\value{
Returns a character specifying a link to an image in html.
}
\description{
Gets an image link as html. Also plots and creates the image.
}
\examples{
imageLink = ezImageFileLink(EzPlotterIris$new())
theDoc = bsdoc(title = 'My document')
theDoc = addParagraph(theDoc, imageLink)
writeDoc(theDoc, "example.html")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzPlotter}}
}

