% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{ezInteractiveTable}
\alias{ezInteractiveTable}
\title{Saves an interactive table}
\usage{
ezInteractiveTable(table, tableLink)
}
\arguments{
\item{table}{a data.frame or table to create an interactive table from.}

\item{tableLink}{a character ending with .html representing the link to the interactive table}
}
\description{
Saves an interactive table accessible with the provided \code{tableLink}.
}
\examples{
tableLink = "exampleTable.html"
table = data.frame(a=1:100, b=201:300)
ezInteractiveTable(table, tableLink)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[DT]{datatable}}

\code{\link[DT]{saveWidget}}
}

