% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ezIntString}
\alias{ezIntString}
\title{Format an integer into a character using only digits}
\usage{
ezIntString(x)
}
\arguments{
\item{x}{an integer.}
}
\value{
Returns the input integer as a character.
}
\description{
convenience function that prevents scientific notation and surrounding white space. Such characters are need for system calls to commandline tools.
}
\examples{
ezIntString(4.5e7)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

