% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ezIsAbsolutePath}
\alias{ezIsAbsolutePath}
\title{Is x an absolute file path?}
\usage{
ezIsAbsolutePath(x)
}
\arguments{
\item{x}{a character representing a file path to check.}
}
\value{
Returns TRUE if \code{x} starts with a /.
}
\description{
Checks whether \code{x} is an absolute file path.
}
\examples{
ezIsAbsolutePath("/absolutepath")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
