% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00defaults.R
\name{ezIsSpecified}
\alias{ezIsSpecified}
\title{Check if a value is specified}
\usage{
ezIsSpecified(x)
}
\arguments{
\item{x}{usually a parameter to check.}
}
\value{
Returns FALSE or TRUE.
}
\description{
A value is specified if it is not \code{NULL}, it is not an empty vector, empty list or empty string.
If the value is a vector, the first element must be different from the empty string.
}
\examples{
ezIsSpecified(5)
ezIsSpecified(c("this","is"))
ezIsSpecified(c("","this isn't"))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

