% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{ezJobStart}
\alias{ezJobStart}
\title{Writes start time of job}
\usage{
ezJobStart(jobName, con = stderr())
}
\arguments{
\item{jobName}{a character representing the job name.}

\item{con}{the connection to the file to write in.}
}
\value{
Returns a list containing \code{jobName} and the running time of R.
}
\description{
Writes the starting date and time of \code{jobName} into a separate file.
}
\examples{
con = file("example.txt", "w")
ezJobStart("a job", con=con)
close(con)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
