% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ezLegend}
\alias{ezLegend}
\title{Plots a legend}
\usage{
ezLegend(legend, fill = NULL, title = "Legend")
}
\arguments{
\item{legend}{passed to \code{legend()}.}

\item{fill}{passed to \code{legend()}.}

\item{title}{passed to \code{legend()}.}
}
\description{
Plots only a legend, removing all the other plot elements.
}
\examples{
ezLegend(1:3)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[graphics]{legend}}
}

