% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezLoadRobj}
\alias{ezLoadRobj}
\title{Load R objects from various serialized formats}
\usage{
ezLoadRobj(filePath, nthreads = 1L)
}
\arguments{
\item{filePath}{Character string specifying the path to the R object file.}

\item{nthreads}{Integer specifying how many threads to use for qs/qs2 loading.
Default is 1.}
}
\value{
The R object stored in the file.
}
\description{
This function loads R objects from files with different extensions (.qs2, .qs, .rds)
by automatically detecting the file format based on the extension.
}
\details{
The function checks if the required packages are available (qs2, qs)
and loads the appropriate package for the file extension.
Supported formats:
\itemize{
  \item .qs2 - Uses qs2::qread() for fastest serialization
  \item .qs - Uses qs::qread() for fast serialization
  \item .rds - Uses base::readRDS() for standard R serialization
}
}
\examples{
\dontrun{
# Load a qs2 file
my_object <- ezLoadRobj("path/to/data.qs2", nthreads=4L)

# Load a qs file
my_object <- ezLoadRobj("path/to/data.qs", nthreads=3L)

# Load an rds file
my_object <- ezLoadRobj("path/to/data.rds")
}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
