% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{ezLog}
\alias{ezLog}
\title{Wrapper for futile.logger}
\usage{
ezLog(message, level = "info")
}
\arguments{
\item{message}{the message to log.}

\item{level}{the log level (trace, debug, info, warn, error, fatal).}
}
\description{
Logs a message at a specified level using futile.logger.
}
\examples{
ezLog("This is an info message")
ezLog("This is a warning message", level="warn")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
