% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{ezLogChangeLevel}
\alias{ezLogChangeLevel}
\title{Change log level}
\usage{
ezLogChangeLevel(level = "info")
}
\arguments{
\item{level}{the log level (fatal, error, warn, success, info, debug, trace).}
}
\description{
Changes the log level for logger. Messages below this level
will not be shown. 'fatal' is the highest level, 'trace' the lowest.
}
\examples{
ezLogChangeLevel("debug")
ezLog("This is a debug message", level="debug")
ezLogChangeLevel("info")
ezLog("This is an info message", level="debug") # will not be shown
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
