% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezLogmeanNorm}
\alias{ezLogmeanNorm}
\alias{ezLogmeanScalingFactor}
\title{Scales columns of a matrix to logarithmic mean}
\usage{
ezLogmeanNorm(x, use = NULL, target = NULL, presentFlag = NULL)

ezLogmeanScalingFactor(x, use = NULL, target = NULL, presentFlag = NULL)
}
\arguments{
\item{x}{the matrix to scale.}

\item{use}{a logical vector defining which rows to use.}

\item{target}{a value or vector defining the means by which columns should be scaled. The default will use the overall logarithmic mean of the matrix for each column.}

\item{presentFlag}{a binary matrix with the same size as \code{x} which indicates if a values is considered as measured correctly.}
}
\value{
Returns a matrix with columns normalized to a logarithmic mean/means.
}
\description{
Columns of the matrix will be scaled to an overall logarithmic mean or to a defined target.
}
\section{Functions}{
\itemize{
\item \code{ezLogmeanScalingFactor()}: Calculates the scaling factor for the main function.

}}
\examples{
m1 = matrix(1:20,5)
m2 = ezLogmeanNorm(m1)
m3 = ezLogmeanNorm(m1,target=10)
m4 = ezLogmeanNorm(m1,use=c(TRUE,FALSE))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
