% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{ezMail}
\alias{ezMail}
\title{Mails information, the nodename and working directory}
\usage{
ezMail(text = "done", subject = "r-mail", to = "")
}
\arguments{
\item{text}{a message to send.}

\item{subject}{the subject of the email.}

\item{to}{the recipient of the email.}
}
\description{
Mails the nodename and working directory to a specified email-address. A text message and the subject can also be defined.
}
\examples{
\dontrun{
ezMail(to="example@example.com")
}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
