% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezMatrix}
\alias{ezMatrix}
\title{Creates a matrix}
\usage{
ezMatrix(x, rows = NULL, cols = NULL, dim = NULL)
}
\arguments{
\item{x}{a vector containing the matrix elements}

\item{rows}{a vector, whose length defines the number of rows.}

\item{cols}{a vector, whose length defines the number of columns.}

\item{dim}{a vector of length 2 defining the dimensions of the matrix.}
}
\value{
Returns a matrix with \code{rows} and \code{cols} defining the names or using generic ones, if \code{dim} is used instead.
}
\description{
Either use the rows and columns to define the matrix or the dimensions.
}
\examples{
ezMatrix(1,rows=1:4,cols=1:3)
ezMatrix(3:6,dim=c(4,6))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
