% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{ezMclapply}
\alias{ezMclapply}
\title{Parallel version of \code{lapply()}}
\usage{
ezMclapply(
  x,
  FUN,
  ...,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = min(length(x), ezThreads())
)
}
\arguments{
\item{x}{a list to apply the function to.}

\item{FUN}{the function to apply to each list element.}

\item{...}{additional arguments to be passed to \code{lapply() and mclapply()}.}

\item{mc.preschedule}{a logical passed to \code{mclapply()}.}

\item{mc.set.seed}{a logical passed to \code{mclapply()}.}

\item{mc.silent}{a logical passed to \code{mclapply()}.}

\item{mc.cores}{an integer passed to \code{mclapply()}.}
}
\value{
Returns a list of the same length as \code{x} with \code{FUN} applied to its elements.
}
\description{
This function is a modified version of \code{mclapply()} of the parallel package and allows a parallel usage of lapply.
}
\examples{
l1 = list(a=1:3, b=c(2,5), c=4:8)
ezMclapply(l1,sum)
}
\seealso{
\code{\link[parallel]{mclapply}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
