% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsPlots.R
\name{ezMdsPlot}
\alias{ezMdsPlot}
\title{Plots the multi dimensional scaling}
\usage{
ezMdsPlot(signal, sampleColors, main)
}
\arguments{
\item{signal}{the signal to use for plotting.}

\item{sampleColors}{a character vector containing colors in hex format.}

\item{main}{a character representing the plot title.}
}
\description{
Plots the multi dimensional scaling.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
