% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{ezMedianNorm}
\alias{ezMedianNorm}
\alias{getMedianScalingFactor}
\title{Scales columns of a matrix to median}
\usage{
ezMedianNorm(x, use = NULL, target = NULL, presentFlag = NULL)

getMedianScalingFactor(x, use = NULL, target = NULL, presentFlag = NULL)
}
\arguments{
\item{x}{the matrix to scale.}

\item{use}{a logical vector defining which rows to use.}

\item{target}{a value or vector defining the median(s) by which columns should be scaled. The default will use the overall median of the matrix for each column.}

\item{presentFlag}{a binary matrix with the same size as \code{x} which indicates if a values is considered as measured correctly.}
}
\value{
Returns a matrix with columns normalized to a median/medians.
}
\description{
Columns of the matrix will be scaled to an overall median or to a defined target.
}
\section{Functions}{
\itemize{
\item \code{getMedianScalingFactor}: Calculates the scaling factor for the main function.
}}
\examples{
m1 = matrix(1:20,5)
ezMedianNorm(m1)
ezMedianNorm(m1,target=10)
ezMedianNorm(m1,use=c(TRUE,FALSE))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

