% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-countOverlaps.r
\name{ezMethodCountOverlaps}
\alias{.countPairedBamHits}
\alias{countBamHitsSingleChrom}
\alias{countNonredundant}
\alias{countPairedBamHitsSingleChrom}
\alias{countTranscriptBam}
\alias{ezMethodCountOverlaps}
\alias{getFeatureCounts}
\alias{getTargetRanges}
\title{The Count Overlaps method to run an app}
\usage{
ezMethodCountOverlaps(input = NULL, output = NULL, param = NULL)

.countPairedBamHits(input = NULL, output = NULL, param = NULL)

countPairedBamHitsSingleChrom(chr, bamFile = NULL, param = NULL,
  gff = NULL)

countNonredundant(bamFile, param = param, gff = gff)

countBamHitsSingleChrom(chr, bamFile = NULL, param = NULL, gff = NULL)

getTargetRanges(gff, param, chrom = NULL)

getFeatureCounts(chrom, gff, reads, param)

countTranscriptBam(bamFile, isFirstMateRead = NA, isSecondMateRead = NA,
  isUnmappedQuery = FALSE, isProperPair = NA)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\value{
Returns "Success" if the method runs until the end. Also creates output files.
}
\description{
This is the method to run the app. Actually running it should be done using the \code{run} method of the app.
}
\section{Functions}{
\itemize{
\item \code{.countPairedBamHits}: Counts the paired BAM hits.

\item \code{countPairedBamHitsSingleChrom}: Counts the paired BAM hits for a single chromosome.

\item \code{countNonredundant}: Counts the non-redundant overlaps.

\item \code{countBamHitsSingleChrom}: Counts the BAM hits for a single chromosome.

\item \code{getTargetRanges}: Gets the target ranges depending on \code{param$featureLevel}.

\item \code{getFeatureCounts}: Gets the feature counts from the target ranges.

\item \code{countTranscriptBam}: Counts the transcripts from a BAM file.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzDataset}}

\code{\link{EzAppCountOverlaps}}
}

