% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\name{ezMethodDEXSeqAnalysis}
\alias{ezMethodDEXSeqAnalysis}
\title{RunMethod for reference class EzAppDEXSeqAnalysis}
\usage{
ezMethodDEXSeqAnalysis(input = NA, output = NA, param = NA)
}
\arguments{
\item{input}{either EzDataSet reference object or path to input dataset file}

\item{output}{}

\item{param}{}
}
\description{
Differential exon usage is assessed using the function \code{DEXSeq}
from the DEXSeq package.
}

