% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqCounting.R
\name{ezMethodDEXSeqCounting}
\alias{ezMethodDEXSeqCounting}
\title{Run-method for ezApp EzAppDEXSeqCounting}
\usage{
ezMethodDEXSeqCounting(input = NA, output = NA, param = NA)
}
\description{
Run-method for ezApp EzAppDEXSeqCounting
}

