% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt.R
\name{ezMethodExceRpt}
\alias{ezMethodExceRpt}
\title{execute exceRpt_smallRNA.}
\usage{
ezMethodExceRpt(input = NA, output = NA, param = NA)
}
\arguments{
\item{input}{ezDataFrame()}

\item{output}{ezDataFrame()}

\item{param}{list() exceRpt_smallRNA and environment parameters.}
}
\description{
R wrapper to execute exceRpt_smallRNA pipeline to count smallRNAs in bulk RNA-seq experiments.
}
\author{
Miquel Anglada Girotto
}
