% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt_report.R
\name{ezMethodExceRptReport}
\alias{ezMethodExceRptReport}
\title{execute exceRpt_smallRNA's output processing.}
\usage{
ezMethodExceRptReport(input = NA, output = NA, param = NA)
}
\arguments{
\item{input}{ezDataFrame()}

\item{output}{ezDataFrame()}

\item{param}{list() environment parameters.}
}
\description{
create QC plots and count smallRNAs detected after executing ezMethodExceRpt().
}
\author{
Miquel Anglada Girotto
}
