% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-trim.R
\name{ezMethodFastpTrim}
\alias{ezMethodFastpTrim}
\title{Trims input reads using fastp (Chen et al. 2018)}
\usage{
ezMethodFastpTrim(input = NA, output = NA, param = NA)
}
\description{
Trims input reads. The trimming happens in the following order:
\itemize{
  \item{global trimming fron and tail}
  \item{quality window}
  \item{avg quality}
  \item{adapter}
  \item{trim polyG (by default, only for NextSeq and NovaSeq)}
  \item{fixed trimming}
  \item{length filtering}
}
This means if you specify a fixed trimming this comes on top of the adapter trimming
}
