% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-fastQC.R
\name{ezMethodFastQC}
\alias{ezMethodFastQC}
\alias{getQualityMatrix}
\alias{plotQualityHeatmap}
\alias{plotQualityMatrixAsHeatmap}
\alias{plotReadCountToLibConc}
\title{The Fast QC method to run an app}
\usage{
ezMethodFastQC(input = NA, output = NA, param = NA,
  htmlFile = "00index.html")

plotReadCountToLibConc(dataset, colname)

getQualityMatrix(inputFile)

plotQualityMatrixAsHeatmap(qualMatrixList, isR2 = FALSE, xScale = 1,
  yScale = 1)

plotQualityHeatmap(result, name = NULL, colorRange = c(0, sqrt(40)),
  colors = gray((1:256)/256), main = NULL, pngFileName = NULL,
  xScale = 1, yScale = 1)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\value{
Returns "Success" if the method runs until the end. Also creates output files.
}
\description{
This is the method to run the app. Actually running it should be done using the \code{run} method of the app.
}
\section{Functions}{
\itemize{
\item \code{plotReadCountToLibConc}: Plots \code{colname} from \code{dataset} against read counts in millions.

\item \code{getQualityMatrix}: Gets a quality count matrix from a fastq or gziped fastq.gz file with dimensions read quality and read length.

\item \code{plotQualityMatrixAsHeatmap}: Returns a png table of quality matrices interpreted as heatmaps.

\item \code{plotQualityHeatmap}: Creates and returns the images used by \code{plotQualityMatrixAsHeatmap()}.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzDataset}}

\code{\link{EzAppFastqc}}
}

