% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-fastqscreen.R
\name{ezMethodFastqScreen}
\alias{collectBowtie2Output}
\alias{collectFastqscreenOutput}
\alias{executeBowtie2CMD}
\alias{executeFastqscreenCMD}
\alias{ezMethodFastqScreen}
\alias{fastqscreenReport}
\title{The Fastq Screen method to run an app}
\usage{
ezMethodFastqScreen(input = NA, output = NA, param = NA,
  htmlFile = "00index.html")

executeFastqscreenCMD(param, files)

collectFastqscreenOutput(dataset, files, resultFiles)

executeBowtie2CMD(param, files)

collectBowtie2Output(param, dataset, countFiles)

fastqscreenReport(dataset, param, htmlFile = "00index.html", fastqData,
  speciesPercentageTop)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}

\item{htmlFile}{a character, ending with .html, representing the file path of an html file.}
}
\value{
Returns "Success" if the method runs until the end. Also creates output files.
}
\description{
This is the method to run the app. Actually running it should be done using the \code{run} method of the app.
}
\section{Functions}{
\itemize{
\item \code{executeFastqscreenCMD}: Executes the fastqscreen command with given parameters on the input files.

\item \code{collectFastqscreenOutput}: Collects the fastqscreen output after the result files have been obtained by \code{executeFastqscreenCMD()}.

\item \code{executeBowtie2CMD}: Executes the bowtie2 command with given parameters on the input files.

\item \code{collectBowtie2Output}: Collects the bowtie2 output after the count files have been obtained by \code{executeBowtie2CMD()}.

\item \code{fastqscreenReport}: Generates a report with plots and other information about the outcome of the run.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzDataset}}

\code{\link{EzAppFastqScreen}}
}

