% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-SCEVANApp.R
\name{ezMethodSCEVANApp}
\alias{ezMethodSCEVANApp}
\title{SCEVAN CNV Analysis Method}
\usage{
ezMethodSCEVANApp(
  input = NA,
  output = NA,
  param = NA,
  htmlFile = "00index.html"
)
}
\arguments{
\item{input}{an object of class EzDataset with input metadata}

\item{output}{an object of class EzDataset with output metadata}

\item{param}{a list of parameters including:
\itemize{
  \item{name: }{Output name for the analysis}
  \item{dataRoot: }{Root directory for data files}
  \item{cellTypeColumn: }{Metadata column containing cell type annotations}
  \item{referenceCellTypes: }{Cell types to use as normal reference}
  \item{sampleColumn: }{Metadata column containing sample identifiers}
  \item{organism: }{Organism: "human" or "mouse"}
  \item{cores: }{Number of cores for parallel processing}
  \item{subclones: }{Enable subclone analysis}
  \item{betaVega: }{Segmentation granularity parameter}
}}

\item{htmlFile}{the name of the HTML output file (default: "00index.html")}
}
\value{
Returns "Success" if analysis completes successfully
}
\description{
Main processing function for SCEVAN CNV inference from scRNA-seq data
}
