% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-mapping.r
\name{ezMethodSTAR}
\alias{ezMethodSTAR}
\title{The STAR method to run an app}
\usage{
ezMethodSTAR(input = NA, output = NA, param = NA)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\value{
Returns "Success" if the method runs until the end. Also creates output files.
}
\description{
This is the method to run the app. Actually running it should be done using the \code{run} method of the app.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzDataset}}

\code{\link{EzAppSTAR}}

\code{\link{getSTARReference}}

\code{\link{ezMethodTrim}}
}

