% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezNorm}
\alias{ezNorm}
\title{Normalization}
\usage{
ezNorm(x, method = "none", presentFlag = NULL)
}
\arguments{
\item{x}{the matrix to normalize.}

\item{method}{which normalization method to use. The default does not normalize. Possible methods:
\itemize{
 \item{"none"}
 \item{"quantile"}
 \item{"logMean"}
 \item{"median"}
 \item{"vsn"}
}}

\item{presentFlag}{a binary matrix with the same size as \code{x} which indicates if a values is considered as measured correctly. Default: NULL}
}
\value{
Returns the modified value.
}
\description{
Normalizes a matrix according to the provided method; normalization is performed column-wise.
}
\examples{
x = ezNorm(runif(100))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
