% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00defaults.R
\name{ezParam}
\alias{ezParam}
\alias{parseOptions}
\title{Gets and processes all parameters}
\usage{
ezParam(userParam = list(), globalDefaults = EZ_PARAM_DEFAULTS,
  appDefaults = ezFrame())

parseOptions(optString)
}
\arguments{
\item{userParam}{a list of parameters defined by the user.}

\item{globalDefaults}{global package defaults. These should normally not be changed.}

\item{appDefaults}{a data.frame containing application specific defaults.}
}
\value{
Returns the complete list of parameters used by the package.
}
\description{
Gets and processes all parameters. Removes any duplicates and recognizes the correct datatype.
}
\section{Functions}{
\itemize{
\item \code{parseOptions}: Used to parse additional options specified in \code{userParam$specialOptions}.
}}
\examples{
ezParam()
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

