% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ezParamFromTsv}
\alias{ezParamFromTsv}
\title{Create ezParam from parameters.tsv}
\usage{
ezParamFromTsv(tsvPath)
}
\arguments{
\item{tsvPath}{Path to the parameters.tsv file on disk}
}
\value{
Returns an ezParam object
}
\description{
Read in a parameters.tsv file from disc into an ezParam object
}
\examples{
params <- ezParamFromTsv("parameters.tsv")
params$cores
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
