% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03plots.R
\docType{class}
\name{EzPlotter-class}
\alias{EzPlotter}
\alias{EzPlotter-class}
\title{The R5 class representing the basic plotter}
\description{
This class is usually not used, but all other plotters inherit from it.
}
\section{Fields}{

\describe{
\item{\code{name}}{a character specifying the name of the plotter instance.}

\item{\code{data}}{a data.frame, matrix, table or another sensible data object to create plots from.}

\item{\code{helpText}}{a character specifying the help text of the plotter.}

\item{\code{mouseOverText}}{a character specifying the mouseOver text of images created with the plotter.}
}}
\section{Methods}{

\describe{
\item{\code{plot(...)}}{Plots \code{data} with the default plot function from the graphics package.}

\item{\code{plotPdf(file = NULL, width = 480, height = 480, ...)}}{Creates a .pdf file of a plot.}

\item{\code{plotPng(file = NULL, width = 480, height = 480, ...)}}{Creates a .png file of a plot.}
}}
\examples{
EzPlotter$new()
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

