% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentAnalysis.R
\name{ezPresentFlags}
\alias{ezPresentFlags}
\title{Which values to keep?}
\usage{
ezPresentFlags(x, presentFlag = NULL, param = NULL, isLog = FALSE)
}
\arguments{
\item{x}{a matrix with row and column names.}

\item{presentFlag}{a binary matrix with the same size as \code{x} which indicates if a values is considered as measured correctly.}

\item{param}{contains a threshold and whether it should be used.
\itemize{
 \item{sigThresh}{ the threshold...}
 \item{useSigThresh}{ ...and whether it should be used.}
}}

\item{isLog}{a logical indicating whether the \code{log2()} of the threshold should be used.}
}
\value{
Returns a logical matrix.
}
\description{
Returns TRUE for values that are considered as measured correctly and that are over a threshold, if one is used.
}
\examples{
m1 = matrix(1:20,5)
rownames(m1) = letters[1:5]
colnames(m1) = letters[6:9]
ezPresentFlags(m1, param=list(useSigThresh=TRUE, sigThresh=10))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
