% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezQuantileNorm}
\alias{ezQuantileNorm}
\title{Quantile Normalization}
\usage{
ezQuantileNorm(x)
}
\arguments{
\item{x}{the matrix to normalize.}
}
\value{
Returns the normalized matrix.
}
\description{
A convenience call to the corresponding method from the preprocessCore package that keeps row and column names of the input matrix.
}
\examples{
m1 = matrix(1:20,4)
m2 = ezQuantileNorm(m1)
}
\seealso{
\code{\link[preprocessCore]{normalize.quantiles}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
