% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamio.r
\name{ezReadGappedAlignments}
\alias{ezReadGappedAlignments}
\title{Reads gapped alignments from bam}
\usage{
ezReadGappedAlignments(
  bamFile,
  seqname = NULL,
  start = NULL,
  end = NULL,
  strand = "*",
  tag = character(0),
  what = character(0),
  use.names = TRUE,
  isFirstMateRead = NA,
  isSecondMateRead = NA,
  isUnmappedQuery = FALSE,
  isProperPair = NA,
  isSecondaryAlignment = NA,
  minMapQuality = 0,
  tagFilter = list(),
  keepMultiHits = TRUE
)
}
\arguments{
\item{bamFile}{a character representing the bam file path.}

\item{seqname}{an optional character vector to keep only the specified sequence names in the returned reads.}

\item{start}{an optional integer vector to limit the start position of the range of returned reads.}

\item{end}{an optional integer vector to limit the start position of the range of returned reads.}

\item{strand}{a character specifying which strand to use (+, - or *).}

\item{tag}{passed further to \code{ScanBamParam()}.}

\item{what}{passed further to \code{ScanBamParam()}.}

\item{use.names}{passed further to \code{readGAlignments()}.}

\item{isFirstMateRead}{passed further to \code{scanBamFlag()}.}

\item{isSecondMateRead}{passed further to \code{scanBamFlag()}.}

\item{isUnmappedQuery}{passed further to \code{scanBamFlag()}.}

\item{isProperPair}{passed further to \code{scanBamFlag()}.}

\item{minMapQuality}{an integer specifying the minimal mapping quality to use.}

\item{keepMultiHits}{a logical specifying whether to keep multiple hits.}
}
\value{
Returns an object of the class GAlignments.
}
\description{
Reads gapped alignments from a bam file.
}
\examples{
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
ga = ezReadGappedAlignments(bamFile)
}
\seealso{
\code{\link[Rsamtools]{scanBam}}

\code{\link[Rsamtools]{ScanBamParam}}

\code{\link[GenomicAlignments]{readGAlignments}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
