% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{ezReadGff}
\alias{ezReadGff}
\alias{ezWriteGff}
\title{Reads an annotation table from a file.}
\usage{
ezReadGff(gffFile, nrows = -1)

ezWriteGff(x, file)
}
\arguments{
\item{gffFile}{the gtf or gff file to read the information from.}

\item{nrows}{an integer specifying the maximum number of rows to read in.}

\item{x}{an annotation matrix or data.frame.}

\item{file}{a character representing the connection to write the table to.}
}
\value{
Returns a data.frame containing the annotation information.
}
\description{
Reads an annotation table from a gtf or gff file into a data.frame.
}
\section{Functions}{
\itemize{
\item \code{ezWriteGff()}: Writes a gtf or gff annotation table passed by with \code{x} to the connection \code{file}, keeping only the original columns.

}}
\examples{
gtf = ezReadGff(system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE))
ezWriteGff(gtf,"newgtf")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
