% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamio.r
\name{ezReadPairedAlignments}
\alias{ezMergeLeftRightAlignments}
\alias{ezReadPairedAlignments}
\title{Reads paired alignments from bam}
\usage{
ezReadPairedAlignments(bamFile, seqname = NULL, start = NULL, end = NULL,
  strand = "*", tag = c("NH"), keepUnpaired = "both", fillGap = "N",
  minMapQuality = 0, keepMultiHits = TRUE)

ezMergeLeftRightAlignments(gaLeft, gaRight, fillGap = "N")
}
\arguments{
\item{bamFile}{a character representing the file path to the bam file to read.}

\item{seqname}{an optional character vector to keep only the specified sequence names.}

\item{start}{passed further to \code{ezReadGappedAlignments()}.}

\item{end}{passed further to \code{ezReadGappedAlignments()}.}

\item{strand}{a character specifying which strand to use (+, - or *).}

\item{tag}{passed further to \code{ezReadGappedAlignments()}.}

\item{keepUnpaired}{a character indicating which aligned but unpaired reads to keep. Possible options: "none", "first", "second", "both".}

\item{fillGap}{a character to fill into the gaps produced by \code{ezMergeLeftRightAlignments()}.}

\item{minMapQuality}{passed further to \code{ezReadGappedAlignments()}.}

\item{keepMultiHits}{passed further to \code{ezReadGappedAlignments()}.}
}
\value{
Returns an object of the class GAlignments.
}
\description{
Reads paired alignments from a bam file.
}
\section{Functions}{
\itemize{
\item \code{ezMergeLeftRightAlignments}: Merges left and right alignments and fills gaps with \code{fillGap}.
}}
\examples{
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
ezReadPairedAlignments(bamFile)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{ezReadGappedAlignments}}
}

