% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01classes.R
\docType{class}
\name{EzRef-class}
\alias{EzRef}
\alias{EzRef-class}
\alias{[,EzRef,ANY,ANY-method}
\alias{initialize,EzRef-method}
\title{The S4 class representing a collection of information about the reference build}
\usage{
\S4method{initialize}{EzRef}(.Object, param = list())

\S4method{[}{EzRef,ANY,ANY}(x, i)
}
\description{
Objects of this class are a collection of references derived from a list of parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: Initializes the slots of EzRef. It will fill also try to specify some fields and if necessary get full file paths.

\item \code{[}: Allows selection with square brackets [ ].
}}
\section{Slots}{

\describe{
\item{\code{refBuild}}{a character specifying the file path to the reference build.}

\item{\code{refBuildName}}{a character specifying the name of the reference build.}

\item{\code{refBuildDir}}{a character specifying the directory of the reference build.}

\item{\code{refIndex}}{a character specifying the location of the index that is used in the alignment.}

\item{\code{refFeatureFile}}{a character specifying the file path to the annotation feature file (.gtf).}

\item{\code{refAnnotationFile}}{a character specifying the file path to the annotation file (.txt).}

\item{\code{refFastaFile}}{a character specifying the file path to the fasta file.}

\item{\code{refChromDir}}{a character specifying the file path to the directory of the chromosome information.}

\item{\code{refChromSizesFile}}{a character specifying the file path to the file containing the chromosome sizes.}
}}
\examples{
EzRef(param=list(refBuild="foo"))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

