% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezReplicateNumber}
\alias{ezReplicateNumber}
\title{Count how often a value has been seen before}
\usage{
ezReplicateNumber(x)
}
\arguments{
\item{x}{a vector with discrete values}
}
\value{
Returns a vector of the same length as the input. If the value at an element is n, then this means, in the original value was the nth occurrence.
}
\description{
This can be used to get replicate identifiers. For each value unique value in the input it counts incrementally how often it occurs
}
\examples{
x = c("a", "c", "a", "b", "c")
ezReplicateNumber(x)
}
