% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03results.R
\docType{class}
\name{EzResult-class}
\alias{EzResult-class}
\alias{EzResult}
\title{The object that represents a result of an expression analysis}
\description{
The object that represents a result of an expression analysis
}
\section{Fields}{

\describe{
\item{\code{param}}{\code{EzParam} object that was used during the result generation}

\item{\code{rawData}}{list with the rawData that was used during the result generation}

\item{\code{result}}{list holding the actual results, in particular, the log2 ration and the p-value}
}}


\section{Functions}{

\itemize{
  \item{\code{saveToFile(file): }}{Saves the object to the given filepath. Saved files can be loaded with \code{EzResult$new(file="mystoredResult.RData"}}
}
}

\examples{
deResult = EzResult$new()
deResult = EzResult$new(param=list(p1="p1", p2="p2"), rawData=list(a=0, b=1), result=list(u=0, v=1))
rdFile = tempfile(fileext = ".RData")
deResult$saveToFile(file = rdFile)
deResult2 = EzResult$new(file=rdFile)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
