\name{ezRun-package}
\alias{ezRun-package}
\alias{ezRun}
\docType{package}
\title{a meta-package that integrates NGS related R-packages and command line tools}
\description{
Designed for NGS data analysis, ezRun provides a collection of ezApps that are  from commandline tools and generate comprehensive reports. ezApps 
}
\details{
\tabular{ll}{
Package: \tab ezRun\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-27\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Hubert Rehrauer

Maintainer: Hubert Rehrauer <Hubert.Rehrauer@fgcz.uzh.ch>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
\code{\link[Biostrings:XString-class]{Biostrings}}
}
\examples{
ls() ## this is an example
}
