% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{ezScaleColumns}
\alias{ezScaleColumns}
\title{Scales columns of a matrix}
\usage{
ezScaleColumns(x, scaling)
}
\arguments{
\item{x}{the matrix to scale.}

\item{scaling}{a vector containing the scale for each column of \code{x}.}
}
\value{
Returns a matrix with scaled columns.
}
\description{
The columns will be scaled by multiplying them with the \code{scaling}.
}
\examples{
m1 = matrix(1:20,5)
ezScaleColumns(m1,1:4)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

