% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bamio.r
\name{ezScanBam}
\alias{ezScanBam}
\title{Scans a bam file}
\usage{
ezScanBam(bamFile, seqname = NULL, start = NULL, end = NULL,
  strand = "*", tag = character(0), what = scanBamWhat(),
  isFirstMateRead = NA, isSecondMateRead = NA, isUnmappedQuery = FALSE,
  isProperPair = NA, countOnly = FALSE)
}
\arguments{
\item{bamFile}{a character representing the file path to the bam file to scan.}

\item{seqname}{an optional character vector to keep only the specified sequence names in the returned reads.}

\item{start}{an optional integer vector to limit the start position of the range of returned reads.}

\item{end}{an optional integer vector to limit the start position of the range of returned reads.}

\item{strand}{a character specifying which strand to use (+, - or *).}

\item{tag}{passed further to \code{ScanBamParam()}.}

\item{what}{passed further to \code{ScanBamParam()}.}

\item{isFirstMateRead}{passed further to \code{scanBamFlag()}.}

\item{isSecondMateRead}{passed further to \code{scanBamFlag()}.}

\item{isUnmappedQuery}{passed further to \code{scanBamFlag()}.}

\item{isProperPair}{passed further to \code{scanBamFlag()}.}

\item{countOnly}{a logical specifying whether only counts should be returned or the whole scan of the bam file.}
}
\value{
Returns a list representing a scanned bam file.
}
\description{
Scans a bam file with the option to select only a part of the output.
}
\examples{
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
ezScanBam(bamFile)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[Rsamtools]{scanBam}}

\code{\link[Rsamtools]{ScanBamParam}}
}

