% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ezScatter}
\alias{ezScatter}
\alias{ezXYScatterScatter}
\title{1}
\usage{
ezScatter(x = NULL, y, xlab = NULL, ylab = NULL, nPlotsPerRow = 6,
  shrink = FALSE, lim = range(x, y, na.rm = TRUE), isPresent = NULL,
  types = NULL, pch = 16, colors = rainbow(ncol(types)),
  legendPos = "bottomright", cex.main = 1, cex = 0.8, ...)

ezXYScatterScatter(xVec, yVec, absentColor = "gray", shrink = FALSE,
  frame = TRUE, axes = TRUE, xlim = range(xVec, yVec, na.rm = TRUE),
  ylim = xlim, isPresent = NULL, types = NULL, pch = 16,
  colors = rainbow(ncol(types)), legendPos = "bottomright", ...)
}
\arguments{
\item{lim}{integers specifying the x and y limits of the plot.}

\item{isPresent}{a logical specifying whether the signal is present.}

\item{types}{a character vector containing the types.}

\item{pch}{an integer specifying the look of plotted points.}

\item{colors}{a character vector containing colors.}

\item{legendPos}{a character vector or integer specifying the position of the legend.}

\item{cex.main}{a numeric specifying the size of main titles relative to \code{cex}.}

\item{cex}{a numeric specifying the size of text and symbols.}

\item{...}{additional arguments to be passed further.}
}
\description{
1
}
\section{Functions}{
\itemize{
\item \code{ezXYScatterScatter}: Does the XY scatter plot.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

