% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/messaging.r
\name{ezSessionInfo}
\alias{ezSessionInfo}
\title{Saves session info in a .txt file}
\usage{
ezSessionInfo(path = ".")
}
\arguments{
\item{path}{an existing directory to save the session info in.}
}
\description{
Writes the current session info into a file called sessionInfo.txt
}
\examples{
ezSessionInfo()
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

