% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ezSmoothScatter}
\alias{ezSmoothScatter}
\title{1}
\usage{
ezSmoothScatter(x = NULL, y, xlab = NULL, ylab = NULL, nPlotsPerRow = 6,
  lim = range(x, y, na.rm = TRUE), isPresent = NULL, types = NULL,
  pch = 16, colors = rainbow(ncol(types)), legendPos = "bottomright",
  cex.main = 1, cex = 0.8, ...)
}
\arguments{
\item{lim}{integers specifying the x and y limits of the plot.}

\item{isPresent}{a logical specifying whether the signal is present.}

\item{types}{a character vector containing the types.}

\item{pch}{an integer specifying the look of plotted points.}

\item{colors}{a character vector containing colors.}

\item{legendPos}{a character vector or integer specifying the position of the legend.}

\item{cex.main}{a numeric specifying the size of main titles relative to \code{cex}.}

\item{cex}{a numeric specifying the size of text and symbols.}

\item{...}{additional arguments to be passed further.}
}
\description{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

