% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezSplit}
\alias{ezSplit}
\title{Separates a character vector into a matrix by splitting it.}
\usage{
ezSplit(x, split, ...)
}
\arguments{
\item{x}{the character vector to be split element-wise.}

\item{split}{a character vector defining with what to split.}

\item{...}{additional arguments to be passed to \code{strsplit()}.}
}
\value{
Returns a matrix containing the split vector and \code{x} as the rownames.
}
\description{
The \code{split} vector needs to divide the \code{x} vector into pieces of equal lengths. The split is removed from the original input.
}
\examples{
ezSplit(letters[1:5], "b")
ezSplit(rep("abcde", 4), letters[1:4])
}
\seealso{
\code{\link[base]{strsplit}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
