% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{ezSystem}
\alias{ezSystem}
\title{Invokes a system command}
\usage{
ezSystem(cmd, echo = TRUE, intern = FALSE, stopOnFailure = !intern, ...)
}
\arguments{
\item{cmd}{a system command input as a character.}

\item{echo}{a logical defining whether \code{cmd} should be written by \code{ezWrite}.}

\item{intern}{a logical which indicates whether to capture the output of the command as an R character vector.}

\item{stopOnFailure}{defaults to the opposite of intern and specifies whether the command should stop.}

\item{...}{additional arguments to be passed to \code{system()}.}
}
\value{
Returns input as character vector if \code{echo} is set to TRUE. Returns an error code if \code{intern} is set to FALSE.
}
\description{
Invokes the system command specified by \code{cmd}.
}
\examples{
try(ezSystem("who"))
}
\seealso{
\code{\link[base]{system}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
