% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{ezThreads}
\alias{ezThreads}
\title{Determines the number of CPU cores to be used}
\usage{
ezThreads()
}
\value{
Returns an integer specifying the amount of CPU cores R is using.
}
\description{
The function will try to find the amount of CPU cores on the current host or how many should be used. First, it will try \code{getOption("cores")}, then \code{Sys.getenv("NSLOTS")} and lastly \code{detectCores(logical=TRUE)}.
}
\examples{
ezThreads()
}
\seealso{
\code{\link[parallel]{detectCores}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
