% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/messaging.r
\name{ezTime}
\alias{ezTime}
\title{Modified version of Sys.time()}
\usage{
ezTime()
}
\value{
Returns the system time.
}
\description{
Displays the system time in a different format.
}
\examples{
ezTime()
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

