% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{ezTranscriptDbFromRef}
\alias{ezTranscriptDbFromRef}
\title{1}
\usage{
ezTranscriptDbFromRef(reference, dataSource = "FGCZ")
}
\arguments{
\item{reference}{a reference to read the information from.}

\item{dataSource}{a character specifying the source of the data.}
}
\value{
Returns
}
\description{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}
}

