% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezUpdateMissingParam}
\alias{ezUpdateMissingParam}
\title{Updates one list with another list.}
\usage{
ezUpdateMissingParam(list1, with_list)
}
\arguments{
\item{list1}{a list or ezParam object}

\item{with_list}{a list or ezParam object which with to update list1}
}
\description{
Updates a list with the values from another list if the list is missing them or if the values are NULL/NA
}
\examples{
l1 <- list(a = 1, b = c("foo", "bar", "bar"), c = 3, e="foo", f="clause")
l2 <- list(a = 1, b = c("baz", "qux"), d = 3, f=NULL)
ezUpdateMissingParam(l2, l1)
# Returns list(a = 1, b = c("baz", "qux"), d = 3, f = "clause", c = 3, e = "foo")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
