% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{ezUtrSequences}
\alias{ezUtrSequences}
\title{Gets UTR sequences}
\usage{
ezUtrSequences(gtf, chromSeqs)
}
\arguments{
\item{gtf}{the annotation data.frame to get the gene data from.}

\item{chromSeqs}{the chromosome sequences to get the Views from.}
}
\value{
Returns a list with two objects of the class DNAStringSet from the packages Biostrings that represent the 3 and 5 primer.
}
\description{
Gets sequences from untranslated regions.
}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param,"./inst/extdata/genes.gtf")
param$ezRef@refFeatureFile = "./inst/extdata/genes.gtf"
fp = "/srv/GT/reference/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/WholeGenomeFasta/genome.fa"
param$ezRef@refFastaFile = fp
genomeSeq = getTranscriptSequences(param)
ezUtrSequences(gtf)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[Biostrings]{DNAStringSet}}

\code{\link[IRanges]{Views}}

\code{\link[Biostrings]{reverseComplement}}
}

