% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ezValidFilename}
\alias{ezValidFilename}
\title{Enforces a valid file name}
\usage{
ezValidFilename(name, replace = "_")
}
\arguments{
\item{name}{a character representing a file name.}

\item{replace}{a character defining the replacement character(s).}
}
\value{
Returns a character where invalid characters for file names are replaced.
}
\description{
Replaces invalid characters and returns a file name with these replaced
}
\examples{
ezValidFilename("example:filename")
ezValidFilename("or?to/remove(them",replace="")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

