% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{ezValidMail}
\alias{ezValidMail}
\title{Is the email-address valid?}
\usage{
ezValidMail(addressString)
}
\arguments{
\item{addressString}{the character to check.}
}
\value{
Returns TRUE if \code{addressString} contains an @.
}
\description{
Checks, whether \code{addressString} is a valid email-address.
}
\examples{
ezValidMail("example@example.com")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
