% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezVsnNorm}
\alias{ezVsnNorm}
\title{Variance Stabilizing Normalization}
\usage{
ezVsnNorm(x, lts.quantile = 0.6)
}
\arguments{
\item{x}{the matrix to normalize.}

\item{lts.quantile}{a numeric passed to \code{justvsn()}.}
}
\value{
Returns the normalized matrix.
}
\description{
Normalizes the matrix with the Variance Stabilizing Normalization and keeps row and column names.
}
\examples{
m1 = matrix(1:200,50)
m2 = ezVsnNorm(m1)
}
\seealso{
\code{\link[vsn]{justvsn}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
