% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{ezWrite}
\alias{ezWrite}
\title{Write in a single line}
\usage{
ezWrite(..., sep = "", collapse = " ", con = stdout())
}
\arguments{
\item{...}{the arguments to concatenate.}

\item{sep}{a string specifying how to seperate the arguments from each other.}

\item{collapse}{a string specifying how to seperate entries of each argument from each other.}

\item{con}{the connection to the file to write in.}
}
\value{
Returns a single line written into a seperate file.
}
\description{
Concatenates its arguments and writes it as a single line.
}
\examples{
con = file("example.txt", "w")
ezWrite("concatenates","these",4,"arguments",con=con,sep=" ")
close(con)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

