% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ezWrite.table}
\alias{ezWrite.table}
\title{Modified version of write.table}
\usage{
ezWrite.table(
  values,
  file = file,
  head = "Identifier",
  row.names = TRUE,
  col.names = TRUE,
  append = FALSE,
  quote = FALSE,
  sep = "\\t",
  na = "NA",
  digits = NA
)
}
\arguments{
\item{values}{a vector, matrix of data.frame to write a table from.}

\item{file}{the name of the output file.}

\item{head}{the names of the header.}

\item{digits}{the number of digits to round to, if rounding is desired.}
}
\value{
Returns a table written into a separate file.
}
\description{
Modified version of \code{write.table()} with a different default. Best used together with \code{ezRead.table()}.
}
\examples{
m1 = matrix(seq(0.01, 1, 0.01), 10)
colnames(m1) = letters[1:10]
ezWrite.table(m1, "exampleTable", digits=1)
}
\seealso{
\code{\link[utils]{write.table}}

\code{\link{ezRead.table}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
