% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatranscriptomicsAssociatedFunctions.R
\name{filterCountsForDiffExpr}
\alias{filterCountsForDiffExpr}
\title{filterCountsForDiffExpr}
\usage{
filterCountsForDiffExpr(DF, desMat, groupName, group1, group2)
}
\arguments{
\item{a}{count matrix, a design matrix, the groups in the contrast}
}
\value{
Returns data.frame
}
\description{
Prepares matrix for differential analysis with specific contrast.
}
