% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{filterEnrichrResults}
\alias{filterEnrichrResults}
\title{Result filtering}
\usage{
filterEnrichrResults(resList, p = 1, pAdj = 1, z = 0, combinedScore = 0)
}
\arguments{
\item{resList}{list of results, i.e. \code{success} field in the list returned by \code{\link{enrichrEnrich}}}

\item{p}{p-value threshold (maximum)}

\item{pAdj}{adjusted p-value threshold (maximum)}

\item{z}{z-value threshold (maximum)}

\item{combinedScore}{combined score threshold (minimum)}
}
\value{
\code{list} that contains the results for libraries with at least one significant gene
that satisfies the criteria.
}
\description{
Filters the results returned by Enrichr. Libraries that fail to yield any
significant results are removed. However, libraries that contain significant results, retain all
records whether significant or not)
}
\seealso{
\code{\link{runEnrichr}, \link{enrichrEnrich}}
}
\author{
Rehrauer, Hubert

Schmid, Peter

Roman Briskine
}
