% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsio.r
\name{filterFastqByBam}
\alias{filterFastqByBam}
\alias{removeReadsFromFastq}
\title{Filters FastQ files by bam}
\usage{
filterFastqByBam(
  fqFiles,
  bamFile,
  fqOutFiles = NULL,
  doGzip = TRUE,
  keepUnmapped = TRUE,
  isProperPair = NA
)

removeReadsFromFastq(fqFiles, readIds, fqOutFiles = NULL, doGzip = TRUE)
}
\arguments{
\item{fqFiles}{a character vector representing file paths to FastQ files.}

\item{bamFile}{a character representing the bam file path.}

\item{fqOutFiles}{an optional character vector representing file paths to the FastQ output files.}

\item{doGzip}{a logical indicating whether to archive the output files in a gzip archive.}

\item{keepUnmapped}{passed further to \code{scanBamFlag()}.}

\item{isProperPair}{passed further to \code{scanBamFlag()}.}

\item{readIds}{a character containing the read ID's from the \code{bamFile}.}
}
\value{
Returns the names of the filtered FastQ files.
}
\description{
Filters FastQ files by bam and writes them into new files.
}
\section{Functions}{
\itemize{
\item \code{removeReadsFromFastq}: Removes reads from the FastQ files according to the bam filter.
}}

\examples{
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
file = system.file("extdata/yeast_10k/dataset.tsv", package="ezRun", mustWork = TRUE)
param = list()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
input = EzDataset$new(file=file, dataRoot=param$dataRoot)
fqFiles = input$getFullPaths("Read1")
fqFiltered = filterFastqByBam(fqFiles, bamFile, doGzip = FALSE)
}
\seealso{
\code{\link[Rsamtools]{scanBam}}

\code{\link[Rsamtools]{ScanBamParam}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
