% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsTools.R
\name{fixStrand}
\alias{fixStrand}
\alias{flipStrand}
\title{Fixes the strand information}
\usage{
fixStrand(strandValues, strandMode = "sense")

flipStrand(strandValues)
}
\arguments{
\item{strandValues}{a set of strand values + - and *.}

\item{strandMode}{a character defining what should be done with the values:
\itemize{
  \item{sense}{ returns \code{strandValues} without modifying them.}
  \item{antisense}{ returns the flipped \code{strandValues}.}
  \item{both}{ sets all the \code{strandValues} to "*".}
}}
}
\value{
Returns a character vector or Rle object containing the strand information.
}
\description{
Fixes the strand information by either leaving it as it is, flipping it or setting both.
}
\section{Functions}{
\itemize{
\item \code{flipStrand()}: This function performs the flipping of the strand values.

}}
\examples{
strandValues = c("+","+","-","*")
fixStrand(strandValues, "sense")
flipStrand(strandValues)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
