% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage.r
\name{gageSigGenes}
\alias{gageSigGenes}
\title{Gets significant genes}
\usage{
gageSigGenes(x, gene.pValue = NULL, signal = NULL)
}
\arguments{
\item{x}{the gage results to get significant genes from.}

\item{gene.pValue}{a numeric specifying the p-Value threshold.}

\item{signal}{a character specifying which signal to analyse.}
}
\value{
Returns the gage results with labeled significant genes.
}
\description{
Gets significant genes from the gage analysis.
}
\examples{
param = ezParam(userParam = list('refBuild' = 'Schizosaccharomyces_pombe/Ensembl/EF2/Annotation/Version-2013-03-07'))
pValue = param$gageGeneThreshold
x = "gageResults"
gageSigGenes(x, pValue, "both")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

