% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGSeqAssociatedFunctions.R
\name{generateFlattenedFileForSGeq}
\alias{generateFlattenedFileForSGeq}
\title{Generate flattened file for SGSeq-based differential expression}
\usage{
generateFlattenedFileForSGeq(x)
}
\arguments{
\item{a}{list of count files generated with SGseq functions analyzeFeatures, analyzeVariants and counts}
}
\value{
a text file
}
\description{
Generate flattened file for SGSeq-based differential expression
}
