% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamio.r
\name{getBamMultiMatching}
\alias{getBamMultiMatching}
\title{Gets bam multi matching}
\usage{
getBamMultiMatching(param, bamFile, nReads = NULL)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{paired}{ a logical indicating whether the samples are paired.}
  \item{pairedMode}{ a character indicating the flag options. Possible inputs are: "first", "second" and "paired".}
}}

\item{bamFile}{a character representing the file path to the bam file to read.}

\item{nReads}{an integer specifying the number of reads. These additional reads will be counted in "0"}
}
\value{
Returns a integer vector specifying the amount of matching reads.
}
\description{
Gets bam multi matching and returns the result as an integer vector.
}
\examples{
param = ezParam()
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
getBamMultiMatching(param, bamFile, nReads=100)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

