% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getBlueRedScale}
\alias{getBlueRedScale}
\title{Gets a color scale from blue to red}
\usage{
getBlueRedScale(n = 256, whiteLevel = 0.9)
}
\arguments{
\item{n}{an integer specifying the amount of colors to split the scale in.}

\item{whiteLevel}{a numeric specifying the maximum whiteness for red, green and blue.}
}
\value{
Returns a character vector containing colors in hex format.
}
\description{
Gets a color scale from blue to red in hex format.
}
\examples{
rbs = getBlueRedScale()
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
