% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-mapping.r
\name{getBowtie2Reference}
\alias{getBowtie2Reference}
\title{Builds and gets a Bowtie2 reference}
\usage{
getBowtie2Reference(param)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{ezRef@refIndex}{ a character specifying the location of the index that is used in the alignment.}
  \item{ezRef@refBuildDir}{ a character specifying the directory of the reference build.}
  \item{ezRef@refFastaFile}{ a character specifying the file path to the fasta file.}
}}
}
\value{
Returns a character specifying the file path to the Bowtie2 indexed reference genome.
}
\description{
Builds and gets a Bowtie2 reference by returning the file path to it.
}
\examples{
param = ezParam(userParam = list('refBuild' = 'Schizosaccharomyces_pombe/Ensembl/EF2/Annotation/Version-2013-03-07'))
getBowtie2Reference(param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

