% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{getEnrichrLibNames}
\alias{getEnrichrLibNames}
\title{Get Enrichr library names}
\usage{
getEnrichrLibNames()
}
\value{
a vector containing library names
}
\description{
reads Enrichr library names from the internal file
  (\code{extdata/enrichr_libnames.txt})
}
\author{
Rehrauer, Hubert

Schmid, Peter

Roman Briskine
}
\seealso{
\code{\link{runEnrichr}}
}

