% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\name{getGeneIdExprLinks}
\alias{getGeneIdExprLinks}
\title{Get Links from GeneIds to Expression result files}
\usage{
getGeneIdExprLinks(pvGeneIds, psdexseq_report_path)
}
\arguments{
\item{pvGeneIds}{vector of gene ids}

\item{psdexseq_report_path}{path to where DEXSeqHTML report is saved}
}
\description{
Get Links from GeneIds to Expression result files
}
