% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.r
\name{getGeneMapping}
\alias{getGeneMapping}
\alias{hasGeneMapping}
\title{Gets the gene mapping}
\usage{
getGeneMapping(param, seqAnnoDF)

hasGeneMapping(param, seqAnnoDF)
}
\arguments{
\item{param}{the parameters to specify which names to use. This is taken from \code{geneColumn} or \code{geneColumnSet} if the first one is not specified.}

\item{seqAnnoDF}{a data.frame containing the sequence annotation.}
}
\value{
Returns a named character vector containing the gene mapping.
}
\description{
Gets the gene mapping.
}
\section{Functions}{
\itemize{
\item \code{hasGeneMapping}: Returns a logical indicating if the set parameters have gene mapping or not.
}}
\examples{
param = ezParam()
param$ezRef@refFeatureFile = system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE)
param$ezRef@refAnnotationFile = ""
fp = "/srv/GT/reference/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/WholeGenomeFasta/genome.fa"
param$ezRef@refFastaFile = fp
seqAnno = writeAnnotationFromGtf(param)
getGeneMapping(param,seqAnno)
hasGeneMapping(param,seqAnno)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

