% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage.r
\name{getGeneSets}
\alias{getGeneSets}
\alias{readGmt}
\title{Gets the gene set}
\usage{
getGeneSets(param)

readGmt(file)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{KEGGgmt}{ a character representing the file path of the gmt data.}
  \item{genesets}{ a character vector containing the gene sets.}
  \item{MSigDB}{ a character. Used if the species is Homo sapiens.}
  \item{MSigDBPath}{ a character representing the file path to \code{MSigDB}. Used if the species is Homo sapiens.}
}}
}
\value{
Returns the gene set.
}
\description{
Gets the gene set
}
\section{Functions}{
\itemize{
\item \code{readGmt}: Reads the gmt file in.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

