% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\name{getGeneTable}
\alias{getGeneTable}
\title{Generate gene table from DEXSeqResults object}
\usage{
getGeneTable(pdxr, param)
}
\arguments{
\item{pdxr}{DEXSeqResults object}

\item{param}{EzParam object}
}
\description{
Generate gene table from DEXSeqResults object
}

