% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igv.R
\name{getIgvLocusLink}
\alias{getIgvLocusLink}
\title{Gets an IGV locus link in .html format}
\usage{
getIgvLocusLink(chrom, start, end)
}
\arguments{
\item{chrom}{the name of the chromosome.}

\item{start}{the start position of the locus.}

\item{end}{the end position of the locus.}
}
\value{
Returns a character containing a link in .html format.
}
\description{
Gets an IGV locus link in html format using chromosome, start and end information.
}
\examples{
getIgvLocusLink(17,31,465)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
