% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage.r
\name{getKeggId}
\alias{getKeggId}
\title{Gets the KEGG ID}
\usage{
getKeggId(x, param)
}
\arguments{
\item{x}{a character representing the species name.}

\item{param}{a list of parameters to extract \code{KEGGdb} and \code{KEGGOrgId} from.}
}
\value{
Returns the KEGG ID as a character.
}
\description{
Gets the KEGG ID.
}
\examples{
param = ezParam(userParam = list('refBuild' = 'Schizosaccharomyces_pombe/Ensembl/EF2/Annotation/Version-2013-03-07'))
x = getSpeciesName(param)
getKeggId(x, param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

