% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rangesTools.r
\name{getRangeValues}
\alias{getRangeValues}
\title{1}
\usage{
getRangeValues(cov, targetRanges, doRev = as.character(strand(targetRanges))
  == "-", asMatrix = TRUE)
}
\arguments{
\item{x}{}
}
\value{
Returns
}
\description{
1
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

