% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsReferenceFiles.r
\name{getRefChromSizesFile}
\alias{getRefChromSizesFile}
\title{Gets the file containing the chromosome sizes}
\usage{
getRefChromSizesFile(param)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{ezRef@refChromSizesFile}{ a character representing the path to the file to get.}
  \item{ezRef@refChromDir}{ a character representing the path to the directory to get the information to calculate chromosome sizes. Used if the file does not yet exist.}
}}
}
\value{
Returns the path to the file containing the chromosome sizes.
}
\description{
Gets the file containing the chromosome sizes either directly or by the fasta files in the chromosome directory.
}
\examples{
param = ezParam()
param$ezRef@refChromSizesFile = "example.txt"
param$ezRef@refChromDir = "./script/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/Chromosomes"
getRefChromSizesFile(param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

