% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsReferenceFiles.R
\name{getReferenceFeaturesBed}
\alias{getReferenceFeaturesBed}
\title{Gets a bed file}
\usage{
getReferenceFeaturesBed(param)
}
\arguments{
\item{param}{a list of parameters to extract \code{ezRef@refFeatureFile} from.}
}
\value{
Returns the path to the created bed file.
}
\description{
Gets a bed file from a gtf annotation file.
}
\examples{
\dontrun{
param = ezParam()
param$ezRef@refFeatureFile = system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE)
rfbed = getReferenceFeaturesBed(param)
}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
