% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rawData.R
\name{getRpkm}
\alias{getRpkm}
\title{Gets rpkm}
\usage{
getRpkm(rawData)
}
\arguments{
\item{rawData}{a list of raw data. Usually obtained from \code{loadCountDataset()}.}
}
\value{
Returns rpkm
}
\description{
Gets rpkm.
}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
input = EzDataset$new(file=system.file("extdata/yeast_10k_STAR_featureCounts/dataset.tsv", package="ezRun", mustWork = TRUE))
rawData = loadCountDataset(input, param)
getRpkm(rawData)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

