% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getSampleColors}
\alias{getSampleColors}
\alias{getSamplePch}
\alias{getSampleLty}
\title{Gets the sample colors}
\usage{
getSampleColors(conds, colorNames = names(conds), hueStep = 0)

getSamplePch(conds, pchNames = names(conds))

getSampleLty(conds, ltyNames = names(conds), maxLineTypes = 5)
}
\arguments{
\item{conds}{conditions obtained from \code{ezConditionsFromDesign()} or \code{ezConditionsFromDataset()}.}

\item{colorNames}{a character vector containing the names for the colors.}

\item{hueStep}{a numeric specifying the hue step for \code{hsv()}.}
}
\value{
Returns a character vector containing colors in hex format.
}
\description{
Gets the sample colors from the experimental conditions.
}
\section{Functions}{
\itemize{
\item \code{getSamplePch}: Gets the sample pch from the experimental conditions.

\item \code{getSampleLty}: Gets the sample line types from the experimental conditions.
}}

\examples{
file = system.file("extdata/yeast_10k/dataset.tsv", package="ezRun", mustWork = TRUE)
ds = EzDataset$new(file=file, dataRoot=NULL)
cond = ezConditionsFromDataset(ds$meta)
getSampleColors(cond)
getSamplePch(cond)
getSampleLty(cond)
}
\seealso{
\code{\link[grDevices]{hsv}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
