% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{getSignal}
\alias{getLog2Signal}
\alias{getSignal}
\title{Gets the signal}
\usage{
getSignal(rawData)

getLog2Signal(rawData)
}
\arguments{
\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}
}
\value{
Returns the signal.
}
\description{
Gets the signal from raw data and modifies it if necessary due to logarithm.
}
\section{Functions}{
\itemize{
\item \code{getLog2Signal}: Does the same but returns the log2 instead.
}}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
file = system.file("extdata/yeast_10k_STAR_counts/dataset.tsv", package="ezRun", mustWork = TRUE)
input = EzDataset$new(file=file)
rawData = loadCountDataset(input, param)
sig = getSignal(rawData)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

