% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{getSignificantRatioCountsTable}
\alias{getSignificantRatioCountsTable}
\title{Gets significant ratios}
\usage{
getSignificantRatioCountsTable(result, pThresh = 1/10^(1:5),
  ratioThresh = c(1, 1.5, 2, 3, 4, 8, 10), genes = NULL)
}
\arguments{
\item{result}{a list of results.}

\item{pThresh}{a numerical indicating the p-value threshold.}

\item{ratioThresh}{a ratio threshold .....}

\item{genes}{if not NULL, the function counts the number of different genes that are significant.}
}
\value{
Returns a matrix containing the counts of significant ratios.
}
\description{
Gets significant ratios and counts them in a table.
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

