% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage.r
\name{getSpeciesName}
\alias{getSpeciesName}
\title{Gets the species name}
\usage{
getSpeciesName(param)
}
\arguments{
\item{param}{a list of parameters to extract \code{refBuild} from.}
}
\value{
Returns the name of the species derived from \code{refBuild}.
}
\description{
Gets the species name.
}
\examples{
ls = list('refBuild' = 'Schizosaccharomyces_pombe/Ensembl/EF2/Annotation/Version-2013-03-07')
param = ezParam(userParam = ls)
getSpeciesName(param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
