% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-mapping.r
\name{getSTARReference}
\alias{getSTARReference}
\title{Builds and gets a STAR reference}
\usage{
getSTARReference(param)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{ezRef@refIndex}{ a character specifying the location of the index that is used in the alignment.}
  \item{ezRef@refFeatureFile}{ a character specifying the file path to the annotation feature file (.gtf).}
  \item{ezRef@refChromDir}{ a character specifying the file path to the directory of the chromosome information.}
  \item{ram}{ an integer specifying how many gigabytes of RAM to use.}
  \item{ezRef@refFastaFile}{ a character specifying the file path to the fasta file.}
}}
}
\value{
Returns a character specifying the file path to the STAR indexed reference genome.
}
\description{
Builds and gets a STAR reference by returning the file path to it.
}
\examples{
param = ezParam(userParam = list('refBuild' = 'Schizosaccharomyces_pombe/Ensembl/EF2/Annotation/Version-2013-03-07'))
getSTARReference(param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

