% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{removeSuffix}
\alias{removeSuffix}
\alias{getSuffix}
\title{Gets the suffix of a file name}
\usage{
removeSuffix(filename)

getSuffix(filename)
}
\arguments{
\item{filename}{a filename to extract the suffix from.}
}
\value{
Returns a character with either the suffix of the file name or the file name without suffix.
}
\description{
Gets the suffix of a file name or the file name without the suffix.
}
\section{Functions}{
\itemize{
\item \code{removeSuffix()}: Gets and removes the suffix of a file name.

}}
\examples{
getSuffix("example.file")
removeSuffix("example.file")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
