\name{getTpm}
\alias{getTpm}
\title{Gets the tpm measurement}
\usage{
  getTpm(rawData)
}
\arguments{
  \item{rawData}{\code{SummarizedExperiment} or \code{SingleCellExperiment} object.}
}
\value{
  Returns the tpm measurement.
}
\description{
  Gets the transcripts per million measurement.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
