% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{getTranscript2Gene}
\alias{getTranscript2Gene}
\alias{countIsoformsPerGene}
\title{Gets gene names from annotation}
\usage{
getTranscript2Gene(gtf)

countIsoformsPerGene(tr2Gene)
}
\arguments{
\item{gtf}{the annotation data.frame to get the gene data from.}

\item{tr2Gene}{a character vector containing the transcript ID's obtained from the gene ID's.}
}
\value{
Returns a character vector with the gene names.
}
\description{
Gets gene names from an annotation data.frame of gtf or gff format.
}
\section{Functions}{
\itemize{
\item \code{countIsoformsPerGene()}: Returns a contingency table of the transcript names counting isoforms per gene.

}}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param, system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE))
tr2Gene = getTranscript2Gene(gtf)
ipg = countIsoformsPerGene(tr2Gene)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
