% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gff.r
\name{getTranscriptGcAndWidth}
\alias{getTranscriptGcAndWidth}
\title{Gets GC proportions and gene widths}
\usage{
getTranscriptGcAndWidth(param)
}
\arguments{
\item{param}{the parameters to load the annotation and sequence files from.}
}
\value{
Returns a list containing named vectors for the GC proportion and the gene width.
}
\description{
Gets GC proportions and gene widths from annotation (gtf or gff) and sequence (fasta) information.
}
\examples{
param = ezParam()
param$ezRef@refFeatureFile = "./inst/extdata/genes.gtf"
param$ezRef@refFastaFile = "./script/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/WholeGenomeFasta/genome.fa"
getTranscriptGcAndWidth(param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{ezLoadFeatures}}

\code{\link[Biostrings]{readDNAStringSet}}

\code{\link[IRanges]{Views}}

\code{\link[Biostrings]{letterFrequency}}
}

