% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsTools.R
\name{getTuxedoLibraryType}
\alias{getTuxedoLibraryType}
\title{Gets the tuxedo library type}
\usage{
getTuxedoLibraryType(strandMode)
}
\arguments{
\item{strandMode}{a character defining the strand mode and what to return.
\itemize{
  \item{sense}{ returns "fr-secondstrand".}
  \item{antisense}{ returns "fr-firststrand".}
  \item{both}{ returns "fr-unstranded".}
}}
}
\value{
Returns a character.
}
\description{
Gets the tuxedo library type from the strand mode.
}
\examples{
getTuxedoLibraryType("sense")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
