% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gff.r
\name{gffToRanges}
\alias{gffToRanges}
\title{Gets GRanges from annotation}
\usage{
gffToRanges(gff)
}
\arguments{
\item{gff}{the annotation data.frame to get the GRanges from.}
}
\value{
Returns a GRanges object
}
\description{
Gets GRanges from annotation of the format gtf or gff.
}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param,"./inst/extdata/genes.gtf")
gffToRanges(gtf)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[GenomicRanges]{GRanges}}
}

