% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gff.r
\name{gffTrimTranscripts}
\alias{gffTrimSingleTranscript}
\alias{gffTrimTranscripts}
\title{Trims transcripts in annotation data.frames}
\usage{
gffTrimTranscripts(gff, maxLength = 100, start = TRUE)

gffTrimSingleTranscript(gffSingle, maxLength = 100, start = TRUE)
}
\arguments{
\item{gff}{an annotation data.frame in gtf or gff format.}

\item{maxLength}{an integer specifying the maximum length of transcripts after trimming.}

\item{start}{a logical indicating the start direction of each transcript to be trimmed.}
}
\value{
Returns
}
\description{
Trims transcripts in annotation data.frames of gtf or gff format.
}
\section{Functions}{
\itemize{
\item \code{gffTrimSingleTranscript}: Trims a single transcript according to \code{maxLength} and in reversed order depending on the strand (+ / -).
}}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param,"./inst/extdata/genes.gtf")
gffTrimTranscripts(gtf[gtf$type == "exon", ])
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

