% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-reports.R
\name{goClusterTable}
\alias{goClusterTable}
\alias{goResultToHtmlTable}
\title{Gets the GO cluster table}
\usage{
goClusterTable(param, clusterResult)

goResultToHtmlTable(x, pThreshGo, minCount, onto = NA,
  maxNumberOfTerms = 40)
}
\arguments{
\item{param}{a list of parameters to extract \code{pValThreshFisher} and \code{minCountFisher} from}

\item{clusterResult}{a list containing the result of the analysis done by \code{clusterHeatmap()}.}
}
\value{
Returns a flex table containing the GO information of the cluster result.
}
\description{
Gets the GO cluster table
}
\section{Functions}{
\itemize{
\item \code{goResultToHtmlTable}: Parses the GO result into a html table.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{clusterHeatmap}}
}

