% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-reports.R
\name{goClusterTableRmd}
\alias{goClusterTableRmd}
\title{Gets the GO cluster table}
\usage{
goClusterTableRmd(param, clusterResult, seqAnno)
}
\arguments{
\item{param}{a list of parameters to extract \code{pValThreshFisher} and \code{minCountFisher} from}

\item{clusterResult}{a list containing the result of the analysis done by \code{goClusterResults()}.}

\item{seqAnno}{the sequence annotation.}
}
\value{
Returns a flex table containing the GO information of the cluster result.
}
\description{
Gets the GO cluster table
}
\seealso{
\code{\link{goClusterResults}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
