% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R
\name{goStringsToList}
\alias{goStringsToList}
\title{Parses GO strings to a list}
\usage{
goStringsToList(goStrings, listNames = NULL)
}
\arguments{
\item{goStrings}{a character vector containing the GO strings}

\item{listNames}{an optional character vector naming the list elements.}
}
\value{
Returns the GO strings as a list.
}
\description{
Parses GO strings to a list.
}
\examples{
goStrings = c("GO 1;GO 2;GO 3", "GO 4")
listNames = letters[1:2]
goStringsToList(goStrings, listNames)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
