% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{groupGff}
\alias{groupGff}
\alias{gffGroupToRanges}
\title{Groups annotation}
\usage{
groupGff(
  gtf,
  grouping = gtf$Parent,
  type = "transcript",
  skipTransSpliced = FALSE
)

gffGroupToRanges(gtf, grouping, skipTransSpliced = FALSE)
}
\arguments{
\item{gtf}{the annotation data.frame to group.}

\item{grouping}{a vector of the annotation data.frame to group it by.}

\item{type}{a character representing the new \code{type} name.}

\item{skipTransSpliced}{a logical indicating whether to skip transspliced elements if they exist.}
}
\value{
Returns a grouped annotation data.frame.
}
\description{
Groups annotation according to the \code{grouping} factor. Duplicated values are dropped and the function checks for transspliced elements.
}
\section{Functions}{
\itemize{
\item \code{gffGroupToRanges}: Groups annotation and returns a GRanges object from the grouped annotation.
}}

\examples{
param = ezParam()
gtf = ezLoadFeatures(param, system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE))
groupedGtf = groupGff(gtf)
gffGroupToRanges(gtf, grouping = gtf$Parent)
}
\seealso{
\code{\link[GenomicRanges]{GRanges}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
