% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{groupModRichPlot}
\alias{groupModRichPlot}
\title{Modified phyloseq richness plot}
\usage{
groupModRichPlot(
  physeq,
  x,
  color = NULL,
  shape = NULL,
  title = NULL,
  scales = "free_y",
  nrow = 1,
  shsi = NULL,
  measures = c("Shannon"),
  sortby = NULL
)
}
\arguments{
\item{physeqFullObject}{(phyloseq object)}
}
\value{
A ggplot
}
\description{
Modified richness plot to include box plots statistical comparison when groups are present
}
