% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hasFilesafeCharacters}
\alias{hasFilesafeCharacters}
\title{Checks if the argument can be safely used as a filename}
\usage{
hasFilesafeCharacters(x)
}
\value{
TRUE or FALSE
}
\description{
only supports alphanumeric characters and "+-_"
}
\examples{
hasFilesafeCharacters("a")
hasFilesafeCharacters("a\n")
hasFilesafeCharacters(c("1", "2"))
hasFilesafeCharacters(list("1", "2 x"))
}
