% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{heatmapForPhylotseqPlotPheatmap}
\alias{heatmapForPhylotseqPlotPheatmap}
\title{Alternative heatmap with pheatmap plot for phyloseq abundnce-taxonimy matrix}
\usage{
heatmapForPhylotseqPlotPheatmap(
  phyloseqOtuObj,
  areThereMultVar,
  isGroupThere,
  rank
)
}
\arguments{
\item{a}{phyloseq object and the rank to summarize}
}
\value{
Returns a stacked bar  plot.
}
\description{
Alternative heatmap plot for phyloseq abundnce-taxonimy matrix
}
