% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{imgLinks}
\alias{imgLinks}
\title{Pastes image links as html}
\usage{
imgLinks(image)
}
\arguments{
\item{image}{character(s) specifying links to image files.}
}
\value{
Returns html link(s).
}
\description{
A simple wrapper that pastes image links as html.
}
\examples{
imgLinks("link.png")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
