% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interleaveMatricesByColumn}
\alias{interleaveMatricesByColumn}
\title{Combines two matrices by their columns}
\usage{
interleaveMatricesByColumn(x, y, suffixes = c("[Signal]", "[Present]"))
}
\arguments{
\item{x}{first matrix containing both row and column names.}

\item{y}{second matrix containing the same column names as the first one.}

\item{suffixes}{a character vector to append as suffixes to the colnames of the result.}
}
\value{
Returns the combined matrix.
}
\description{
The matrices need to have the same dimensions and column names must be equal. The first matrix needs row names as well. This is used to merge, e.g. matrices for expression signal and present flag by sample.
}
\examples{
m1 = matrix(1:10,2)
m2 = matrix(11:20,2)
colnames(m1) = as.character(1:5)
rownames(m1) = c("a","b")
colnames(m2) = as.character(1:5)
interleaveMatricesByColumn(m1,m2)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
