% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{intHist}
\alias{intHist}
\title{Shrinks values and plots a histogram}
\usage{
intHist(
  x,
  range = c(round(min(x, na.rm = TRUE)) - 0.5, round(max(x, na.rm = TRUE)) + 0.5),
  step = 1,
  ...
)
}
\arguments{
\item{x}{a vector of values to shrink and plot a histogram from.}

\item{range}{two values specifying the range to shrink \code{x} to.}

\item{step}{a value specifying the step distance between the breaks for \code{hist()}.}

\item{...}{additional arguments to be passed to \code{hist()}.}
}
\value{
Returns the histogram.
}
\description{
Shrinks values and plots a histogram.
}
\examples{
intHist(1:10)
}
\seealso{
\code{\link[graphics]{hist}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
