% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{inverseMapping}
\alias{inverseMapping}
\alias{makeMultiMapping}
\title{Inverse mapping}
\usage{
inverseMapping(xList)

makeMultiMapping(xList)
}
\arguments{
\item{xList}{a named list.}
}
\value{
Returns a list for the input values containing all the names that belong to each value in the input list.
}
\description{
Strips the names and values of a list apart and rematches all the names that apply to each value.
}
\section{Functions}{
\itemize{
\item \code{makeMultiMapping()}: Unlists the input and returns a data.frame with separated names and values.

}}
\examples{
l1 = list(a=1:3, b=c(2,5), c=4:8)
inverseMapping(l1)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
