% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt.R
\name{keepOnlyCoreFiles}
\alias{keepOnlyCoreFiles}
\title{keep only files in *CORE_RESULTS* folder}
\usage{
keepOnlyCoreFiles(data.dir)
}
\arguments{
\item{data.dir}{<string> directory in which the function will be executed.}
}
\description{
to reduce the final output size, this function deletes all files within the "processed_output"

directory to keep only the core files required to generate the counts and final report.
}
\author{
Miquel Anglada Girotto
}
