% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lastVal}
\alias{lastVal}
\title{Gets the last value of an object}
\usage{
lastVal(x)
}
\arguments{
\item{x}{a vector, object or a list.}
}
\description{
Gets the last value of an object.
}
\examples{
lastVal(1:10)
obj1 <- 1:5
obj2 <- "a character"
lastVal(c(obj1,obj2))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
