% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ncpro.R
\name{lGetGlobalCountParam}
\alias{lGetGlobalCountParam}
\title{specification of default values used for count splitting}
\usage{
lGetGlobalCountParam()
}
\description{
specification of default values used for count splitting
}
