% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R, R/app-DEXSeqCounting.R
\name{lGetPyScriptPaths}
\alias{lGetPyScriptPaths}
\title{Get list with required python script paths}
\usage{
lGetPyScriptPaths()

lGetPyScriptPaths()
}
\description{
Get list with required python script paths

Get list with required python script paths
}

