% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatranscriptomicsAssociatedFunctions.R
\name{listOfAbundMerge}
\alias{listOfAbundMerge}
\title{listOfAbundMerge}
\usage{
listOfAbundMerge(listOfAbund, names)
}
\arguments{
\item{a}{list of data frame, one per sample.}
}
\value{
Returns a data.frame.
}
\description{
Create summary table from list of individual abundance.
}
