% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitessce-utils.R
\name{load_seurat_object}
\alias{load_seurat_object}
\title{Load Seurat object from various file formats}
\usage{
load_seurat_object(seurat_path, nthreads = 8)
}
\arguments{
\item{seurat_path}{Path to Seurat file (.qs2, .qs, .rds)}

\item{nthreads}{Number of threads for qs/qs2 reading}
}
\value{
Seurat object
}
\description{
Load Seurat object from various file formats
}
