% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsio.r
\name{loadCountDataset}
\alias{loadCountDataset}
\title{Loads the count dataset}
\usage{
loadCountDataset(input, param)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{param}{a list of parameters:
\itemize{
  \item{dataRoot}{ the root directory of the files.}
  \item{expressionName}{ if specified, this will be used as the column name...}
  \item{knownExpressionNames}{ ...or otherwise known expression names that occur in the dataset will be used.}
  \item{ezRef@refBuild}{ if specified, the sequence annotation will be extracted from \code{ezFeatureAnnotation()}.}
  \item{useTranscriptType}{ if specified, only the defined transcript type will be used.}
  \item{sigThresh}{ the threshold...}
  \item{useSigThresh}{ ...and whether it should be used.}
  \item{featureLevel}{ if equal to "gene" and the feature level of the dataset to "isoform", the rawdata will be passed to \code{aggregateCountsByGene()} before returning it.}
}}
}
\value{
Returns a list of raw data.
}
\description{
Loads the count dataset with the given input.
}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
input = EzDataset$new(file=system.file("extdata/yeast_10k_STAR_featureCounts/dataset.tsv", package="ezRun", mustWork = TRUE))
loadCountDataset(input, param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{ezFeatureAnnotation}}

\code{\link{aggregateCountsByGene}}
}

