% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{logMessage}
\alias{logMessage}
\title{Logs time and base directory together with a parameter}
\usage{
logMessage(methodName, param, message)
}
\arguments{
\item{methodName}{the name of the method to display.}

\item{param}{the parameter to display.}

\item{message}{a character to start the expression and provide context information.}
}
\value{
Returns a message about \code{param} with a context.
}
\description{
Logs a time with an additional context provided in \code{message} and the base directory. Also displays a defined method name and the parameters name and Id.
}
\examples{
param = ezParam()
logMessage("a method", param, "Starting")
logMessage("a method", param, "Finished")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
