% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-genome.R
\name{makeCoordinate}
\alias{makeCoordinate}
\title{Make a coordinate}
\usage{
makeCoordinate(chrom, strand, start, stop)
}
\arguments{
\item{chrom}{the chromosome name.}

\item{strand}{the strand of the chromosome.}

\item{start}{where the coordinate starts.}

\item{stop}{where the coordinate stops.}
}
\value{
Returns a chromosome coordinate.
}
\description{
Creates a chromosome coordinate with the provided input.
}
\examples{
makeCoordinate("chrm", 3, 150, 300)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

