% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{makeRmdReport}
\alias{makeRmdReport}
\title{make ezRun Rmd report}
\usage{
makeRmdReport(
  ...,
  htmlFile = "00index.html",
  rmdFile = "",
  selfContained = TRUE,
  reportTitle = "SUSHI Report",
  use.qs2 = FALSE
)
}
\arguments{
\item{...}{any number of R object}

\item{htmlFile}{The name of the output html}

\item{rmdFile}{The name of the Rmd template}

\item{selfContained}{If FALSE, will copy over the css and other aesthetic assets}

\item{reportTitle}{Title to use the header of the html}

\item{use.qs2}{If TRUE, use qs2 rather than rds}
}
\description{
Makes an ezRun Rmd report and additionally writes the R objects to disk using eather RDS or the qs2 format
}
\examples{
makeRmdReport(param=param, output=output, scData=scData, rmdFile = "ScSeuratCombine.Rmd", reportTitle = "Seurat Report", use.qs2=TRUE) 
# Outputs the report as '00index.html' in addition to files 'param.qs2', 'output.qs2', 'scData.qs2'
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
