% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGSeqAssociatedFunctions.R
\name{makeSGSeqtxdbObject}
\alias{makeSGSeqtxdbObject}
\title{Create and save SGSea database}
\usage{
makeSGSeqtxdbObject(refBuil, refFeat, txdbDatabase)
}
\arguments{
\item{a}{gtf file}
}
\value{
a makeTxDbFromGFF database object
}
\description{
Create and save SGSea database
}
