% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{mergeRawData}
\alias{mergeRawData}
\title{Merges raw data}
\usage{
mergeRawData(rawData1, rawData2)
}
\arguments{
\item{rawData1}{a list of raw data. This is the main dataset where the information from the second gets added to.}

\item{rawData2}{the second list of raw data. Both are obtained from \code{loadCountDataset()}.}
}
\value{
Returns a list containing the merged raw data.
}
\description{
Merges raw data by combining two lists of raw data into one.
}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
file = system.file("extdata/yeast_10k_STAR_counts/dataset.tsv", package="ezRun", mustWork = TRUE)
input = EzDataset$new(file=file, dataRoot=param$dataRoot)
rawData1 = loadCountDataset(input$subset(1), param)
rawData2 = loadCountDataset(input$subset(2), param)
rdm = mergeRawData(rawData1, rawData2)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
