% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotationQC_associatedFunctions.R
\name{mergeSummaryBinFiles}
\alias{mergeSummaryBinFiles}
\title{mergeSummaryBinFile}
\usage{
mergeSummaryBinFiles(binFile)
}
\arguments{
\item{a}{list of summary bin files generated from the summaryMetagenomeBins fun}
}
\value{
Returns a data frame
}
\description{
Merged summary bin files for plots
}
