% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatranscriptomicsAssociatedFunctions.R
\name{metagMetatrDifferentialAnalysis_tableAndPlots}
\alias{metagMetatrDifferentialAnalysis_tableAndPlots}
\title{diffAbundReport}
\usage{
metagMetatrDifferentialAnalysis_tableAndPlots(
  countMatrix,
  designMatrix,
  group,
  sampleGroup,
  refGroup,
  mode = "expression",
  DB = NULL,
  N = 30
)
}
\arguments{
\item{a}{count and a design matrix}
}
\value{
Returns a list of tables and plots.
}
\description{
Differential abundance analysis between groups.
}
