% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{mioString}
\alias{mioString}
\title{Converts a number into a string representing millions}
\usage{
mioString(x)
}
\arguments{
\item{x}{an integer.}
}
\value{
Returns a string representing the input integer in millions.
}
\description{
used for labels etc. where typically millions of reads are indicated.
}
\examples{
mioString(4e7)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

