% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurDataCleanAssociatedFunctions.R
\name{mixedDatasetToMothur}
\alias{mixedDatasetToMothur}
\title{Creates Mothur input files from  Sushi dataset.}
\usage{
mixedDatasetToMothur(sushiInputDataset, param)
}
\arguments{
\item{taxaFileName, }{mothur taxonomy file.}
}
\value{
Returns the .groups and .fasta files
}
\description{
Converts Sushi dataset into Mothur input.
}
