% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{mkCurlQryString}
\alias{mkCurlQryString}
\title{Query string generator}
\usage{
mkCurlQryString(...)
}
\arguments{
\item{...}{Named parameters that will form the query}
}
\value{
A string in the following format: \code{param1=value1&param2=value2}.
}
\description{
Helper function that generates a query string out of named parameters. The main
purpose of this function is to generate query strings for curl requests. All special characters
in the values are escaped.
}
\author{
Rehrauer, Hubert

Schmid, Peter

Roman Briskine
}

