% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ncpro.R
\name{modifyInput}
\alias{modifyInput}
\title{input is modified to show the count results}
\usage{
modifyInput(psCountCategory, pdataset, psInputFn, psReportDir)
}
\description{
\code{modifyInput} for a given count category, tsv-formatted
metadata is modified such that paths to read files are replaced
with paths to output count files
}
