% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-edgerMultiGroups.R
\name{multiGroupCountComparison}
\alias{multiGroupCountComparison}
\alias{runDeseq2MultiGroup}
\alias{runEdgerGlmMultiGroup}
\title{Compares the counts of many groups}
\usage{
multiGroupCountComparison(rawData, param)

runDeseq2MultiGroup(x, sampleGroup, refGroup, grouping, grouping2 = NULL)

runEdgerGlmMultiGroup(x, refGroup, grouping, normMethod, grouping2 = NULL)
}
\arguments{
\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}

\item{param}{a list of parameters:
\itemize{
  \item{testMethod}{ defines the method to run: deseq2, exactTest, glm, sam or limma. Defaults to glm.}
  \item{grouping}{ a character specifying the grouping.}
  \item{grouping2}{ a character vector specifying a secondary factor or numeric co-variate.}
  \item{refGroup}{ a character specifying the reference group.}
  \item{normMethod}{ a character specifying the normalization method for the edger and glm test methods.}
}}
}
\value{
Returns a list containing the results of the comparison.
}
\description{
Compares the counts of many groups with the option to choose from several methods to test them.
}
\section{Functions}{
\itemize{
\item \code{runDeseq2MultiGroup}: Runs the Deseq2 test method for many groups.

\item \code{runEdgerGlmMultiGroup}: Runs the EdgeR GLM test method for many groups.
}}

\author{
Rehrauer, Hubert

Schmid, Peter
}
