% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ncpro.R
\name{ncpro}
\alias{ncpro}
\title{Analysis of small RNA sequences using ncpro}
\usage{
ncpro(input, dataset, param = NULL)
}
\arguments{
\item{input}{list of input configuration}

\item{dataset}{dataframe with description of input dataset}

\item{param}{further configuration parameters}
}
\description{
Trimming is done using ezMethodTrim which is similar to trimMirna().
}
