% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-edgerTwoGroups.R
\name{ngsTwoGroupAnalysis}
\alias{ngsTwoGroupAnalysis}
\alias{runNgsTwoGroupAnalysis}
\title{Prepares the two group analysis}
\usage{
ngsTwoGroupAnalysis(input = NA, output = NA, param = NULL,
  htmlFile = "00index.html")

runNgsTwoGroupAnalysis(dataset, htmlFile = "00index.html", param = param,
  rawData = NULL, types = NULL)
}
\arguments{
\item{input}{an object of the class EzDataset.}

\item{output}{an object of the class EzDataset.}

\item{param}{a list of parameters:
\itemize{
  \item{grouping}{ a character specifying the grouping.}
  \item{sampleGroup}{ a character specifying the group to sample.}
  \item{refGroup}{ a character specifying the reference group.}
  \item{useFactorsAsSampleName}{ a logical indicating whether to use the factors as sample names.}
  \item{removeOutliers}{ a logical indicating whether to remove outliers.}
  \item{batch}{ a character vector specifying the batch groups.}
  \item{markOutliers}{ a logical indicating whether to mark outliers.}
}}

\item{htmlFile}{a character representing the path to write the report in.}
}
\value{
Returns the results of the analysis.
}
\description{
Prepares the two group analysis by checking some parameters and setting up the options. Then \code{runNgsTwoGroupAnalysis()} gets called to obtain the result.
}
\section{Functions}{
\itemize{
\item \code{runNgsTwoGroupAnalysis}: Checks for errors and calls \code{twoGroupCountComparison()}, then \code{writeNgsTwoGroupReport()}.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{twoGroupCountComparison}}

\code{\link{writeNgsTwoGroupReport}}
}

