% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reports.R
\name{openBsdocReport}
\alias{closeBsdocReport}
\alias{openBsdocReport}
\title{Opens an html report}
\usage{
openBsdocReport(title = "", dataset = NULL)

closeBsdocReport(doc, file)
}
\arguments{
\item{title}{a character specifying the title of the html report.}

\item{dataset}{usually a data.frame from the meta field of an EzDataset.}
}
\value{
Returns an object of the class bsdoc to add further elements.
}
\description{
Opens an html report using \code{bsdoc()} from the ReporteRs package. Also adds some introductory elements.
}
\section{Functions}{
\itemize{
\item \code{closeBsdocReport}: Adds a paragraph showing the finishing time and writes the document. \code{file} must have a .html suffix.
}}
\examples{
theDoc = openBsdocReport(title="My html report")
closeBsdocReport(doc=theDoc, file="example.html")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[ReporteRs]{bsdoc}}

\code{\link[ReporteRs]{writeDoc}}
}

