% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{openHtmlReport}
\alias{ezBannerFile}
\alias{ezCSSFile}
\alias{openHtmlReport}
\title{Writes an html report}
\usage{
openHtmlReport(htmlFile, param = param, title = "", dataset = NULL)

ezCSSFile()

ezBannerFile()
}
\arguments{
\item{htmlFile}{a character representing the path to write the report in.}

\item{param}{a list of parameters to extract the \code{projectId} from.}

\item{title}{a character specifying the title of the html report.}

\item{dataset}{a data.frame from the meta field of an EzDataset.}
}
\value{
Returns a connection to the written html report.
}
\description{
Writes an html report and returns the connection to it.
}
\section{Functions}{
\itemize{
\item \code{ezCSSFile}: Returns a link to the report css file if it exists.

\item \code{ezBannerFile}: Returns a link to the banner png file if it exists.
}}
\examples{
param = ezParam()
htmlFile = "example_html"
openHtmlReport(htmlFile,param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{writeJavaScriptIgvStarter}}

\code{\link{writeTxtLinksToHtml}}
}

