% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{ordPlot}
\alias{ordPlot}
\title{Rank-specific tax-based ordination plot}
\usage{
ordPlot(rank, fullObject, type, areThereMultVar, numTopRanks, isGroupThere)
}
\arguments{
\item{fullObject}{(phyloseq object),x (rank)}
}
\value{
Returns a ggplot
}
\description{
ordination plot for a specific rank
}
