% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatranscriptomicsAssociatedFunctions.R
\name{orgFuncHeatmapPrep}
\alias{orgFuncHeatmapPrep}
\title{Organism/function heatmap}
\usage{
orgFuncHeatmapPrep(listOfAnnotatedAbundTable, diffTableFunc)
}
\arguments{
\item{a}{list of annotated org/func abundance table}
}
\value{
Returns a data.frame
}
\description{
Prepares organsim/function matrix for heatmap
}
