% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatranscriptomicsAssociatedFunctions.R
\name{orgFuncHeatmaptoBoxPlotPrep}
\alias{orgFuncHeatmaptoBoxPlotPrep}
\title{Organism/function heatmap-to-boxplot preparation}
\usage{
orgFuncHeatmaptoBoxPlotPrep(x, designMatrix, group, N)
}
\arguments{
\item{a}{heatmap-like dataframe}
}
\value{
Returns a data.frame
}
\description{
Gathers organsim/function from heatmap to boxplot DF
}
