% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R,
%   R/mothurDataCleanAssociatedFunctions.R
\name{otuSaturationPlot}
\alias{otuSaturationPlot}
\title{OTUs saturation}
\usage{
otuSaturationPlot(x, type)

otuSaturationPlot(x, type)
}
\arguments{
\item{x, }{mothur shared abundance  file or already read-in table (dep on sec. param).}

\item{sharedFile, }{mothur shared abundance  file.}
}
\value{
Returns a table

Returns a grid of plots
}
\description{
HOw many OTUs do we really have?

HOw many OTUs do we really have?
}
