% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R,
%   R/mothurErrorEstimateAssociatedFunctions.R
\name{otuSaturationTable}
\alias{otuSaturationTable}
\title{OTUs saturation table}
\usage{
otuSaturationTable(sharedFile)

otuSaturationTable(sharedFile)
}
\arguments{
\item{sharedFile, }{mothur shared abundance  file.}
}
\value{
Returns a table

Returns a table
}
\description{
HOw many OTUs do we really have?

HOw many OTUs do we really have?
}
