% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt.R
\name{pasteExcerptCmd}
\alias{pasteExcerptCmd}
\title{create appropriate command line call.}
\usage{
pasteExcerptCmd(readFile, outputDir, param)
}
\arguments{
\item{readFile}{<string> full path of the *.fastq(.gz) input file.}

\item{outputDir}{<string> directory in which to save the processed outputs created by exceRpt_smallRNA.}

\item{param}{list() exceRpt_smallRNA parameters.}
}
\description{
paste character strings to create the adequate shell command to execute exceRpt_smallRNA.
}
\author{
Miquel Anglada Girotto
}
