% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{pcaForPhyloseqPlot}
\alias{pcaForPhyloseqPlot}
\title{Alternative pca plot for phyloseq abundnce-taxonimy matrix}
\usage{
pcaForPhyloseqPlot(phySeqObject, type, group)
}
\arguments{
\item{phySeq, rank}{a phyloseq otu object and the number or top ranked categpries}
}
\value{
Returns an MDS   plot.
}
\description{
Alternative pca plot for phyloseq abundnce-taxonimy matrix
}
