% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqCommunityComp}
\alias{phyloSeqCommunityComp}
\title{Rank summary plots}
\usage{
phyloSeqCommunityComp(physeq)
}
\arguments{
\item{phySeq, rank}{a phyloseq object and the rank to summarize}
}
\value{
Returns a stacked bar  plot.
}
\description{
Summarized the top present genera
}
