% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqDivPlotAndPercUnclassified}
\alias{phyloSeqDivPlotAndPercUnclassified}
\title{Richness plot for a certain taxa}
\usage{
phyloSeqDivPlotAndPercUnclassified(taxaFileName, rank)
}
\arguments{
\item{taxaFileName, rank}{mothur taxonomy file.}
}
\value{
Returns a pie chart plot.
}
\description{
Create pie chart richness plot
}
