% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R,
%   R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqOTU}
\alias{phyloSeqOTU}
\title{Phyloseq OTU object}
\usage{
phyloSeqOTU(otuDF)

phyloSeqOTU(otuDF)
}
\arguments{
\item{otuFileName, }{mothur shared clustered OTU  files.}

\item{a}{data frame in the format of mothur shared clustered OTU  files.}
}
\value{
Returns a Phyloseq OTU object.

Returns a Phyloseq OTU object.
}
\description{
Create Phyloseq OTU object mothur OTU files.

Create Phyloseq OTU object mothur OTU files.
}
