% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqOTUFromFile}
\alias{phyloSeqOTUFromFile}
\title{Phyloseq OTU object}
\usage{
phyloSeqOTUFromFile(otuFile)
}
\arguments{
\item{a}{data frame in the format of mothur shared clustered OTU  files.}
}
\value{
Returns a Phyloseq OTU object.
}
\description{
Create Phyloseq OTU object mothur OTU files.
}
