% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqPreprocess}
\alias{phyloSeqPreprocess}
\title{Phyloseq preprocess}
\usage{
phyloSeqPreprocess(phyloseqObj, rawCount, sampleFraction)
}
\arguments{
\item{phyloseqObj, }{a phyloseq object.}
}
\value{
Returns a  filtered Phyloseq  object.
}
\description{
Preprocesses a phyloseq object.
}
