% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R,
%   R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqSample}
\alias{phyloSeqSample}
\title{Phyloseq Sample object}
\usage{
phyloSeqSample(sampleFileName)

phyloSeqSample(sampleFileName)
}
\arguments{
\item{taxaFileName, }{mothur taxonomy file.}

\item{path}{to the tsv desing file}
}
\value{
Returns a Phyloseq Taxa object.

Returns a Phyloseq Taxa object.
}
\description{
Create Phyloseq sample object from dataset.

Create Phyloseq sample object from dataset.
}
