% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R,
%   R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqTaxa}
\alias{phyloSeqTaxa}
\title{Phyloseq Taxa object}
\usage{
phyloSeqTaxa(taxaDB)

phyloSeqTaxa(taxaDB)
}
\arguments{
\item{taxaDB, }{a DF in the format of mothur taxonomy file.}

\item{taxaFileName, }{mothur taxonomy file.}
}
\value{
Returns a Phyloseq Taxa object.

Returns a Phyloseq Taxa object.
}
\description{
Create Phyloseq taxa object from mothur taxonomy files.

Create Phyloseq taxa object from mothur taxonomy files.
}
