% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqTaxaFromFile}
\alias{phyloSeqTaxaFromFile}
\title{Phyloseq Taxa object}
\usage{
phyloSeqTaxaFromFile(taxaFile)
}
\arguments{
\item{taxaDB, }{a DF in the format of mothur taxonomy file.}
}
\value{
Returns a Phyloseq Taxa object.
}
\description{
Create Phyloseq taxa object from mothur taxonomy files.
}
