% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{phyloSeqToDeseq2_tableAndPlots}
\alias{phyloSeqToDeseq2_tableAndPlots}
\title{Differential abundance analysis between groups}
\usage{
phyloSeqToDeseq2_tableAndPlots(phyloseqObj, rank, group, sampleGroup, refGroup)
}
\arguments{
\item{phyloseqObj, }{a phyloseq object.}
}
\value{
Returns a list of tables and plots.
}
\description{
Comaprison of metagenomics communities from files stored as a phyloseq object.
}
