% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamStats-reports.r
\name{plotBamStat}
\alias{getTypeCountTable}
\alias{getTypeCoverageTable}
\alias{makeAlignmentCountBarPlot}
\alias{plotBamStat}
\alias{plotPosSpecificErrorRate}
\title{Plots the BAM statistics}
\usage{
plotBamStat(resultList, dataset, param, htmlFile = NULL)

getTypeCountTable(resultList, name)

getTypeCoverageTable(resultList, name)

makeAlignmentCountBarPlot(file, mmCounts)

plotPosSpecificErrorRate(errorRate, png, main = "Per base mismatch rate",
  writeTxt = TRUE)
}
\arguments{
\item{resultList}{a list of results.}

\item{dataset}{a data.frame from the meta field of an EzDataset.}

\item{param}{a list of parameters:
\itemize{
  \item{name}{ a character representing the name of the app.}
  \item{writeIgvSessionLink}{ a logical indicating whether to write an IGV session link.}
  \item{ezRef@refBuild}{ a character containing the path of the reference build.}
  \item{sigThresh}{ an integer specifying the significance threshold.}
  \item{normMethod}{ a character specifying the normalization method to use.}
  \item{runGO}{ a logical indicating wheter to run the GO analysis.}
}}

\item{htmlFile}{a character, ending with .html, representing the file path of an html file.}
}
\description{
Plots the BAM statistics and writes the report.
}
\section{Functions}{
\itemize{
\item \code{getTypeCountTable}: Gets the table containing the counts of each type.

\item \code{getTypeCoverageTable}: Gets the table containing the coverage of each type.

\item \code{makeAlignmentCountBarPlot}: Plots the alignment counts and returns the image file link.

\item \code{plotPosSpecificErrorRate}: Plots the position specific error rates.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

