% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{plotBarMod}
\alias{plotBarMod}
\title{modified phyloseq barplot}
\usage{
plotBarMod(xx, x, fill = NULL, title = NULL, facet_grid = NULL, group)
}
\arguments{
\item{physeqFullObject}{(phyloseq object)}
}
\value{
A ggplot
}
\description{
Removed stacks border
}
