% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt_report.R
\name{PlotData}
\alias{PlotData}
\title{Make QC plots}
\usage{
PlotData(
  sampleIDs,
  output.dir,
  sampleGroups = NA,
  minPercent_exogenousRibosomal = 0.5,
  minPercent_exogenousGenomes = 0.5
)
}
\arguments{
\item{sampleIDs}{<string> names of the samples obtained through readData()}

\item{output.dir}{<string> directory to which output the files to create the report.}

\item{sampleGroups}{<string> if any sample groups were defined. Deprecated.}

\item{minPercent_exogenousRibosomal}{<double> Not used.}

\item{minPercent_exogenousGenomes}{<double> Not used.}
}
\description{
Create QC plots from data obtained through readData()
}
\author{
(Rob Kitchen) Miquel Anglada Girotto
}
