% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurePlots.R
\name{plotLocusCoverageProfile}
\alias{plotLocusCoverageProfile}
\title{Title}
\usage{
plotLocusCoverageProfile(
  gRanges,
  bamFiles,
  gtfFile = NULL,
  ylim = c(0, 40),
  sizes = NULL,
  isSecondaryAlignment = NA,
  height = 10,
  width = 20,
  plotType = c("coverage", "sashimi")
)
}
\arguments{
\item{width}{}
}
\description{
Title
}
\examples{
\dontrun{
require(GenomicRanges)
require(ezRun)
gtfFile = "/srv/GT/reference/Homo_sapiens/Ensembl/GRCh38.p10/Annotation/Release_89-2017-05-31/Genes/genes.gtf"
myGeneRange = GRanges(seqnames="2", ranges=IRanges(start=1, end=1000, name="foo"))
plotLocusCoverageProfile(myGeneRange, bamFiles, gtfFile=gtfFile,
                         height=20, width=20)
}
}
