% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt_report.R
\name{printMessage}
\alias{printMessage}
\title{print info messages}
\usage{
printMessage(message = "")
}
\arguments{
\item{message}{<string>}
}
\description{
to print formatted info messages.
}
\author{
(Rob Kitchen)
}
