% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotationQC_associatedFunctions.R
\name{prodigalFileReport}
\alias{prodigalFileReport}
\title{Prepares all the prodigal-associated files}
\usage{
prodigalFileReport(x, meth)
}
\arguments{
\item{a}{pridgal.gff prediction file}
}
\value{
Returns the full DF and the subest to partial =00
}
\description{
Prepares all the prodigal-associated plots
}
