% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{rarefactionPlot}
\alias{rarefactionPlot}
\title{OTUs saturation plot}
\usage{
rarefactionPlot(physeqFullObject, type)
}
\arguments{
\item{x, }{mothur shared abundance  file or already read-in table (dep on sec. param).}
}
\value{
Returns a grid of plots
}
\description{
HOw many OTUs do we really have?
}
