% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-exceRpt_report.R
\name{readData}
\alias{readData}
\title{Get smallRNA counts and stats}
\usage{
readData(samplePathList, output.dir)
}
\description{
Get smallRNA counts and stats to return as final output and make QC plots.
}
\author{
(Rob Kitchen) Miquel Anglada Girotto
}
