% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-XeniumSeurat.R
\name{RunBanksy_v5}
\alias{RunBanksy_v5}
\title{Patched RunBanksy function for Seurat v5 compatibility
Uses 'layer' instead of deprecated 'slot' argument (SeuratObject >= 5.0.0)
Issue: https://github.com/satijalab/seurat/issues/10250}
\usage{
RunBanksy_v5(
  object,
  lambda,
  assay = "RNA",
  layer = "data",
  use_agf = FALSE,
  dimx = NULL,
  dimy = NULL,
  dimz = NULL,
  ndim = 2,
  features = "variable",
  group = NULL,
  split.scale = TRUE,
  k_geom = 15,
  n = 2,
  sigma = 1.5,
  alpha = 0.05,
  k_spatial = 10,
  spatial_mode = "kNN_median",
  assay_name = "BANKSY",
  M = NULL,
  chunk_size = NULL,
  parallel = FALSE,
  num_cores = NULL,
  verbose = TRUE
)
}
\description{
Patched RunBanksy function for Seurat v5 compatibility
Uses 'layer' instead of deprecated 'slot' argument (SeuratObject >= 5.0.0)
Issue: https://github.com/satijalab/seurat/issues/10250
}
