% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R, R/app-DEXSeqCounting.R
\name{runCountSingleBam}
\alias{runCountSingleBam}
\title{Run counts for a single BAM file}
\usage{
runCountSingleBam(psBamFile, psGffFile, psCountfileExt)

runCountSingleBam(psBamFile, psGffFile, psCountfileExt)
}
\description{
Run counts for a single BAM file

Run counts for a single BAM file
}

