% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{runEnrichr}
\alias{runEnrichr}
\title{Run Enrichr}
\usage{
runEnrichr(genes, minScore = 12, maxAdjP = 0.01, connectionN = 10)
}
\arguments{
\item{genes}{A list or vector containing gene names}

\item{minScore}{Combined score (\eqn{log(AdjP) * z}) threshold to filter the results.}

\item{maxAdjP}{Adjusted p (Benjamini-Hochberg method) threshold to filter the results.}

\item{connectionN}{Maximum number of connections to open simultaneously for the enrichment analysis.}
}
\value{
A named \code{list} with library names as names and \code{data.frame} objects as values.
An entire result set is retained if there is at least one significant gene based on the supplied
criteria.
}
\description{
Runs Enrichr on the specified genes against all libraries

If you find that this function is not flexible enough, you can run Enrichr as follows.

\itemize{
\item Update the internal list of the libraries (if necessary) by running \code{\link{parseEnrichrLibNames}}
\item Register gene list with \code{\link{enrichrAddList}}
\item Run enrichment analysis with \code{\link{enrichrEnrich}}
\item Check the length \code{failure} element of the response to see if there were any failures
\item Filter the \code{success} element of the response by using \code{\link{filterEnrichrResults}}
}
}
\examples{
\dontrun{
genes <- c("HPS1", "HPS3", "HPS5", "PSMD11", "CUL7", "SNAPC4", "GNB2")
resList <- runEnrichr(genes, minScore = 15, maxAdjP = 1e-4, connectionN = 20)
}

# Alternative usage
\dontrun{
addResp <- enrichrAddList(genes)
names(addResp)
# [1] "shortId"    "userListId"
# This will check all libraries specified in the internal list (extdata/enrichr_libnames.txt)
res <- enrichrEnrich(genes)
# Check if any requests failed
if (length(res$failure) > 0) {
  cat(length(res$failure), " requests failed")
  allLibs <- getEnrichrLibNames()
  # Libraries that failed
  setdiff(allLibs, names(res$success))
}
# Filter the data sets
resFilt <- filterEnrichrResults(res$success, pAdj = 0.01, z = -3, combinedScore = 12)
}
}
\author{
Roman Briskine
}
