% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R
\name{separateGoIdsByOnto}
\alias{separateGoIdsByOnto}
\title{Separates GO ID's by ontology}
\usage{
separateGoIdsByOnto(goIdStrings)
}
\arguments{
\item{goIdStrings}{the GO ID's to separate.}
}
\value{
Returns the separated GO ID's
}
\description{
Separates GO ID's by ontology
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

