% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{setwdNew}
\alias{setwdNew}
\title{Switches to a working directory}
\usage{
setwdNew(dir)
}
\arguments{
\item{dir}{a character specifying the desired directory.}
}
\description{
If the directory does not exist, the function will create it recursively.
}
\examples{
cd = getwd()
setwdNew("newDirectory")
setwd(cd)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
