% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitessce-utils.R
\name{seurat_to_h5ad}
\alias{seurat_to_h5ad}
\title{Convert Seurat object to h5ad format}
\usage{
seurat_to_h5ad(seurat_obj, h5ad_path, nthreads = 8)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{h5ad_path}{Output path for h5ad file}

\item{nthreads}{Number of threads for parallel operations}
}
\value{
Path to created h5ad file
}
\description{
Convert Seurat object to h5ad format
}
