% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratUtils.R
\name{seuratMergeClusters}
\alias{seuratMergeClusters}
\title{Merge Seurat clusters}
\usage{
seuratMergeClusters(scData, clustList)
}
\description{
Given an input mapping, group Seurat clusters of the active Seurat Ident together, i.e. merging them
}
\section{Slots}{

\describe{
\item{\code{scData}}{the Seurat object}

\item{\code{clustList}}{a list() object where the names are the new cluster names and their values are all the clusters to be merged. If the list consists of multiple sets of clusters to be merged, ensure the sets do not overlap.}
}}

\examples{
data("pbmc_small")
clustList <- list("foo"=as.character(c(1, 2)))
seuratMergeClusters(pbmc_small, clustList)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
