% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsTools.R
\name{shiftZeros}
\alias{shiftZeros}
\title{Shifts zeros}
\usage{
shiftZeros(counts, minSignal, seed = 42)
}
\arguments{
\item{counts}{a set of numeric or integer values.}

\item{minSignal}{a numeric or integer specifying the minimal signal amount.}
}
\value{
Returns the modified count values.
}
\description{
Shifts zeros by changing values lower than \code{minSignal} to a random numeric in the range of (1/4 to 3/4) * \code{minSignal}. This is done to prevent zero signals to stack at 0.
}
\examples{
shiftZeros(1:10, 5)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
