% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{shrinkToRange}
\alias{shrinkToRange}
\title{Shrink the dynamic range of a numeric vector, matrix, or data frame}
\usage{
shrinkToRange(x, theRange)
}
\arguments{
\item{x}{the values to modify.}

\item{theRange}{two values specifying the lower and upper boundary of the range.}
}
\value{
Returns the shrunk object.
}
\description{
Values outside the range will be set as the lower or upper boundary of the range.
}
\examples{
shrinkToRange(1:10,c(2,6))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
