% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratUtils.R
\name{SpatiallyVariableFeatures_workaround}
\alias{SpatiallyVariableFeatures_workaround}
\title{This is work around function to replace SeuratObject::SpatiallyVariableFeatures function.
return ranked list of Spatially Variable Features}
\usage{
SpatiallyVariableFeatures_workaround(
  object,
  assay = "SCT",
  selection.method = "moransi"
)
}
\description{
This is work around function to replace SeuratObject::SpatiallyVariableFeatures function.
return ranked list of Spatially Variable Features
}
