% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-genome.R
\name{splitCoordinate}
\alias{splitCoordinate}
\alias{splitRegion}
\title{Splits fragment coordinates}
\usage{
splitCoordinate(pos)

splitRegion(pos)
}
\arguments{
\item{pos}{a character vector of chromosome ranges to split.}
}
\value{
Returns a data.frame containing the chromosome name, the start and the stop position.
}
\description{
Use this function to split chromosome range information in the format of chromosome:start-stop.
}
\section{Functions}{
\itemize{
\item \code{splitRegion()}: Does the same as \code{splitCoordinate}, but returns the result as a list and is more robust.

}}
\examples{
splitCoordinate("chrm:1-50")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
