% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ncpro.R
\name{splitCounts}
\alias{splitCounts}
\title{split count result into different files and put them into result dir}
\usage{
splitCounts(psCountType, pvSamples, plCountParam)
}
\arguments{
\item{psCountType}{for which count category {all, mature or precursor}
count results should be split}

\item{pvSamples}{vector of sample names}

\item{plCountParam}{list of parameters specifying existing result files}
}
\description{
\code{splitCounts} for a given count category, count results
for all samples given in pvSamples are written to different
files, according to parameters given in plCountParam
}
