% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rangesTools.r
\name{subSampleRle}
\alias{subSampleRle}
\title{1}
\usage{
subSampleRle(x, idx)
}
\arguments{
\item{x}{}
}
\value{
Returns
}
\description{
1
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

