% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{subsetRankTopN}
\alias{subsetRankTopN}
\title{Select only top10 rank}
\usage{
subsetRankTopN(physeqFullObject, rank, N)
}
\arguments{
\item{physeqFullObject}{(phyloseq object),x (rank)}
}
\value{
A subsetted phyloseq object
}
\description{
subset taxa to retain only top10 renks. Useful for plotting
}
