% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurPhyloseqAssociatedFunctions.R
\name{subsetTaxMod}
\alias{subsetTaxMod}
\title{Modifed subset taxa}
\usage{
subsetTaxMod(physeq, x)
}
\arguments{
\item{physeqFullObject}{(phyloseq object),x (rank)}
}
\value{
Returns a filtered phyloseqobj
}
\description{
Removes NA for a specific rank for better plotting
}
