% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotationQC_associatedFunctions.R
\name{summaryHeatmap}
\alias{summaryHeatmap}
\title{makeHeatmapFromSummbinFile}
\usage{
summaryHeatmap(
  mergedBinAbundTable,
  isGroupThere,
  dataset = NULL,
  numberOfTopNCategories,
  plotTitle = "plotTitle"
)
}
\arguments{
\item{the}{data frame generated by mergeSummaryBinFiles}
}
\value{
Returns a heatmap
}
\description{
taxa-by-sample heatmap
}
