% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotationQC_associatedFunctions.R
\name{summaryMatchScorePlot}
\alias{summaryMatchScorePlot}
\title{Prepares all the IPS-associated plots}
\usage{
summaryMatchScorePlot(x)
}
\arguments{
\item{a}{pridgal.gff prediction file}
}
\value{
Returns ggplots
}
\description{
Prepares all the prodigal-associated plots
}
