% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotationQC_associatedFunctions.R
\name{summaryMetagenomeBins}
\alias{summaryMetagenomeBins}
\title{summaryBins}
\usage{
summaryMetagenomeBins(krakenFile, binFiles)
}
\arguments{
\item{a}{kraken.labels files and fasta bin files}
}
\value{
Returns a DF summarizing bins
}
\description{
Create summary files for bins
}
