% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeAnnotationQC_associatedFunctions.R
\name{summaryScorePlot}
\alias{summaryScorePlot}
\title{Prepares all the prodigal-associated plots}
\usage{
summaryScorePlot(x)
}
\arguments{
\item{a}{pridgal.gff prediction file}
}
\value{
Returns ggplots
}
\description{
Prepares all the prodigal-associated plots
}
