% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-targetEnrichmentQC.R
\name{teqc}
\alias{runTEQC}
\alias{teqc}
\title{Target enrichment quality control}
\usage{
teqc(input, param = NULL)

runTEQC(file, param)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{param}{a list of parameters, some of which are passed to \code{runTEQC()}:
\itemize{
 \item{designFile}{ a file describing the regions selected by the enrichment kit.}
 \item{name}{ a character representing the name of the app.}
 \item{covUniformityPlot}{ a logical indicating whether to generate plots for coverage uniformity.}
 \item{covTargetLengthPlot}{ a logical indicating whether to generate plots for coverage vs. target length.}
 \item{duplicatesPlot}{ a logical indicating whether to generate plots for duplicates.}
 \item{paired}{ a logical indicating whether the samples are paired.}
}}

\item{file}{a character representing the path to the file containing the reads.}
}
\description{
Performs a target enrichment quality control and creates reports of the outcome.
}
\section{Functions}{
\itemize{
\item \code{runTEQC}: Runs the target enrichment QC.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

