% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{transcriptAnnoFromGtf}
\alias{transcriptAnnoFromGtf}
\title{Gets the transcript annotation}
\usage{
transcriptAnnoFromGtf(gtf, id = gtf$transcript_id, types = "exon",
  attributes = NULL)
}
\arguments{
\item{gtf}{an annotation data.frame of the gtf format.}

\item{id}{a column of the gtf data.frame that specifies a factor to group transcripts by.}

\item{types}{a character specifying which transcript types to use.}

\item{attributes}{additional annotation attributes to pass in.}
}
\value{
Returns a transposed transcript annotation data.frame.
}
\description{
Gets the transcript annotation from a gtf annotation.
}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param,"./inst/extdata/genes.gtf")
transcriptAnnoFromGtf(gtf)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

