% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trimWhiteSpace}
\alias{trimWhiteSpace}
\title{Trims white space}
\usage{
trimWhiteSpace(x)
}
\arguments{
\item{x}{the character to trim.}
}
\value{
Returns the trimmed character.
}
\description{
Removes white space at the end or in the front of a character.
}
\examples{
trimWhiteSpace("    bla    ")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
