% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-deseq2TwoGroups.R, R/app-edgerTwoGroups.R
\name{runDeseq2}
\alias{runDeseq2}
\alias{runEdger}
\alias{runGfold}
\alias{runGlm}
\alias{twoGroupCountComparison}
\title{1}
\usage{
runDeseq2(x, sampleGroup, refGroup, grouping, batch = NULL,
  isPresent = NULL)

twoGroupCountComparison(rawData, param)

runGfold(rawData, scalingFactors, isSample, isRef)

runEdger(x, sampleGroup, refGroup, grouping, normMethod)

runGlm(x, sampleGroup, refGroup, grouping, normMethod, batch = NULL)
}
\arguments{
\item{rawData}{}

\item{param}{a list of parameters:
\itemize{
  \item{testMethod}{}
  \item{batch}{}
  \item{grouping}{}
  \item{sampleGroup}{}
  \item{refGroup}{}
  \item{normMethod}{}
  \item{runGfold}{}
}}
}
\value{
Returns
}
\description{
1
}
\section{Functions}{
\itemize{
\item \code{runDeseq2}: Runs the Deseq2 test method.

\item \code{runGfold}: Runs the Gfold test method.

\item \code{runEdger}: Runs the Edger test method.

\item \code{runGlm}: Runs the Glm test method.
}}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

