% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R
\name{twoGroupsGO}
\alias{twoGroupsGO}
\alias{ezGoseq}
\title{Performs the GO analysis for two groups}
\usage{
twoGroupsGO(
  param,
  testResult,
  seqAnno,
  normalizedAvgSignal = NULL,
  method = "Wallenius"
)

ezGoseq(
  param,
  selectedGenes,
  allGenes,
  gene2goList = NULL,
  method = c("Wallenius", "Sampling", "Hypergeometric"),
  onto = NULL,
  normalizedAvgSignal = NULL
)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{featureLevel}{ which feature level to use.}
  \item{pValThreshGO}{ a numeric specifying the threshold for the GO p-value.}
  \item{log2RatioThreshGO}{ a numeric specifying the threshold for the GO log2 ratios.}
  \item{includeGoParentAnnotation}{ a logical indicating whether to include the annotation of the GO parents.}
}}

\item{testResult}{a list containing the results of an earlier test.}

\item{seqAnno}{the sequence annotation.}

\item{normalizedAvgSignal}{an optional normalized average signal.}

\item{method}{which method to pass to \code{ezGoseq()}.}
}
\value{
Returns the results of the GO analysis.
}
\description{
Performs the GO analysis for two groups.
}
\section{Functions}{
\itemize{
\item \code{ezGoseq}: Performs the GO analysis and returns a list of results.
}}

\seealso{
\code{\link{ezGoseq}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
