% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vennFromSets}
\alias{vennFromSets}
\title{Creates a Venn diagram based on the overlapping elements in vectors}
\usage{
vennFromSets(setList)
}
\arguments{
\item{setList}{a named list of length two or three.}
}
\description{
This is a convenience function that converts the list of vectors first into a logical matrix and subsequently applies the venn diagramm function from the limma package
}
\examples{
aList = list(a=1:5,b=3:6)
vennFromSets(aList)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
