% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01classes.R
\name{waitForFreeDiskSpace}
\alias{waitForFreeDiskSpace}
\alias{getGigabyteFree}
\title{Checks if there is enough free disk space}
\usage{
waitForFreeDiskSpace(param)

getGigabyteFree(dirPath)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{\code{scratch}}{ the required disk space in gigabytes.}
  \item{\code{mail}}{ the e-mail address of the recipient.}
}}

\item{dirPath}{a character specifying the path of the directory to check the disk space in.}
}
\description{
Checks if there is enough free disk space. If there is not enough disk space, an e-mail will be sent and the job will be put on hold for up to two hours.
}
\section{Functions}{
\itemize{
\item \code{getGigabyteFree()}: Gets the number of free gigabytes.

}}
\examples{
param = list()
param[['mail']] = ''
param[['scratch']] = '100'
waitForFreeDiskSpace(param)
getGigabyteFree(".")
getGigabyteFree("/")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
