% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeCountResultSummary}
\alias{writeCountResultSummary}
\title{Writes a count result summary}
\usage{
writeCountResultSummary(html, param, result)
}
\arguments{
\item{html}{a connection to an html file.}

\item{param}{a list of parameters to influence the output:
\itemize{
 \item{batch}{ a logical indicating whether the second factor was used.}
 \item{comparison}{ which comparison was used.}
 \item{normMethod}{ the normalization method.}
 \item{sigThresh}{ the threshold...}
 \item{useSigThresh}{ ...and whether it should be used.}
}}

\item{result}{a list of results.}
}
\description{
Writes a count result summary to an html file.
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

