% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\name{writeDEXSeqReport}
\alias{writeDEXSeqReport}
\title{Writing a report for a DEXSeq analysis}
\usage{
writeDEXSeqReport(
  dataset,
  dexResult,
  htmlFile = "00index.html",
  sResultDir = "html",
  BPPARAM = bpparam()
)
}
\description{
Addition experimental conditions to input files
}
