% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeErrorHtml}
\alias{writeErrorHtml}
\title{Writes an error message}
\usage{
writeErrorHtml(htmlFile, param = param, dataset = NULL,
  error = "Unknown Error")
}
\arguments{
\item{htmlFile}{a character representing the path to write the file in.}

\item{param}{a list of parameters to extract the \code{projectId} and \code{name} from.}

\item{dataset}{usually a data.frame from the meta field of an EzDataset.}

\item{error}{a character vector representing the error message(s).}
}
\description{
Writes an error message to an html file. Also creates the file and closes it.
}
\examples{
param = ezParam()
htmlFile = "example_html"
writeErrorHtml(htmlFile, param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{openHtmlReport}}

\code{\link{closeHTML}}
}

