% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reports.R
\name{writeErrorReport}
\alias{writeErrorReport}
\title{Writes an error report}
\usage{
writeErrorReport(htmlFile, param = param, dataset = NULL,
  error = "Unknown Error")
}
\arguments{
\item{htmlFile}{a character representing the path to write the file in. Must have a .html suffix.}

\item{param}{a list of parameters to extract the \code{name} from.}

\item{dataset}{usually a data.frame from the meta field of an EzDataset.}

\item{error}{a character vector representing the error message(s).}
}
\description{
Writes an error report to an html file. Also creates the file and closes it.
}
\examples{
param = ezParam()
htmlFile = "example.html"
writeErrorReport(htmlFile, param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{openBsdocReport}}

\code{\link{closeBsdocReport}}
}

