% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeImageTableToHtml}
\alias{writeImageColumnToHtml}
\alias{writeImageRowToHtml}
\alias{writeImageTableToHtml}
\title{Writes an image table}
\usage{
writeImageTableToHtml(pngMatrix, con = stdout())

writeImageRowToHtml(pngNames, con = stdout())

writeImageColumnToHtml(pngNames, con = stdout())
}
\arguments{
\item{pngMatrix}{a matrix containing picture files.}

\item{con}{the connection to the file to write in.}
}
\description{
Writes an image table to an html file, usually containing plot windows.
}
\section{Functions}{
\itemize{
\item \code{writeImageRowToHtml}: Write the pictures in a row instead.

\item \code{writeImageColumnToHtml}: Write the pictures in a column instead.
}}
\examples{
param = ezParam()
htmlFile = "example_html"
html =  openHtmlReport(htmlFile,param)
writeImageTableToHtml(as.matrix("example"),html)
writeImageRowToHtml("example",html)
writeImageColumnToHtml("example",html)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

