% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igv.r
\name{writeJavaScriptIgvStarter}
\alias{writeJavaScriptIgvStarter}
\title{Writes a jnlp link to a static html}
\usage{
writeJavaScriptIgvStarter(htmlFile, projectId, html)
}
\arguments{
\item{htmlFile}{the file to send to the html connection.}

\item{projectId}{a character representing the project ID.}

\item{html}{a character representing an html address.}
}
\description{
Writes a jnlp link to a static html. The html can be moved as long as the session xml stays in the same directory.
}
\examples{
writeJavaScriptIgvStarter()
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

