% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurDataCleanAssociatedFunctions.R
\name{writeOTUgzFileForVamps}
\alias{writeOTUgzFileForVamps}
\title{OTUs saturation table conversion for VAMPs}
\usage{
writeOTUgzFileForVamps(sharedFile, taxaFile)
}
\arguments{
\item{sharedFile, taxafile}{mothur shared abundance and taxa files.}
}
\value{
Writes a gzipped file
}
\description{
writeOTUgzFileForVamps
}
