% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gff.r
\name{writePresplicedGtf}
\alias{writePresplicedGtf}
\title{Writes prespliced annotation}
\usage{
writePresplicedGtf(param, featureFile = param$ezRef["refFeatureFile"])
}
\arguments{
\item{param}{contains the feature file to load the annotation from.}

\item{featureFile}{the file to load the features in the \code{ezLoadFeatures()} call from.}
}
\description{
Writes prespliced annotation directly from the parameters into a new .gtf file.
}
\examples{
param = ezParam()
writePresplicedGtf(param,"./inst/extdata/genes.gtf")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{ezLoadFeatures}}
}

