% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeQcScatterPlots}
\alias{writeQcScatterPlots}
\title{Writes QC scatter plots}
\usage{
writeQcScatterPlots(html, param, design, conds, rawData, signalCond,
  isPresentCond, seqAnno, colors = ezRedBlueScale(255), types = NULL)
}
\arguments{
\item{html}{a connection to an html file.}

\item{param}{a list of parameters. If \code{writeScatterPlots} is false, the function returns NULL.}

\item{design}{a data.frame containing the factorial design.}

\item{conds}{a named character vector containing the conditions of the factorial design.}

\item{rawData}{a list of raw data. Usually obtained from \code{loadCountDataset()}.}

\item{signalCond}{a set of values containing signals averaged by the conditions.}

\item{isPresentCond}{Either NULL or a data.frame containing coloring information.}

\item{colors}{a character vector containing rgb codes. The default is a scale from blue to red.}

\item{types}{a character vector containing the types.}
}
\description{
Writes QC scatter plots to an html file.
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

