% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeResultCounts}
\alias{writeResultCounts}
\title{1}
\usage{
writeResultCounts(html, param, result, geneIds = NULL, pThresh = c(0.1,
  0.05, 1/10^(2:5)))
}
\arguments{
\item{html}{a connection to an html file.}
}
\description{
1
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{writeTableToHtml}}
}

