% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeResultSummary}
\alias{writeResultSummary}
\title{Writes a summary of statistical tests}
\usage{
writeResultSummary(html, param, result, type)
}
\arguments{
\item{html}{a connection to an html file.}

\item{param}{a list of parameters to influence the output:
\itemize{
 \item{comparison}{ which comparison was used.}
 \item{nativeCdf}{ the chip type.}
 \item{cdfName}{ the name of the CDF that was used.}
 \item{AffyPreprocessing}{ the preprocessing algorithm.}
 \item{AgilentSignalColumn}{ which data was used.}
 \item{normMethod}{ the normalization method.}
 \item{runGO}{ a logical indicating whether the GO analysis ran or not.}
 \item{runMetaCore}{ }
 \item{sigThresh}{ the threshold...}
 \item{useSigThresh}{ ...and whether it should be used.}
}}

\item{result}{\itemize{
 \item{analysis}{ which analysis was used.}
 \item{method}{ which method was used.}
 \item{usedInTest}{ counts the number of probes.}
 \item{isPresentProbe}{ counts the number of probes with present signals.}
}}

\item{type}{a character vector .........}
}
\description{
Writes a summary of statistical tests to an html file.
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

