% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeTableToHtml}
\alias{writeTableToHtml}
\title{Writes a table}
\usage{
writeTableToHtml(x, con = stdout(), bgcolors = matrix("#ffffff", nrow =
  nrow(x), ncol = ncol(x)), valign = "middle", border = 1, head = "")
}
\arguments{
\item{x}{a matrix or data.frame to paste a table from.}

\item{con}{the connection to the file to write in.}

\item{bgcolors}{a matrix specifying the background colors in html format.}

\item{valign}{a character specifying with an html tag where to align the table vertically.}

\item{border}{an integer or numeric specifying the border width.}

\item{head}{a character specifying the header of the table.}
}
\description{
Writes a table to an html file.
}
\examples{
x = matrix(1:25,5)
rownames(x) = letters[1:5]
colnames(x) = LETTERS[1:5]
param = ezParam()
htmlFile = "example_html"
html =  openHtmlReport(htmlFile,param)
writeTableToHtml(x,html)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

