% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeTestScatterPlots}
\alias{writeTestScatterPlots}
\title{Writes test scatter plots}
\usage{
writeTestScatterPlots(html, param, x, result, seqAnno, colorRange = c(-3, 3),
  colors = getBlueRedScale(255), types = NULL)
}
\arguments{
\item{html}{a connection to an html file.}

\item{param}{a list of parameters .....}

\item{x}{a dataset .....}

\item{result}{a list of results.}

\item{seqAnno}{the sequence annotation. This is used if types is NULL.}

\item{types}{a character vector containing the types.}
}
\value{
Returns
}
\description{
Writes test scatter plots to an html file.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

