% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeTxtLinksToHtml}
\alias{writeTxtLinksToHtml}
\title{Writes a text link}
\usage{
writeTxtLinksToHtml(txtNames, mime = "text/plain", con = stdout())
}
\arguments{
\item{txtNames}{a character representing the link name.}

\item{mime}{a character representing the type of the link.}

\item{con}{the connection to the file to write in.}
}
\description{
Writes a text link to an html file.
}
\examples{
param = ezParam()
htmlFile = "example_html"
html =  openHtmlReport(htmlFile,param)
writeTxtLinksToHtml("dataset.tsv", con=html)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

