% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{zipWorkingDir}
\alias{zipFile}
\alias{zipWorkingDir}
\title{Archives files}
\usage{
zipWorkingDir(zipName)

zipFile(inputs, zipped = NULL)
}
\arguments{
\item{zipName}{a character naming the working directory to be archived.}

\item{inputs}{one or several input files to archive.}

\item{zipped}{optional character to name the archive.}
}
\value{
Returns the file name of the zipped archive.
}
\description{
Archives one or several files to a .zip extension. Optionally, a seperate name can be given to the archive.
}
\section{Functions}{
\itemize{
\item \code{zipWorkingDir}: Archives the whole working directory.
}}
\examples{
write("a file","exampleFile")
zipFile("exampleFile")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

