% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addDataset}
\alias{addDataset}
\title{Adds a dataset}
\usage{
addDataset(doc, dataset)
}
\arguments{
\item{doc}{an object of the class bsdoc to add the dataset.}

\item{dataset}{a data.frame from the meta field of an EzDataset.}
}
\description{
Adds a dataset to a bsdoc object.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

