% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-reports.R
\name{addQcScatterPlots}
\alias{addQcScatterPlots}
\title{Adds QC scatter plots}
\usage{
addQcScatterPlots(
  doc,
  param,
  design,
  conds,
  rawData,
  signalCond,
  isPresentCond,
  types = NULL
)
}
\arguments{
\item{doc}{an object of the class bsdoc to add the plots.}

\item{param}{a list of parameters.}

\item{design}{a data.frame containing the factorial design.}

\item{conds}{a named character vector containing the conditions of the factorial design.}

\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}

\item{signalCond}{a set of values containing signals averaged by the conditions.}

\item{isPresentCond}{Either NULL or a data.frame containing coloring information.}

\item{types}{a character vector containing the types.}
}
\description{
Adds QC scatter plots to an html file.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
