% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addTxtLinksToReport}
\alias{addTxtLinksToReport}
\title{Adds text links}
\usage{
addTxtLinksToReport(doc, txtNames, doZip = FALSE, mime = ifelse(doZip,
  "application/zip", "application/text"))
}
\arguments{
\item{doc}{an object of the class bsdoc to add the text links.}

\item{txtNames}{a character representing the file names.}

\item{doZip}{a logical indicating whether to zip the files.}

\item{mime}{a character representing the type of the links.}
}
\description{
Adds texts link to an html file.
}
\examples{
param = ezParam()
htmlFile = "example.html"
doc =  openBsdocReport(title="My report")
addTxtLinksToReport(doc, "dataset.tsv")
closeBsdocReport(doc, htmlFile)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

