% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{addTxtLinksToReport}
\alias{addTxtLinksToReport}
\title{Writes a text link}
\usage{
addTxtLinksToReport(txtNames, mime = "text/plain", doc)
}
\arguments{
\item{txtNames}{a character representing the link name.}

\item{mime}{a character representing the type of the link.}

\item{con}{the connection to the file to write in.}
}
\description{
Writes a text link to an html file.
}
\examples{
param = ezParam()
htmlFile = "example_html"
html =  openHtmlReport(htmlFile,param)
writeTxtLinksToHtml("dataset.tsv", con=html)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

