% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{aggregateCountsByGene}
\alias{aggregateCountsByGene}
\title{Aggregates counts by gene}
\usage{
aggregateCountsByGene(param, rawData)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{geneColumnSet}{ a character naming the columns.}
  \item{sigThresh}{ the threshold...}
  \item{useSigThresh}{ ...and whether it should be used.}
}}

\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}
}
\value{
Returns the aggregates raw data.
}
\description{
Aggregates counts by gene.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{getGeneMapping}}
}

