% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{aggregateCountsByGene}
\alias{aggregateCountsByGene}
\title{Aggregates counts by gene}
\description{
  Aggregates counts from isoform level to gene level.
}

\usage{
  aggregateCountsByGene(rawData)
}
\arguments{
  \item{rawData}{\code{SummarizedExperiment} obtained from \code{loadCountDataset()} or 
    \code{loadSCCountDataset()}.}
}
\value{
  \code{SummarizedExperiment} object with counts on gene level.
}
\seealso{
  \code{\link{getGeneMapping}}
}
\author{
  Ge Tan
}
