% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurDataCleanAssociatedFunctions.R,
%   R/mothurPhyloseqAssociatedFunctions.R
\name{chimeraSummaryPlot}
\alias{chimeraSummaryPlot}
\title{Chimera identification}
\usage{
chimeraSummaryPlot(chimeraDF)

chimeraSummaryPlot(chimeraDF)
}
\arguments{
\item{chimerFile}{ezTable from  mothur chimera file.}

\item{a}{data.frame}
}
\value{
Returns a pie chart plot.

A ggplot
}
\description{
Summarizes chimera rates from chimera file

Generate chimera plot from chimera file
}
