% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-RnaBamStats.R
\name{computeBamStats}
\alias{computeBamStats}
\alias{getPosErrorFromBam}
\alias{ezPosSpecErrorRate}
\alias{getStatsFromBam}
\alias{getStatsFromBamParallel}
\alias{getStatsFromBamSingleChrom}
\alias{getTargetTypeCounts}
\alias{getJunctionPlotsFromBam}
\title{Computes the BAM statistics}
\usage{
computeBamStats(input, htmlFile, param, gff, resultList = NULL)

getPosErrorFromBam(bamFile, param)

ezPosSpecErrorRate(bam, ReferenceGenome, nMaxReads = 1e+05)

getStatsFromBam(param, bamFile, sm, gff = NULL, repeatsGff = NULL, nReads = NA)

getStatsFromBamParallel(
  seqLengths,
  param,
  bamFile,
  sm,
  gff = NULL,
  repeatsGff = NULL,
  mc.cores = ezThreads(),
  nReads = NA
)

getStatsFromBamSingleChrom(
  chrom,
  param,
  bamFile,
  sm,
  gff = NULL,
  repeatsGff = NULL
)

getTargetTypeCounts(param, gff, rr, seqid = NULL, repeatsGff = NULL)

getJunctionPlotsFromBam(bamFile, param)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{htmlFile}{a character, ending with .html, representing the file path of an html file.}

\item{param}{a list of parameters:
\itemize{
  \item{ezRef@refBuild}{ a character containing the path of the reference build.}
  \item{seqNames}{ the sequence names to select.}
  \item{posErrorRates}{ a logical indicating whether to call \code{getPosErrorFromBam()}.}
  \item{saveImage}{ a logical indicating whether to save the image.}
}}

\item{gff}{an annotation data.frame in gtf or gff format.}

\item{resultList}{a list of results.}
}
\description{
Computes the BAM statistics.
}
\section{Functions}{
\itemize{
\item \code{getPosErrorFromBam}: Gets the error positions from the BAM file.

\item \code{ezPosSpecErrorRate}: Calculates the specific error rates for \code{getPosErrorFromBam()}.

\item \code{getStatsFromBam}: Gets the result statistics from the BAM file.

\item \code{getStatsFromBamParallel}: Gets parallel by chromosome statistics for \code{getStatsFromBam()} if the logical \code{param$splitByChrom} is true.

\item \code{getStatsFromBamSingleChrom}: Gets the statistics of a single chromosome for \code{getStatsFromBam()}.

\item \code{getTargetTypeCounts}: Gets the counts of the target types for \code{getStatsFromBam()}.

\item \code{getJunctionPlotsFromBam}: Gets the junction results from the BAM file.
}}

\seealso{
\code{\link{plotBamStat}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
