% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangesTools.R
\name{computeRangeStats}
\alias{computeRangeStats}
\title{Computes stats of genomic ranges}
\usage{
computeRangeStats(cov, targetRanges, FUN = mean)
}
\arguments{
\item{cov}{an object of the class RleList.}

\item{targetRanges}{an object of the class GRanges.}

\item{FUN}{a function to apply to the views.}
}
\value{
Returns the computed values belonging to their genomic ranges.
}
\description{
Computes stats of genomic ranges.
}
\seealso{
\code{\link[IRanges]{RleViewsList}}

\code{\link[IRanges]{viewApply}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
