% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\name{convertGtfToGff}
\alias{convertGtfToGff}
\title{Convert annotation file from GTF format to GFF}
\usage{
convertGtfToGff(psGtfFile, psGffFile)
}
\arguments{
\item{psGtfFile}{name of the GTF annotation file}

\item{psGffFile}{name of the GFF file to be generated}
}
\description{
\code{convertGtfToGff} converts an annotation file from
the GTF format into the GFF format which is required
by the package \code{DEXSeq}. Input file name and the
name of the file to be generated are both given as
function parameters. The conversion is done by a python
script that is given by the content of \code{DEXSEQ_PREPARE}
which is taken from the result of function
\code{lGetPyScriptPaths}
}
