% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2associatedFunctions.R
\name{DADA2CreateSeqTab}
\alias{DADA2CreateSeqTab}
\title{Main DADA2 workflow (sample-wise)}
\usage{
DADA2CreateSeqTab(
  sampleNames,
  maxLen = 0,
  file1PathInDataset,
  database,
  seqTech,
  maxExpErr
)
}
\arguments{
\item{a}{fastq or pairs of ffastq files .}
}
\value{
Returns a DADA2 seqtab object.
}
\description{
Create  OTU table.
}
