% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangesTools.R
\name{expandGRanges}
\alias{expandGRanges}
\title{Expands genomic ranges}
\usage{
expandGRanges(x, width = 2000)
}
\arguments{
\item{x}{an object of the class GRanges.}

\item{width}{an integer specifying how much to expand the ranges.}
}
\value{
Returns an expanded GRanges object.
}
\description{
Expands genomic ranges of a GRanges object. All ranges will be expanded at both ends by \code{width}.
}
\examples{
x = GRanges(c("chr1", "chr1", "chr2"), IRanges(5000:5002, 8000:8002), strand=c("+", "-", "+"))
expandGRanges(x)
}
\seealso{
\code{\link[GenomicRanges]{GRanges}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
