% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{ezAddTable}
\alias{ezAddTable}
\alias{ezAddTableWhite}
\title{Adds a table}
\usage{
ezAddTable(doc, x, bgcolors = NULL, valign = "middle", border = 1,
  head = "")

ezAddTableWhite(x, bgcolors = NULL, valign = "middle", border = 1,
  head = "")
}
\arguments{
\item{doc}{an object of the class bsdoc to add the table.}

\item{x}{a matrix or data.frame to paste a table from.}

\item{bgcolors}{a matrix specifying the background colors.}

\item{valign}{a character specifying where to align the table elements vertically. Use either "top", "middle" or "bottom".}

\item{border}{an integer specifying the border width.}

\item{head}{a character specifying the contents of the upperleft corner of the table.}
}
\description{
Adds a table to a bsdoc object.
}
\section{Functions}{
\itemize{
\item \code{ezAddTableWhite}: Does the same with a white font and returning the table instead of adding it to the document.
}}
\examples{
x = matrix(1:25,5)
rownames(x) = letters[1:5]
colnames(x) = LETTERS[1:5]
html = openBsdocReport()
ezAddTable(html, x, head="Example", bgcolors="red")
closeBsdocReport(html, "example.html")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[ReporteRs]{addFlexTable}}
}

