% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-countOverlaps.R
\docType{class}
\name{EzAppCountOverlaps-class}
\alias{EzAppCountOverlaps-class}
\alias{EzAppCountOverlaps}
\title{The R5 class representing a runnable app using \code{ezMethodCountOverlaps(input=NA, output=NA, param=NA)()}}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\description{
Use this reference class to run
}
\section{Fields}{

\describe{
\item{\code{runMethod}}{the function that will be executed in the \code{run} method.}

\item{\code{name}}{the name of the application.}

\item{\code{appDefaults}}{the defaults to run the application with.}
}}

\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the application using its specific defaults.}
}}

\section{Functions}{

\itemize{
  \item{\code{ezMethodCountOverlaps(input=NA, output=NA, param=NA) }}:
  {The function to run this application.}
}


\itemize{
  \item{\code{countNonredundant(bamFile, param=param, gff=gff): }}
  {Counts the non-redundant overlaps.}
  \item{\code{countBamHitsSingleChrom(chr, bamFile=NULL, param=NULL, gff=NULL): }}
  {Counts the BAM hits for a single chromosome.}
  \item{\code{countPairedBamHitsSingleChrom(chr, bamFile=NULL, param=NULL, gff=NULL): }}
  {Counts the paired BAM hits for a single chromosome.}
  \item{\code{getTargetRanges(gff, param, chrom=NULL): }}
  {Gets the target ranges depending on \code{param$featureLevel}.}
  \item{\code{getFeatureCounts(chrom, gff, reads, param): }}
  {Gets the feature counts from the target ranges.}
}
}

\seealso{
\code{\link{EzApp}}

\code{\link{EzDataset}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
