% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamio.R
\name{ezBam2bigwig}
\alias{ezBam2bigwig}
\title{Converts from bam to bigwig}
\usage{
ezBam2bigwig(bamFile, bigwigPrefix, param = NULL, paired = NULL)
}
\arguments{
\item{bamFile}{a character representing the bam file path.}

\item{bigwigPrefix}{a character representing a prefix to add to the bigwig file names.}

\item{param}{a list of parameter to extract the \code{strandMode} from.}

\item{paired}{a logical indicating whether the samples are paired.}
}
\description{
Converts a bam file into a bigwig file.
}
\examples{
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
ezBam2bigwig(bamFile, bigwigPrefix="bg", param=ezParam(), paired=TRUE)
}
\seealso{
\code{\link[Rsamtools]{scanBamHeader}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
