% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamio.R
\name{ezBamSeqNames}
\alias{ezBamSeqNames}
\alias{ezBamSeqLengths}
\title{Gets sequence names}
\usage{
ezBamSeqNames(bamFile, sizeSorting = "decreasing")

ezBamSeqLengths(bamFile)
}
\arguments{
\item{bamFile}{a character representing the bam file path.}

\item{sizeSorting}{a character. if equal to "decreasing" the names will be sorted decreasingly.}
}
\value{
Returns a character vector of sequence names.
}
\description{
Gets sequence names from a bam file.
}
\section{Functions}{
\itemize{
\item \code{ezBamSeqLengths()}: Gets the sequence lengths from a bam file.

}}
\examples{
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
ezBamSeqLengths(bamFile)
ezBamSeqNames(bamFile)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
