% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rangesTools.r
\name{ezCountOverlaps}
\alias{ezCountOverlaps}
\alias{ezHasOverlaps}
\title{1}
\usage{
ezCountOverlaps(query, subject, ignoreStrand = FALSE, ...)

ezHasOverlaps(query, subject, ignore.strand = FALSE, ...)
}
\arguments{
\item{x}{}
}
\value{
Returns
}
\description{
1
}
\section{Functions}{
\itemize{
\item \code{ezHasOverlaps}: 
}}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

