% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ezDesignFromDataset}
\alias{ezDesignFromDataset}
\alias{ezConditionsFromDesign}
\alias{ezConditionsFromDataset}
\alias{addReplicate}
\title{Gets the design from the dataset}
\usage{
ezDesignFromDataset(dataset, param = NULL)

ezConditionsFromDesign(design, maxFactors = 2)

ezConditionsFromDataset(dataset, param = NULL, maxFactors = 2)

addReplicate(x, sep = "_", repLabels = 1:length(x))
}
\arguments{
\item{dataset}{a data.frame from the meta field of an EzDataset.}

\item{param}{a list of parameters to specify the factors directly using \code{param$factors}.}
}
\value{
Returns the factorial design of the dataset.
}
\description{
Gets the design from the dataset or parameters, if specified.
}
\section{Functions}{
\itemize{
\item \code{ezConditionsFromDesign()}: Gets the conditions from the design. Use \code{maxFactors} to limit the amount of factors.

\item \code{ezConditionsFromDataset()}: A wrapper to get the conditions directly from the dataset.

\item \code{addReplicate()}: add a replicate id to make values unique

}}
\examples{
file = system.file("extdata/yeast_10k/dataset.tsv", package="ezRun", mustWork = TRUE)
ds = EzDataset$new(file=file, dataRoot=NULL)
design = ezDesignFromDataset(ds$meta)
cond = ezConditionsFromDesign(design)
addReplicate(apply(design, 1, paste, collapse="_"))
addReplicate(cond)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
