% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.r
\name{ezFeatureAnnotation}
\alias{aggregateGoAnnotation}
\alias{ezFeatureAnnotation}
\alias{writeAnnotationFromGtf}
\title{Gets the sequence annotation}
\usage{
ezFeatureAnnotation(param, ids, dataFeatureType)

writeAnnotationFromGtf(param, featureFile = param$ezRef["refFeatureFile"],
  featAnnoFile = param$ezRef["refAnnotationFile"])

aggregateGoAnnotation(seqAnno, genes, goColumns = c("GO BP", "GO CC",
  "GO MF"))
}
\arguments{
\item{param}{the parameters to load the annotation and sequence files from and possibly write to.}

\item{ids}{a character vector containing the gene ID's to return.}

\item{dataFeatureType}{either \code{"isoform"} or \code{"gene"}.}
}
\value{
Returns a data.frame containing information about the genes in an easily readable way.
}
\description{
Gets the sequence annotation from annotation (gtf or gff) and sequence (fasta) information.
}
\section{Functions}{
\itemize{
\item \code{writeAnnotationFromGtf}: Gets the annotation from a .gtf file and transforms it into a more readable .txt file.

\item \code{aggregateGoAnnotation}: Aggregates the Go annotation.
}}
\examples{
param = ezParam()
param$ezRef@refFeatureFile = system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE)
param$ezRef@refAnnotationFile = ""
fp = "/srv/GT/reference/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/WholeGenomeFasta/genome.fa"
param$ezRef@refFastaFile = fp
x = writeAnnotationFromGtf(param)
ezFeatureAnnotation(param, rownames(x), "gene")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

