% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{ezFlexTable}
\alias{ezFlexTable}
\alias{ezGrid}
\title{Wrapper for \code{FlexTable()}}
\usage{
ezFlexTable(
  x,
  border = 1,
  valign = "top",
  talign = "left",
  header.columns = FALSE,
  ...
)

ezGrid(x, header.columns = FALSE, valign = "top", ...)
}
\arguments{
\item{x}{a matrix or data.frame to turn into an object of the class FlexTable.}

\item{border}{an integer specifying the width of the table borders.}

\item{valign}{"bottom", "middle" or "top" specifying the position of table cell contents.}

\item{talign}{"left", "middle" or "right" specifying the position of text within the cells.}

\item{header.columns}{a logical indicating whether to use a header for the table.}

\item{...}{additional arguments to be passed to \code{FlexTable()}.}
}
\value{
Returns an object of the class FlexTable.
}
\description{
Wraps \code{FlexTable()} with defaults to remove the cell header and cell borders.
}
\section{Functions}{
\itemize{
\item \code{ezGrid}: A flex table without borders.
}}

\examples{
ezFlexTable(data.frame(a=1:5, b=11:15))
}
\seealso{
\code{\link[ReporteRs]{FlexTable}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
