% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00defaults.R
\name{ezFrame}
\alias{ezFrame}
\title{Wrapper for data.frame suitable for data processing rather than stastistical modelling}
\usage{
ezFrame(
  ...,
  row.names = NULL,
  check.rows = TRUE,
  check.names = FALSE,
  stringsAsFactors = FALSE
)
}
\arguments{
\item{...}{additional arguments to be passed to \code{data.frame()}.}
}
\value{
Returns a data.frame.
}
\description{
The original \code{data.frame()} is suitable for statistical modelling with factors etc.
Our version is designed to be more suitable for data processing, i.e. we
* do keep column names the way they are; we do not require that they represent valid variable names
* by default we do not convert strings to factors
There exist, similarities with data_frame 
in Hadleys package. The difference is that we encourage to use rownames while he discourages this.
}
\examples{
ezFrame(first=1:3, second=5, "with space"="text", row.names=letters[1:3])
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
