% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R
\name{ezGroupGO}
\alias{ezGroupGO}
\alias{ezGetGoByLevels}
\title{Groups GO terms and information}
\usage{
ezGroupGO(selectedGenes, go2GeneList, onto = "CC", levels = 2:4, goSlim = NULL)

ezGetGoByLevels(onto, levels, goSlim = NULL)
}
\arguments{
\item{selectedGenes}{a character vector containing the selected genes.}

\item{go2GeneList}{a character vector.}

\item{onto}{a character representing the ontology to use.}

\item{levels}{an integer vector selecting the levels to get GO information by.}

\item{goSlim}{a character vector.}
}
\value{
Returns the results of grouping the GO information.
}
\description{
Groups GO terms and information.
}
\section{Functions}{
\itemize{
\item \code{ezGetGoByLevels}: Gets the gene ontology by level.
}}

\author{
Rehrauer, Hubert

Schmid, Peter
}
