% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ezInteractiveTableRmd}
\alias{ezInteractiveTableRmd}
\title{Generates an interactive table}
\usage{
ezInteractiveTableRmd(
  values,
  digits = NULL,
  colNames = colnames(values),
  title = "",
  format = NULL,
  envir = parent.frame()
)
}
\arguments{
\item{values}{a data.frame or table to create an interactive table from.}

\item{digits}{the number of digits to round to, if rounding is desired.}

\item{colNames}{a character vector specifying the column names of the interactive table.}

\item{title}{a character representing the title of the interactive table.}

\item{format}{formatting options passed as an expression. The table argument in formatting functions must be named \code{interactiveTable}.}

\item{envir}{the environment to evaluate \code{format} in.}
}
\description{
Generates an interactive table embeded in rmarkdown document.
}
\examples{
table = data.frame(a=c(1.11, 2:100), b=201:300)
ezInteractiveTableRmd(table)
}
\seealso{
\code{\link[DT]{datatable}}

\code{\link[DT]{saveWidget}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
