% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00defaults.R
\name{ezIsSpecified}
\alias{ezIsSpecified}
\title{Is x specified?}
\usage{
ezIsSpecified(x)
}
\arguments{
\item{x}{usually a parameter to check.}
}
\value{
Returns FALSE or TRUE.
}
\description{
Checks whether \code{x} is an existing parameter and whether its first entry not an empty character.
}
\examples{
ezIsSpecified(c("this","is"))
ezIsSpecified(c("","this isn't"))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

