% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{ezLog}
\alias{ezLog}
\title{Wrapper for logger}
\usage{
ezLog(..., level = "info")
}
\arguments{
\item{...}{the message to log, will be concatenated.}

\item{level}{the log level (fatal, error, warn, success, info, debug, trace).}
}
\description{
Logs a message at a specified level using logger.
}
\examples{
ezLog("This is an info message")
ezLog("This is a warning message", level="warn")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
