% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-fastQC.R
\name{ezMethodFastQC}
\alias{ezMethodFastQC}
\title{The Fast QC method to run an app}
\usage{
ezMethodFastQC(input = NA, output = NA, param = NA,
  htmlFile = "00index.html")
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{a list, file path or an object of the class EzDataset containing the output information.}

\item{param}{a list of parameters to customize the application run.}
}
\value{
Returns "Success" if the method runs until the end. Also creates output files.
}
\description{
This is the method to run the app. Actually running it should be done using the \code{run} method of the app.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{EzDataset}}

\code{\link{EzAppFastqc}}
}

