% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-trim.R
\name{ezMethodTrim}
\alias{ezMethodSubsampleReads}
\alias{ezMethodTrim}
\alias{ezSubsampleFastq}
\title{Trims input reads}
\usage{
ezMethodTrim(input = NA, output = NA, param = NA)

ezMethodSubsampleReads(input = NA, output = NA, param = NA)

ezSubsampleFastq(full, sub, subsampleFactor = NA, nYield = 1e+05)
}
\arguments{
\item{input}{an object of the class EzDataset.}

\item{output}{an object of the class EzDataset or NA. If it is NA, it will be copied from the input.}

\item{param}{a list of parameters:
\itemize{
  \item{paired}{ a logical specifying whether the samples have paired ends.}
  \item{subsampleReads}{ an integer specifying how many subsamples there are. This will call \code{ezMethodSubsampleReads()} if > 1.}
  \item{trimAdapter}{ a logical specifying whether to use a trim adapter.}
  \item{minTailQuality}{ an integer specifying the minimal tail quality to accept. Only used if > 0.}
  \item{minAvgQuality}{ an integer specifying the minimal average quality to accept. Only used if > 0.}
  \item{minReadLength}{ an integer specifying the minimal read length to accept.}
  \item{dataRoot}{ a character specifying the path of the data root to get the full column paths from.}
}}
}
\value{
Returns the output after trimming as an object of the class EzDataset.
}
\description{
Trims input reads. There are several options to influence trimming with parameters.
}
\section{Functions}{
\itemize{
\item \code{ezMethodSubsampleReads}: Gets the subsample files, calls \code{ezSubsampleFastq()} on them and returns the output, which is an object of the class EzDataset.

\item \code{ezSubsampleFastq}: Performs the fastq for the subsamples using the package ShortRead.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

