% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-trim.R
\name{ezMethodTrim}
\alias{ezMethodTrim}
\title{Trims input reads}
\usage{
ezMethodTrim(input = NA, output = NA, param = NA)
}
\arguments{
\item{input}{a list, file path or an object of the class EzDataset containing the input.}

\item{output}{an object of the class EzDataset or NA. If it is NA, it will be copied from the input.}

\item{param}{a list of parameters:
\itemize{
  \item{paired}{ a logical specifying whether the samples have paired ends.}
  \item{subsampleReads}{ an integer specifying how many subsamples there are. This will call \code{ezMethodSubsampleReads()} if > 1.}
  \item{trimAdapter}{ a logical specifying whether to use a trim adapter.}
  \item{minTailQuality}{ an integer specifying the minimal tail quality to accept. Only used if > 0.}
  \item{minTrailingQuality}{ an integer specifying the minimal trailing quality to accept. Only used if > 0.}
  \item{minAvgQuality}{ an integer specifying the minimal average quality to accept. Only used if > 0.}
  \item{minReadLength}{ an integer specifying the minimal read length to accept.}
  \item{dataRoot}{ a character specifying the path of the data root to get the full column paths from.}
}}
}
\value{
Returns the output after trimming as an object of the class EzDataset.
}
\description{
Trims input reads. The trimming happens in the following order:
\itemize{
  \item{adapter}
  \item{quality window}
  \item{avg quality}
  \item{minimum length}
  \item{fixed trimming}
}
This means if you specify a fixed trimming this comes on top of the adaptive trimming
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
