% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00defaults.R
\name{ezParam}
\alias{ezParam}
\alias{parseOptions}
\title{Build parameter object}
\usage{
ezParam(userParam = list(), globalDefaults = EZ_PARAM_DEFAULTS,
  appDefaults = ezFrame())

parseOptions(optString)
}
\arguments{
\item{userParam}{a list of parameters defined by the user}

\item{globalDefaults}{a data.frame containing the global defaults}

\item{appDefaults}{a data.frame containing application specific defaults}
}
\value{
Returns the merged list of parameters
}
\description{
Parameters can be specified in the global defaults, app-specific defaults and user-given.
If a parameter is specified in multiple places, the user parameters override the app defaults which again
override the global defaults
}
\section{Functions}{
\itemize{
\item \code{parseOptions}: Used to parse additional options specified in \code{userParam$specialOptions}.
}}
\examples{
ezParam()
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

