% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00defaults.R
\name{ezParam}
\alias{ezParam}
\alias{parseOptions}
\title{Build parameter object}
\usage{
ezParam(
  userParam = list(),
  globalDefaults = getGlobalDefaults(),
  appDefaults = ezFrame()
)

parseOptions(optString)
}
\arguments{
\item{userParam}{a list of parameters defined by the user}

\item{globalDefaults}{a data.frame containing the global defaults for parameters. Includes for each parameter a name, 
a default value, the type, and a description. The global defaults are read from a file. 
See the file EZ_PARAM_DEFAULTS.txt in the package directory.}

\item{appDefaults}{a data.frame containing application specific defaults.
Must have the same columns as the globalDefaults data.frame.}
}
\value{
Returns the merged list of parameters
}
\description{
Parameters can be specified in the global defaults, app-specific defaults and user-given.
If a parameter is specified in multiple places, the user parameters override the app defaults which again
override the global defaults
}
\section{Functions}{
\itemize{
\item \code{parseOptions}: Used to parse additional options specified in \code{userParam$specialOptions}.
Converts an option specification in the from "key1=value1 key2=value2" into a named list
}}

\examples{
library(ezRun)
head(EZ_PARAM_DEFAULTS)
globalDefaultTypedParams = ezParam()
modifiedParams = ezParam(list(ram=40))
modifiedParams$ram
globalDefaultTypedParams$ram
parseOptions("a=5 b='foo with space' c=foo")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
