% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01classes.R
\docType{class}
\name{EzRef-class}
\alias{EzRef}
\alias{EzRef-class}
\alias{[,EzRef,ANY,ANY-method}
\alias{initialize,EzRef-method}
\title{The S4 class representing a collection of references}
\usage{
\S4method{initialize}{EzRef}(.Object, param = list())

\S4method{[}{EzRef,ANY,ANY}(x, i)
}
\description{
Objects of this class are a collection of references derived from a list of parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: Initializes the slots of EzRef. It will fill also try to specify some fields and if necessary get full file paths.

\item \code{[}: Allows selection with square brackets [ ].
}}
\section{Slots}{

\describe{
\item{\code{refBuild}}{}

\item{\code{refBuildName}}{}

\item{\code{refBuildDir}}{}

\item{\code{refIndex}}{}

\item{\code{refFeatureFile}}{}

\item{\code{refAnnotationFile}}{}

\item{\code{refFastaFile}}{}

\item{\code{refChromDir}}{}

\item{\code{refChromSizesFile}}{}
}}
\examples{
EzRef(param=list(refBuild="foo"))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

