% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02references.R
\docType{class}
\name{EzRef-class}
\alias{EzRef}
\alias{EzRef-class}
\alias{[,EzRef,ANY,ANY,ANY-method}
\alias{buildIgvGenome,EzRef-method}
\alias{buildRefDir,EzRef-method}
\alias{getOrganism,EzRef-method}
\alias{initialize,EzRef-method}
\title{The S4 class representing a collection of information about the reference build}
\usage{
\S4method{initialize}{EzRef}(.Object, param = list(),
  genomesRoot = GENOMES_ROOT)

\S4method{[}{EzRef,ANY,ANY,ANY}(x, i)

\S4method{getOrganism}{EzRef}(.Object)

\S4method{buildRefDir}{EzRef}(.Object, genomeFile, genesFile,
  genomesRoot = ".")

\S4method{buildIgvGenome}{EzRef}(.Object, param = NULL, addList = FALSE)
}
\description{
Objects of this class are a collection of references derived from a list of parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: Initializes the slots of EzRef. It will also try to specify some fields and if necessary get full file paths.

\item \code{[}: Allows selection with square brackets [ ].

\item \code{getOrganism}: Gets the organism name from the reference build.

\item \code{buildRefDir}: Builds the reference directory and copies the annotation and fasta file into the right folders.

\item \code{buildIgvGenome}: Builds the IGV genome.
}}
\section{Slots}{

\describe{
\item{\code{refBuild}}{a character specifying the file path to the reference build.}

\item{\code{refBuildName}}{a character specifying the name of the reference build.}

\item{\code{refBuildDir}}{a character specifying the directory of the reference build.}

\item{\code{refIndex}}{a character specifying the location of the index that is used in the alignment.}

\item{\code{refFeatureFile}}{a character specifying the file path to the annotation feature file (.gtf).}

\item{\code{refAnnotationFile}}{a character specifying the file path to the annotation file (.txt).}

\item{\code{refFastaFile}}{a character specifying the file path to the fasta file.}

\item{\code{refChromDir}}{a character specifying the file path to the directory of the chromosome information.}

\item{\code{refChromSizesFile}}{a character specifying the file path to the file containing the chromosome sizes.}

\item{\code{refAnnotationVersion}}{a character specifying the annotation version.}
}}
\examples{
refBuild = "Saccharomyces_cerevisiae/Ensembl/EF4/Annotation/Version-2013-03-18"
gtf = system.file("extdata/genes.gtf", package="ezRun", mustWork = TRUE)
fp = system.file("extdata/genome.fa", package="ezRun", mustWork = TRUE)
genomesRoot = "./refExample"
myRef = EzRef(param=ezParam(list(refBuild=refBuild)), genomesRoot=genomesRoot)
buildRefDir(myRef, fp, gtf)
buildIgvGenome(myRef)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{cleanGenomeFiles}}
}

