% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02references.R
\docType{class}
\name{EzRef-class}
\alias{EzRef-class}
\alias{EzRef}
\alias{[,EzRef,ANY-method}
\alias{[<-,EzRef,ANY,ANY,ANY-method}
\title{The S4 class representing a collection of information about the reference build}
\usage{
EzRef(param)

\S4method{[}{EzRef,ANY}(x, i)

\S4method{[}{EzRef,ANY,ANY,ANY}(x, i) <- value
}
\description{
Objects of this class are a collection of references derived from a list of parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{x[i}: Access of slots by name with square brackets [ ]

\item \code{`[`(x = EzRef, i = ANY, j = ANY) <- value}: Assignment to slots by name with square brackets [ ]

}}
\section{Functions}{
\itemize{
\item \code{EzRef()}: Initializes the slots of EzRef. It will also try to specify some fields and if necessary get full file paths.

}}
\section{Slots}{

\describe{
\item{\code{refBuild}}{a character specifying the file path to the reference build.}

\item{\code{refBuildName}}{a character specifying the name of the reference build.}

\item{\code{refBuildDir}}{a character specifying the directory of the reference build.}

\item{\code{refIndex}}{a character specifying the location of the index that is used in the alignment.}

\item{\code{refFeatureFile}}{a character specifying the file path to the annotation feature file (.gtf).}

\item{\code{refAnnotationFile}}{a character specifying the file path to the annotation file (.txt). Subsetting species are automatically removed}

\item{\code{refFastaFile}}{a character specifying the file path to the fasta file.}

\item{\code{refChromSizesFile}}{a character specifying the file path to the file containing the chromosome sizes.}

\item{\code{refAnnotationVersion}}{a character specifying the annotation version.}
}}

\examples{
refBuild = "Saccharomyces_cerevisiae/Ensembl/EF4/Annotation/Version-2013-03-18"
genomesRoot = "./refExample"
param = ezParam(list(refBuild=refBuild, genomesRoot=genomesRoot))
gtf = system.file("extdata/genes.gtf", package="ezRun", mustWork = TRUE)
fp = system.file("extdata/genome.fa", package="ezRun", mustWork = TRUE)
buildRefDir(param$ezRef, fp, gtf)
buildIgvGenome(param$ezRef)
seqAnno = writeAnnotationFromGtf(param=param)
}
\seealso{
\code{\link{cleanGenomeFiles}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
