% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ezScaleColumns}
\alias{ezScaleColumns}
\title{Scales columns of a matrix}
\usage{
ezScaleColumns(x, scaling)
}
\arguments{
\item{x}{the matrix to scale.}

\item{scaling}{a vector containing the scale for each column of \code{x}.}
}
\value{
Returns a matrix with scaled columns.
}
\description{
The columns will be scaled by multiplying them with the \code{scaling}.
}
\examples{
x = ezScaleColumns(matrix(1:20, 5), 1:4)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
