% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ezScatter}
\alias{ezScatter}
\alias{ezXYScatter}
\title{Does scatter plots}
\usage{
ezScatter(
  x = NULL,
  y,
  xlab = NULL,
  ylab = NULL,
  nPlotsPerRow = 6,
  shrink = FALSE,
  lim = range(x, y, na.rm = TRUE),
  isPresent = NULL,
  types = NULL,
  pch = 16,
  colors = rainbow(ncol(types)),
  legendPos = "bottomright",
  cex.main = 1,
  cex = 1,
  ...
)

ezXYScatter(
  xVec,
  yVec,
  absentColor = "gray",
  shrink = FALSE,
  frame = TRUE,
  axes = TRUE,
  xlim = range(xVec, yVec, na.rm = TRUE),
  ylim = xlim,
  isPresent = NULL,
  types = NULL,
  pch = 16,
  colors = rainbow(ncol(types)),
  legendPos = "bottomright",
  ...
)
}
\arguments{
\item{x}{an optional reference vector or matrix.}

\item{y}{a matrix of values to plot.}

\item{xlab}{a character for labeling the reference. If \code{x} is a matrix, its colnames will be used as labels.}

\item{ylab}{a character vector for the labels of the plots. If \code{ylab} is NULL, the colnames from \code{y} will be used as labels.}

\item{nPlotsPerRow}{an integer specifying the number of plots per row.}

\item{shrink}{a logical specifying whether to shrink the values to range.}

\item{lim}{integers specifying the x and y limits of the plot.}

\item{isPresent}{a logical specifying whether the signal is present.}

\item{types}{a character vector containing the types.}

\item{pch}{an integer specifying the look of plotted points.}

\item{colors}{a character vector containing colors.}

\item{legendPos}{a character vector or integer specifying the position of the legend.}

\item{cex.main}{a numeric specifying the size of main titles relative to \code{cex}.}

\item{cex}{a numeric specifying the size of text and symbols.}

\item{...}{additional arguments to be passed further.}
}
\description{
Does scatter plots.
}
\section{Functions}{
\itemize{
\item \code{ezXYScatter}: Does the XY scatter plot.
}}

\examples{
x = runif(n=1000)
y = runif(n=1000)
isPresent = x > 0.2 & y > 0.2
ezScatter(y=data.frame(a=1:10, b=21:30, c=41:50))
ezXYScatter(x, y, isPresent=isPresent)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
