% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ezSmoothScatter}
\alias{ezSmoothScatter}
\title{Does smooth scatter plots}
\usage{
ezSmoothScatter(
  x = NULL,
  y,
  xlab = NULL,
  ylab = NULL,
  nPlotsPerRow = 6,
  lim = range(x, y, na.rm = TRUE),
  isPresent = NULL,
  types = NULL,
  pch = 16,
  colors = rainbow(ncol(types)),
  legendPos = "bottomright",
  cex.main = 1,
  cex = 1,
  ...
)
}
\arguments{
\item{x}{an optional reference vector or matrix.}

\item{y}{a matrix of values to plot.}

\item{xlab}{a character for labeling the reference. If \code{x} is a matrix, its colnames will be used as labels.}

\item{ylab}{a character vector for the labels of the plots. If \code{ylab} is NULL, the colnames from \code{y} will be used as labels.}

\item{nPlotsPerRow}{an integer specifying the number of plots per row.}

\item{lim}{integers specifying the x and y limits of the plot.}

\item{isPresent}{a logical specifying whether the signal is present.}

\item{types}{a character vector containing the types.}

\item{pch}{an integer specifying the look of plotted points.}

\item{colors}{a character vector containing colors.}

\item{legendPos}{a character vector or integer specifying the position of the legend.}

\item{cex.main}{a numeric specifying the size of main titles relative to \code{cex}.}

\item{cex}{a numeric specifying the size of text and symbols.}

\item{...}{additional arguments to be passed further.}
}
\description{
Does smooth scatter plots.
}
\examples{
ezSmoothScatter(y=data.frame(a=1:10, b=21:30, c=41:50))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
