% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamio.r
\name{ezSortIndexBam}
\alias{ezSortIndexBam}
\title{Sorts and indexes a bam file}
\usage{
ezSortIndexBam(inBam, bam, samtools = SAMTOOLS, maxMem = "10000000000",
  removeBam = TRUE, cores = 2)
}
\arguments{
\item{inBam}{a character representing the file path of the input bam file to sort.}

\item{bam}{a character representing the file path of the output bam file.}

\item{samtools}{a character representing the file path to the samtools directory.}

\item{maxMem}{a character the maximum amount of memory to be used for the new file.}

\item{removeBam}{a logical indicating whether to remove the file specified in \code{inBam}.}

\item{cores}{an integer specifying how many cores to use for the command.}
}
\description{
Sorts and indexes a bam file.
}
\examples{
bamFile <- system.file("extdata", "ex1.bam", package="Rsamtools", mustWork=TRUE)
file.create("test.bam")
ezSortIndexBam("test.bam", basename(bamFile), maxMem="500M", cores = ezThreads())
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

