% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{ezTranscriptDbFromRef}
\alias{ezTranscriptDbFromRef}
\title{1}
\usage{
ezTranscriptDbFromRef(reference, dataSource = "FGCZ")
}
\arguments{
\item{gtfFile}{the gtf file to read the information from.}
}
\value{
Returns
}
\description{
1
}
\examples{
refBuild = "Saccharomyces_cerevisiae/Ensembl/EF4/Annotation/Version-2013-03-18"
genomesRoot = "~/refExample"
myRef = EzRef(param=ezParam(list(refBuild=refBuild)), genomesRoot=genomesRoot)
myTrdb = ezTranscriptDbFromRef(myRef)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}
}

