% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{ezTranscriptDbFromRef}
\alias{ezTranscriptDbFromRef}
\title{Gets the transcripts database from the reference}
\usage{
ezTranscriptDbFromRef(reference, dataSource = "FGCZ")
}
\arguments{
\item{reference}{a reference to read the information from.}

\item{dataSource}{a character specifying the source of the data.}
}
\value{
Returns an object of the class TxDb
}
\description{
Gets the transcripts database from the reference.
}
\examples{
refBuild = "Mus_musculus/UCSC/mm10/Annotation/Version-2012-05-23"
param = ezParam(list(refBuild=refBuild))
tdb = ezTranscriptDbFromRef(param$ezRef)
}
\seealso{
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
