% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{ezUtrSequences}
\alias{ezUtrSequences}
\title{Gets UTR sequences}
\usage{
ezUtrSequences(gtf, chromSeqs)
}
\arguments{
\item{gtf}{the annotation data.frame to get the gene data from.}

\item{chromSeqs}{the chromosome sequences to get the Views from.}
}
\value{
Returns a list with two objects of the class DNAStringSet from the packages Biostrings that represent the 3 and 5 primer.
}
\description{
Gets sequences from untranslated regions.
}
\seealso{
\code{\link[Biostrings]{DNAStringSet}}

\code{\link[IRanges]{Views}}

\code{\link[Biostrings]{reverseComplement}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
