% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{ezWriteElapsed}
\alias{ezWriteElapsed}
\alias{getElapsed}
\title{Writes elapsed time of job}
\usage{
ezWriteElapsed(job, status = "done", con = stderr())

getElapsed(x)
}
\arguments{
\item{job}{a list with fields \code{$name} and \code{$start}.}

\item{status}{a character describing the status of the job.}

\item{con}{the connection to the file to write in.}

\item{x}{corresponds to \code{job$start}.}
}
\description{
Writes the current date and time, the status and how much time elapsed since the jobs' start.
}
\section{Functions}{
\itemize{
\item \code{getElapsed()}: Gets the elapsed time in minutes since a reference time.

}}
\examples{
con = file("example.txt", "w")
job = ezJobStart("a job",con=con)
ezWriteElapsed(job,con=con)
close(con)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
