% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{filterEnrichrResults}
\alias{filterEnrichrResults}
\title{Result filtering}
\usage{
filterEnrichrResults(
  resList,
  maxP = 1,
  maxPAdj = 1,
  maxZ = 0,
  minCombinedScore = 0,
  minOverlapGenes = 3,
  softFilter = F,
  maxResult = NA
)
}
\arguments{
\item{resList}{list of results, i.e. \code{success} field in the list returned by \code{\link{enrichrEnrich}}}

\item{maxP}{p-value threshold (maximum)}

\item{maxPAdj}{adjusted p-value threshold (maximum)}

\item{maxZ}{z-value threshold (maximum)}

\item{minCombinedScore}{combined score threshold (minimum)}

\item{minOverlapGenes}{minimum number of genes in the overlap}

\item{softFilter}{select libraries with at least one significant result but report all results}

\item{maxResult}{maximum number of results to return}
}
\value{
\code{list} that contains the results for libraries with at least one significant gene
that satisfies the criteria.
}
\description{
Filters the results returned by Enrichr. Libraries that fail to yield any
significant results are removed. However, libraries that contain significant results, retain all
records whether significant or not)
}
\seealso{
\code{\link{runEnrichr}, \link{enrichrEnrich}}
}
\author{
Rehrauer, Hubert

Schmid, Peter

Roman Briskine
}
