% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage.r
\name{gageSigGenes}
\alias{gageSigGenes}
\title{Gets significant genes}
\usage{
gageSigGenes(x, gene.pValue = NULL, signal = NULL)
}
\arguments{
\item{x}{the gage results to get significant genes from.}

\item{gene.pValue}{a numeric specifying the p-value threshold.}

\item{signal}{a character specifying which signal to analyse.}
}
\value{
Returns the gage results with labeled significant genes.
}
\description{
Gets significant genes from the gage analysis.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

