% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{getExonNumber}
\alias{getExonNumber}
\title{Gets the exon numbers}
\usage{
getExonNumber(gtf)
}
\arguments{
\item{gtf}{the annotation data.frame to get transcripts from.}
}
\value{
Returns a vector containing the exon numbers.
}
\description{
Gets the exon numbers of an annotation data.frame of the gtf or gff format.
}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param, system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE))
en = getExonNumber(gtf)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
