% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{getGeneMapping}
\alias{getGeneMapping}
\alias{hasGeneMapping}
\title{Gets the isoform-to-gene mapping}
\usage{
getGeneMapping(param, seqAnnoDF)

hasGeneMapping(param, seqAnnoDF)
}
\arguments{
\item{param}{the global parameter object. We need the entries \code{geneColumn} or \code{geneColumnSet} to figure out 
which columns of the annotation data frame 
should be used for the mapping}

\item{seqAnnoDF}{a data.frame containing the sequence annotation.}
}
\value{
Returns a named character vector containing the gene mapping.
}
\description{
The annotation in the reference folders is at the isoform level but for gene-level analyses we need to map the annotation to the gene level.
Especially all enrichment analyses should be run at the gene-level
}
\section{Functions}{
\itemize{
\item \code{hasGeneMapping()}: Returns a logical indicating if the set parameters have gene mapping or not.

}}
\examples{
param = ezParam()
annoFile = system.file("extdata/genes_annotation.txt", package="ezRun", mustWork=TRUE)
param$ezRef["refAnnotationFile"] = annoFile
param$ezRef["refFeatureFile"] = system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE)
param$ezRef["refFastaFile"] = system.file("extdata/genome.fa", package="ezRun", mustWork=TRUE)
seqAnno = writeAnnotationFromGtf(param)
seqAnno = ezFeatureAnnotation(param, rownames(seqAnno), dataFeatureType="gene")
gm = getGeneMapping(param, seqAnno)
hasGeneMapping(param, seqAnno)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
