% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R
\name{getGOparents}
\alias{getGOparents}
\alias{addGoParents}
\title{Gets the GO parents}
\usage{
getGOparents(id, onto = "BP")

addGoParents(gene2goList, onto)
}
\arguments{
\item{id}{the ID to check}

\item{onto}{the ontology to use}
}
\value{
Returns the ancestor terms if they are specified.
}
\description{
Gets the GO parents.
}
\section{Functions}{
\itemize{
\item \code{addGoParents()}: Adds the GO parents.

}}
\examples{
getGOparents("GO:0034767")
addGoParents(c("GO:0034767", "GO:0034768"), "BP")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
