% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{getRpkm}
\alias{getRpkm}
\title{Gets the rpkm measurement}
\description{
  Gets the rpkm measurement from \code{SummarizedExperiment} or 
  \code{SingleCellExperiment} object.
}

\usage{
  getRpkm(rawData)
}
\arguments{
  \item{rawData}{
    \code{SummarizedExperiment} or \code{SingleCellExperiment} object.
  }
}
\value{
  Returns the \code{matrix} of rpkm measurement.
}
\seealso{
  \code{\link{getTpm}}
}
\author{
  Ge Tan
}
\examples{
  \dontrun{
    file <- "/srv/gstore/projects/p1688/CountQC_20201_2017-08-15--20-52-44/input_dataset.tsv"
    param <- ezParam(list(dataRoot="/srv/gstore/projects",
                          refBuild="Homo_sapiens/Ensembl/GRCh38.p10/Annotation/Release_89-2017-05-31"))
    input <- EzDataset$new(file=file, dataRoot=param$dataRoot)
    rawData <- loadCountDataset(input, param)
    getRpkm(rawData)
  }
}