% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{getRpkm}
\alias{getRpkm}
\title{Gets the rpkm measurement}
\usage{
getRpkm(rawData)
}
\arguments{
\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}
}
\value{
Returns the rpkm measurement.
}
\description{
Gets the rpkm measurement.
}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
file = system.file("extdata/yeast_10k_STAR_counts/dataset.tsv", package="ezRun", mustWork = TRUE)
input = EzDataset$new(file=file, dataRoot=param$dataRoot)
rawData = loadCountDataset(input, param)
rpkm = getRpkm(rawData)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

