% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-rsem.R
\name{getRSEMReference}
\alias{getRSEMReference}
\title{Builds and gets a RSEM reference}
\usage{
getRSEMReference(param)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{trinityFasta}{ an optional character specifying the path to a fasta file extracted from the trinity method. If specified, the reference will be prepared using it.}
  \item{ezRef@refIndex}{ a character specifying the location of the index that is used in the alignment.}
  \item{ezRef@refFeatureFile}{ a character specifying the path to the annotation feature file (.gtf).}
  \item{ezRef@refFastaFile}{ a character specifying the path to the fasta file.}
}}
}
\value{
Returns a character specifying the file path to the RSEM indexed reference genome.
}
\description{
Builds and gets a RSEM reference by returning the file path to it.
}
\examples{
param = ezParam(userParam = list('refBuild' = 'Schizosaccharomyces_pombe/Ensembl/EF2/Annotation/Version-2013-03-07'))
getRSEMReference(param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

