% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-salmon.R
\name{getSalmonReference}
\alias{getSalmonReference}
\title{Builds and gets a salmon reference}
\usage{
getSalmonReference(param)
}
\arguments{
\item{param}{a list of parameters:
\itemize{
  \item{transcriptFasta}{ an optional character specifying the path to a fasta file. If specified, the reference will be prepared using it.}
  \item{ezRef@refIndex}{ a character specifying the location of the index that is used in the alignment.}
  \item{ezRef@refFeatureFile}{ a character specifying the path to the annotation feature file (.gtf).}
  \item{ezRef@refFastaFile}{ a character specifying the path to the fasta file.}
}}
}
\value{
Returns a character specifying the file path to the salmon indexed reference genome.
}
\description{
Builds and gets a salmon reference by returning the file path to it.
}
\examples{
ls = list('refBuild' = 'Schizosaccharomyces_pombe/Ensembl/EF2/Annotation/Version-2013-03-07')
param = ezParam(userParam = ls)
getsalmonReference(param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

