% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{getTpm}
\alias{getTpm}
\title{Gets the tpm measurement}
\usage{
getTpm(rawData)
}
\arguments{
\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}
}
\value{
Returns the tpm measurement.
}
\description{
Gets the transcripts per million measurement.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

