% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurePlots.r
\name{getTranscriptCoverage}
\alias{getRangesCoverageChrom}
\alias{getTranscriptCoverage}
\title{Gets transcripts coverages}
\usage{
getTranscriptCoverage(chrom, gff, reads, strandMode = "both")

getRangesCoverageChrom(chrom = NULL, ranges, reads, strandMode = "both")
}
\arguments{
\item{chrom}{a character vector containing chromosome names.}

\item{gff}{an annotation data.frame in gtf or gff format.}

\item{reads}{an object of the class GAlignments.}

\item{strandMode}{a character specifying the mode of the strand.}

\item{ranges}{an object of the class GRanges.}
}
\value{
Returns a list of transcript coverages.
}
\description{
Gets transcripts coverages.
}
\section{Functions}{
\itemize{
\item \code{getRangesCoverageChrom}: Gets the range coverages.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

