% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{getTranscriptGcAndWidth}
\alias{getTranscriptGcAndWidth}
\title{Gets GC proportions and gene widths}
\usage{
getTranscriptGcAndWidth(param = NULL, genomeFn = NULL, featureFn = NULL)
}
\arguments{
\item{param}{the parameters to load the annotation and sequence files from.}
}
\value{
Returns a list containing named vectors for the GC proportion and the gene width.
}
\description{
Gets GC proportions and gene widths from annotation (gtf or gff) and sequence (fasta) information.
}
\examples{
param = ezParam()
param$ezRef@refFeatureFile = system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE)
fp = "/srv/GT/reference/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/WholeGenomeFasta/genome.fa"
param$ezRef@refFastaFile = fp
gw = getTranscriptGcAndWidth(param)
}
\seealso{
\code{\link{ezLoadFeatures}}

\code{\link[Biostrings]{readDNAStringSet}}

\code{\link[IRanges]{Views}}

\code{\link[Biostrings]{letterFrequency}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
