% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{getTranscriptSequences}
\alias{getTranscriptSequences}
\title{Gets transcript sequences}
\usage{
getTranscriptSequences(param, useFivePrimeAsStart = TRUE)
}
\arguments{
\item{param}{the parameters to load the annotation and sequence files from.}

\item{useFivePrimeAsStart}{a logical indicating whether to start from the 5' primer or not}
}
\value{
Returns an object of the class DNAStringSet from the package Biostrings
}
\description{
Gets transcript sequences from annotation (gtf or gff) and sequence (fasta) information.
}
\examples{
param = ezParam()
param$ezRef@refFeatureFile = system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE)
fp = "/srv/GT/reference/Saccharomyces_cerevisiae/Ensembl/EF4/Sequence/WholeGenomeFasta/genome.fa"
param$ezRef@refFastaFile = fp
ts = getTranscriptSequences(param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
