% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{gffToRanges}
\alias{gffToRanges}
\title{Gets GRanges from annotation}
\usage{
gffToRanges(gff)
}
\arguments{
\item{gff}{the annotation data.frame to get the GRanges from.}
}
\value{
Returns a GRanges object
}
\description{
Gets GRanges from annotation of the format gtf or gff.
}
\examples{
param = ezParam()
gtf = ezLoadFeatures(param, system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE))
gRanges = gffToRanges(gtf)
gRanges
}
\seealso{
\code{\link[GenomicRanges]{GRanges}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
