% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-analysis.R, R/go-reports.R
\name{getChildTerms}
\alias{.getGoTermsAsTd}
\alias{getChildTerms}
\alias{goClusterTable}
\title{Gets the GO cluster table}
\usage{
getChildTerms(x, subset, goRelatives, indent = "", childEnvir)

goClusterTable(param, clusterResult)

.getGoTermsAsTd(x, pThreshGo, minCount, onto = NA, maxNumberOfTerms = 40)
}
\arguments{
\item{param}{a list of parameters to extract \code{pValThreshFisher} and \code{minCountFisher} from}

\item{clusterResult}{a list containing the result of the analysis done by \code{goClusterResults()}.}
}
\value{
Returns a flex table containing the GO information of the cluster result.
}
\description{
Gets the GO cluster table
}
\section{Functions}{
\itemize{
\item \code{getChildTerms}: Gets the child terms from a GO term and returns them together.

\item \code{.getGoTermsAsTd}: Gets the GO terms and pastes them into a table.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{goClusterResults}}
}

