% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\name{makeMinimalSingleEndReadDataset}
\alias{makeMinimalSingleEndReadDataset}
\alias{makeMinimalPairedEndReadDataset}
\title{Makes a minimal single end dataset}
\usage{
makeMinimalSingleEndReadDataset(
  fqDir,
  species = "",
  adapter1 = "GATCGGAAGAGCACACGTCTGAACTCCAGTCAC",
  strandMode = "both",
  readCount = NULL,
  dataRoot = DEFAULT_DATA_ROOT
)

makeMinimalPairedEndReadDataset(
  fqDir,
  species = "",
  adapter1 = "GATCGGAAGAGCACACGTCTGAACTCCAGTCAC",
  adapter2 = "AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT",
  strandMode = "both",
  readCount = NULL,
  readTypeSuffix = c("-R1.fastq.gz", "-R2.fastq.gz"),
  dataRoot = DEFAULT_DATA_ROOT
)
}
\arguments{
\item{fqDir}{a character specifying the path to the directory holding the fastq files. The paths stored in the dataset will be relative to the dataRoot}

\item{species}{a character specifying the species name.}

\item{adapter1}{a character representing the adapter sequence.}

\item{strandMode}{a character specifying the strand mode for the dataset.}

\item{dataRoot}{a string specifying the root directory of the data}

\item{adapter2}{a character representing the second adapter sequence in the case of a paired end dataset.}
}
\value{
Returns a data.frame containing the provided information.
}
\description{
Makes a minimal single end dataset by combining the arguments to a data.frame.
You have to switch directory to the root of the data store first.
}
\section{Functions}{
\itemize{
\item \code{makeMinimalPairedEndReadDataset}: Does the same for paired end reads.
}}

\examples{
fqDir = system.file("extdata/yeast_10k", package="ezRun", mustWork=TRUE)
species = "Example"
ds = makeMinimalSingleEndReadDataset(fqDir, species)
ds2 = makeMinimalPairedEndReadDataset(fqDir, species)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
