% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-edgerTwoGroups.R
\name{ngsTwoGroupAnalysis}
\alias{ngsTwoGroupAnalysis}
\alias{runNgsTwoGroupAnalysis}
\title{1}
\usage{
ngsTwoGroupAnalysis(input = NA, output = NA, param = NULL,
  htmlFile = "00index.html")

runNgsTwoGroupAnalysis(dataset, htmlFile = "00index.html", param = param,
  rawData = NULL, types = NULL)
}
\arguments{
\item{input}{an object of the class EzDataset.}

\item{output}{an object of the class EzDataset.}

\item{param}{a list of parameters:
\itemize{
  \item{sampleGroup}{}
  \item{refGroup}{}
  \item{useFactorsAsSampleName}{}
  \item{removeOutliers}{}
  \item{grouping}{}
  \item{batch}{}
  \item{markOutliers}{}
}}

\item{htmlFile}{}
}
\value{
Returns
}
\description{
1
}
\section{Functions}{
\itemize{
\item \code{runNgsTwoGroupAnalysis}: Checks for errors and calls \code{twoGroupCountComparison()}, then \code{writeNgsTwoGroupReport()}.
}}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{twoGroupCountComparison}}

\code{\link{writeNgsTwoGroupReport}}
}

