% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R
\name{phyloSeqOTU}
\alias{phyloSeqOTU}
\title{Phyloseq OTU object}
\usage{
phyloSeqOTU(otuFileName)
}
\arguments{
\item{otuFileName, }{mothur shared clustered OTU  files.}
}
\value{
Returns a Phyloseq OTU object.
}
\description{
Create Phyloseq OTU object mothur OTU files.
}
