% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothurAssociatedFunctions.R
\name{phyloSeqTaxa}
\alias{phyloSeqTaxa}
\title{Phyloseq Taxa object}
\usage{
phyloSeqTaxa(taxaFileName, technology)
}
\arguments{
\item{taxaFileName, }{mothur taxonomy file.}
}
\value{
Returns a Phyloseq Taxa object.
}
\description{
Create Phyloseq taxa object from mothur taxonomy files.
}
