% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurePlots.r
\name{plotLocusCoverageProfile}
\alias{plotLocusCoverageProfile}
\title{Title}
\usage{
plotLocusCoverageProfile(
  gRanges,
  bamFiles,
  gtfFile = NULL,
  height = 10,
  width = 20
)
}
\arguments{
\item{gRanges}{}

\item{bamFiles}{}

\item{gtfFile}{}

\item{height}{}

\item{width}{}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
require(GenomicRanges)
require(ezRun)
gtfFile = "/srv/GT/reference/Homo_sapiens/Ensembl/GRCh38.p10/Annotation/Release_89-2017-05-31/Genes/genes.gtf"
myGeneRange = GRanges(seqnames="2", ranges=IRanges(start=1, end=1000, name="foo"))
plotLocusCoverageProfile(myGeneRange, bamFiles, gtfFile=gtfFile,
                         height=20, width=20)
}
}
