% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsio.R
\name{readNcProResult}
\alias{readNcProResult}
\title{Reads NcPro results}
\usage{
readNcProResult(ncproDir)
}
\arguments{
\item{ncproDir}{a character representing the file path to the directory that contains the NcPro results.}
}
\value{
Returns the rawdata read from the NcPro results.
}
\description{
Reads NcPro results and performs some checks on the data.
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
