% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\name{runCountSingleBam}
\alias{runCountSingleBam}
\title{Run counts for a single BAM file}
\usage{
runCountSingleBam(
  psBamFile,
  psGffFile,
  psCountfileExt,
  strandMode,
  Paired,
  ramPerJob,
  bgExpression
)
}
\description{
Run counts for a single BAM file
}
