% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{selectFeatures}
\alias{selectFeatures}
\title{Selects features from raw data}
\usage{
selectFeatures(rawData, keep)
}
\arguments{
\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}

\item{keep}{a character vector specifying which signals or counts to keep.}
}
\value{
Returns the selected raw data.
}
\description{
Selects features from raw data by looking at "signal" (if specified) or "counts" from \code{rawData}.
}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
file = system.file("extdata/yeast_10k_STAR_counts/dataset.tsv", package="ezRun", mustWork = TRUE)
input = EzDataset$new(file=file, dataRoot=param$dataRoot)
rawData = loadCountDataset(input, param)
keep = "YFR014C"
rd2 = selectFeatures(rawData, keep)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
