% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rawData.R
\name{selectSamples}
\alias{selectSamples}
\title{Select samples from raw data}
\usage{
selectSamples(rawData, samples)
}
\arguments{
\item{rawData}{a list of raw data. Usually obtained from \code{loadCountDataset()}.}

\item{samples}{a character vector specifying which samples to select.}
}
\value{
Returns a list of the selected samples.
}
\description{
Select samples from raw data by specifying either the signal names or the sample names.
}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
input = EzDataset$new(file=system.file("extdata/yeast_10k_STAR_featureCounts/dataset.tsv", package="ezRun", mustWork = TRUE))
rawData = loadCountDataset(input, param)
selectSamples(rawData, c("wt_1","wt_2"))
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

