% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsTools.R
\name{strandValue}
\alias{strandValue}
\alias{strandName}
\title{Gets a strand value}
\usage{
strandValue(x)

strandName(x)
}
\arguments{
\item{x}{an integer to select the strand value: 1 equals plus and 2 equals minus.}
}
\value{
Returns a named character "+" or "-".
}
\description{
Gets a strand value of plus or minus from an integer.
}
\section{Functions}{
\itemize{
\item \code{strandName()}: Does the same, but returns the character as "plus" or "minus".

}}
\examples{
strandValue(1)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
