% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangesTools.R
\name{subSampleRle}
\alias{subSampleRle}
\title{Subsets an Rle object}
\usage{
subSampleRle(x, idx)
}
\arguments{
\item{x}{an object of the class Rle.}

\item{idx}{an integer vector specifying which indices to keep.}
}
\value{
Returns a subset of an Rle object as an integer or numeric vector.
}
\description{
Subsets an Rle object.
}
\examples{
rleobj = Rle(c(1, 1, 1, 2, 2, 3, 2, 2))
subSampleRle(rleobj, 1:8)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
