% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoGroups.R
\name{twoGroupCountComparison}
\alias{runDeseq2}
\alias{runEdger}
\alias{runGfold}
\alias{runGlm}
\alias{twoGroupCountComparison}
\title{Compares the counts of two groups}
\usage{
twoGroupCountComparison(rawData, param)

runGfold(rawData, scalingFactors, isSample, isRef)

runDeseq2(x, sampleGroup, refGroup, grouping, batch = NULL,
  isPresent = NULL)

runEdger(x, sampleGroup, refGroup, grouping, normMethod)

runGlm(x, sampleGroup, refGroup, grouping, normMethod, batch = NULL)
}
\arguments{
\item{rawData}{a list of raw data obtained from \code{loadCountDataset()}.}

\item{param}{a list of parameters:
\itemize{
  \item{testMethod}{ defines the method to run: deseq2, exactTest, glm, sam or limma. Defaults to glm.}
  \item{batch}{ a character vector specifying the batch groups.}
  \item{grouping}{ a character specifying the grouping.}
  \item{sampleGroup}{ a character specifying the group to sample.}
  \item{refGroup}{ a character specifying the reference group.}
  \item{normMethod}{ a character specifying the normalization method for the edger and glm test methods.}
  \item{runGfold}{ a logical indicating whether to run Gfold.}
}}
}
\value{
Returns a list containing the results of the comparison.
}
\description{
Compares the counts of two groups with the option to choose from several methods to test them.
}
\section{Functions}{
\itemize{
\item \code{runGfold}: Runs the Gfold test method.

\item \code{runDeseq2}: Runs the Deseq2 test method.

\item \code{runEdger}: Runs the EdgeR test method.

\item \code{runGlm}: Runs the Glm test method.
}}
\author{
Rehrauer, Hubert

Schmid, Peter
}

