% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-deseq2TwoGroups.R, R/app-edgerTwoGroups.R
\name{runDeseq2}
\alias{runDeseq2}
\alias{runEdger}
\alias{runGfold}
\alias{runGlm}
\alias{twoGroupCountComparison}
\title{Compares the counts of two groups}
\usage{
runDeseq2(x, sampleGroup, refGroup, grouping, batch = NULL,
  isPresent = NULL)

twoGroupCountComparison(rawData, param)

runGfold(rawData, scalingFactors, isSample, isRef)

runEdger(x, sampleGroup, refGroup, grouping, normMethod)

runGlm(x, sampleGroup, refGroup, grouping, normMethod, batch = NULL)
}
\arguments{
\item{rawData}{a list of raw data. Usually obtained from \code{loadCountDataset()}.}

\item{param}{a list of parameters:
\itemize{
  \item{testMethod}{ defines the method to run: deseq2, exactTest, glm, sam or limma. Defaults to glm.}
  \item{batch}{}
  \item{grouping}{ a character specifying the grouping.}
  \item{sampleGroup}{ a character specifying the group to sample.}
  \item{refGroup}{ a character specifying the reference group.}
  \item{normMethod}{ a character specifying the normalization method for the edger and glm test methods.}
  \item{runGfold}{ a logical indicating whether to run Gfold.}
}}
}
\value{
Returns a list containing the results of the comparison.
}
\description{
Compares the counts of two groups with the option to choose from several methods to test them.
}
\section{Functions}{
\itemize{
\item \code{runDeseq2}: Runs the Deseq2 test method.

\item \code{runGfold}: Runs the Gfold test method.

\item \code{runEdger}: Runs the Edger test method.

\item \code{runGlm}: Runs the Glm test method.
}}
\examples{
param = ezParam()
param$dataRoot = system.file(package="ezRun", mustWork = TRUE)
param$normMethod = "RLE"
input = EzDataset$new(file=system.file("extdata/yeast_10k_STAR_featureCounts/dataset.tsv", package="ezRun", mustWork = TRUE))
rawData = loadCountDataset(input$copy()$subset(1), param)
twoGroupCountComparison(rawData, param)
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

