% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-DEXSeqAnalysis.R
\name{writeCountFilesToMeta}
\alias{writeCountFilesToMeta}
\title{Write names of countfiles back into the input file}
\usage{
writeCountFilesToMeta(pvCountFiles, input)
}
\description{
Write names of countfiles back into the input file
}

