% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igv.R
\name{writeIgvJnlp}
\alias{writeIgvJnlp}
\alias{ezIgvTemplateFile}
\title{Writes a .jnlp file}
\usage{
writeIgvJnlp(jnlpFile, projectId, sessionUrl)

ezIgvTemplateFile()
}
\arguments{
\item{jnlpFile}{the connection to write the IGV.}

\item{projectId}{a character representing the project ID.}

\item{sessionUrl}{a character representing the URL of the session.}
}
\value{
Returns lines written to the specified connection.
}
\description{
Writes a .jnlp file and adds the \code{projectId} and the \code{sessionUrl} to the template.
}
\section{Functions}{
\itemize{
\item \code{ezIgvTemplateFile()}: Gets the IGV template file.

}}
\author{
Rehrauer, Hubert

Schmid, Peter
}
