% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igv.r
\name{writeIgvSession}
\alias{writeIgvSession}
\alias{getIgvGenome}
\alias{addIgvSessionLink}
\alias{writeIgvSessionLink}
\title{Writes a jnlp link to a static html}
\usage{
writeIgvSession(
  genome,
  refBuild,
  file = "igvSession.xml",
  bamUrls = NULL,
  vcfUrls = NULL,
  locus = "All"
)

getIgvGenome(param)

addIgvSessionLink(
  genome,
  refBuild,
  bamFiles,
  doc,
  locus = "All",
  label = "Open Integrative Genomics Viewer",
  baseUrl = PROJECT_BASE_URL
)

writeIgvSessionLink(
  genome,
  refBuild,
  bamFiles,
  html,
  locus = "All",
  label = "Open Integrative Genomics Viewer",
  baseUrl = PROJECT_BASE_URL
)
}
\arguments{
\item{genome}{a character representing the build name.}

\item{refBuild}{a character representing a file path to the reference build.}

\item{file}{a character representing the name of the file to write the IGV session in.}

\item{bamUrls}{a character vector containing BAM file links.}

\item{vcfUrls}{a character vector containing VCF file links.}

\item{locus}{a character describing which loci were used.}

\item{html}{a character representing an html address.}

\item{htmlFile}{a character, ending with .html, representing the file path of an html file.}

\item{projectId}{a character representing the project ID.}
}
\value{
Returns a character containing the name of the newly written .xml file.
}
\description{
Writes a jnlp link to a static html. The html can be moved as long as the session xml stays in the same directory.

Writes an IGV session into a separate .xml file
}
\section{Functions}{
\itemize{
\item \code{getIgvGenome}: Gets the IGV genome if specified or otherwise tries to get the build name from the parameters.

\item \code{addIgvSessionLink}: Adds an IGV session link to an object of the class bsdoc.

\item \code{writeIgvSessionLink}: Writes an IGV session link.
}}

\examples{
param = ezParam()
genome = getIgvGenome(param)
writeIgvSession(genome, param$ezRef["refBuild"])
}
\seealso{
\code{\link[XML]{newXMLNode}}

\code{\link[XML]{saveXML}}

\code{\link[XML]{addChildren}}
}
\author{
Rehrauer, Hubert

Schmid, Peter

Rehrauer, Hubert

Schmid, Peter
}
