% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-edgerTwoGroups.R
\name{writeNgsTwoGroupReport}
\alias{writeNgsTwoGroupReport}
\title{1}
\usage{
writeNgsTwoGroupReport(dataset, result, htmlFile, param = NA, rawData = NA,
  types = NULL)
}
\arguments{
\item{dataset}{}

\item{result}{}

\item{htmlFile}{}

\item{param}{a list of parameters:
\itemize{
  \item{logColorRange}{}
  \item{minSignal}{}
  \item{grouping}{}
  \item{sampleGroup}{}
  \item{refGroup}{}
  \item{pValueHighlightThresh}{}
  \item{log2RatioHighlightThresh}{}
  \item{maxGenesForClustering}{}
  \item{minGenesForClustering}{}
  \item{doZip}{}
  \item{goseqMethod}{}
  \item{maxNumberGroupsDisplayed}{}
}}

\item{rawData}{}

\item{types}{}
}
\value{
Returns
}
\description{
1
}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}

