% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{writePresplicedGtf}
\alias{writePresplicedGtf}
\title{Writes pre-spliced annotation}
\usage{
writePresplicedGtf(param, featureFile = param$ezRef["refFeatureFile"])
}
\arguments{
\item{param}{contains the feature file to load the annotation from.}

\item{featureFile}{the file to load the features in the \code{ezLoadFeatures()} call from.}
}
\description{
Writes pre-spliced annotation directly from the parameters into a new .gtf file.
}
\examples{
param = ezParam()
writePresplicedGtf(param, system.file("extdata/genes.gtf", package="ezRun", mustWork=TRUE))
}
\seealso{
\code{\link{ezLoadFeatures}}
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
