% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmlOutput.r
\name{writeResultCounts}
\alias{getSignificantCountsTable}
\alias{getSignificantFoldChangeCountsTable}
\alias{writeResultCounts}
\title{1}
\usage{
writeResultCounts(html, param, result, geneIds = NULL, pThresh = c(0.1,
  0.05, 1/10^(2:5)))

getSignificantCountsTable(result, pThresh = 1/10^(1:5), genes = NULL)

getSignificantFoldChangeCountsTable(result, pThresh = 1/10^(1:5),
  fcThresh = c(1, 1.5, 2, 3, 4, 8, 10), genes = NULL)
}
\arguments{
\item{html}{a connection to an html file.}

\item{result}{}

\item{geneIds}{}

\item{pThresh}{}
}
\description{
1
}
\section{Functions}{
\itemize{
\item \code{getSignificantCountsTable}: 

\item \code{getSignificantFoldChangeCountsTable}: 
}}
\examples{
1
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
\seealso{
\code{\link{writeTableToHtml}}
}

