% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{zipFile}
\alias{zipFile}
\alias{zipWorkingDir}
\title{Archives files}
\usage{
zipFile(inputs, zipped = NULL)

zipWorkingDir(zipName)
}
\arguments{
\item{inputs}{one or several input files to archive.}

\item{zipped}{optional character to name the archive.}

\item{zipName}{a character naming the working directory to be archived.}
}
\value{
Returns the file name of the zipped archive.
}
\description{
Archives one or several files to a .zip extension. Optionally, a separate name can be given to the archive.
}
\section{Functions}{
\itemize{
\item \code{zipWorkingDir()}: Archives the whole working directory.

}}
\examples{
write("a file","exampleFile")
zipFile("exampleFile")
}
\author{
Rehrauer, Hubert

Schmid, Peter
}
